Test for Hardy Weinberg equilibrium

Test experiment features for Hardy Weinberg equilibrium.

HWETest(
  object,
  experiment = c("hla_alleles", "hla_aa", "hla_g_groups", "hla_supertypes",
    "hla_NK_ligands"),
  HWE_group = NULL,
  HWE_cutoff = NULL,
  as.MiDAS = FALSE
)

Arguments

object
MiDAS object.

experiment
String specifying experiment to test. Valid values includes "hla_alleles", "hla_aa", "hla_g_groups", "hla_supertypes", "hla_NK_ligands".

HWE_group
Expression defining samples grouping to test for Hardy Weinberg equilibrium. By default samples are not grouped.

HWE_cutoff
Number specifying p-value threshold. When HWE_group is specified both groups are thresholded.

as.MiDAS
Logical flag indicating if MiDAS object should be returned.

Value

Data frame with Hardy Weinberg Equilibrium test results or a filtered MiDAS object.

Details

Setting as.MiDAS to TRUE will filter MiDAS object based on p-value cut-off given by HWE_cutoff.

Examples

# create MiDAS object
midas <- prepareMiDAS(hla_calls = MiDAS_tut_HLA,
                      colData = MiDAS_tut_pheno,
                      experiment = "hla_alleles"
)

# get HWE p-values as data frame
HWETest(midas, experiment = "hla_alleles")

#>             var       p.value
#> 1       A*01:01  9.815768e-01
#> 2       A*01:02  9.747476e-01
#> 3      A*01:234  9.873786e-01
#> 4      A*01:237  9.873786e-01
#> 5      A*01:274  9.494700e-01
#> 6       A*01:43  9.873786e-01
#> 7       A*01:98  9.873786e-01
#> 8       A*02:01  9.932791e-01
#> 9       A*02:02  9.747476e-01
#> 10      A*02:03  9.115621e-01
#> 11      A*02:05  4.993169e-04
#> 12      A*02:06  4.712638e-08
#> 13      A*02:07  9.115621e-01
#> 14      A*02:08  9.747476e-01
#> 15      A*02:10  9.621104e-01
#> 16     A*02:211  9.873786e-01
#> 17      A*02:22  9.873786e-01
#> 18     A*02:324  9.873786e-01
#> 19      A*02:60  9.873786e-01
#> 20     A*02:614  9.873786e-01
#> 21     A*02:642  9.873786e-01
#> 22     A*02:686  9.873786e-01
#> 23     A*02:753  9.873786e-01
#> 24     A*02:763  9.873786e-01
#> 25      A*03:01  5.024356e-01
#> 26      A*03:02  8.360935e-01
#> 27     A*03:205  9.873786e-01
#> 28      A*03:24  9.873786e-01
#> 29     A*03:312  9.621104e-01
#> 30      A*03:72  9.747476e-01
#> 31      A*11:01  1.055618e-01
#> 32      A*11:02  9.873786e-01
#> 33      A*11:03  9.873786e-01
#> 34      A*11:12  9.873786e-01
#> 35     A*11:306  9.621104e-01
#> 36      A*11:43  9.873786e-01
#> 37     A*11:50Q  9.873786e-01
#> 38      A*11:88  9.621104e-01
#> 39      A*23:01  4.466456e-01
#> 40      A*23:17  9.873786e-01
#> 41      A*23:18  9.873786e-01
#> 42      A*24:02  7.551682e-01
#> 43      A*24:03  9.747476e-01
#> 44      A*24:04  9.873786e-01
#> 45      A*24:07 1.795833e-219
#> 46     A*24:104  9.747476e-01
#> 47      A*24:20  9.494700e-01
#> 48      A*24:50  9.494700e-01
#> 49      A*24:95  9.873786e-01
#> 50      A*25:01  3.985913e-01
#> 51      A*25:09  9.747476e-01
#> 52      A*25:25  9.873786e-01
#> 53      A*26:01  9.853229e-01
#> 54      A*26:02  9.494700e-01
#> 55      A*26:03  9.747476e-01
#> 56      A*26:08  9.621104e-01
#> 57      A*26:16  9.873786e-01
#> 58      A*29:01  9.241927e-01
#> 59      A*29:02  4.777554e-01
#> 60     A*29:119  9.621104e-01
#> 61      A*30:01  5.511430e-01
#> 62      A*30:02  4.482222e-01
#> 63      A*30:04  9.621104e-01
#> 64      A*31:01  3.449400e-01
#> 65      A*31:08  9.873786e-01
#> 66      A*31:21  9.873786e-01
#> 67      A*32:01  3.449400e-01
#> 68      A*32:02  9.873786e-01
#> 69     A*32:116  9.494700e-01
#> 70      A*33:01  7.494065e-01
#> 71      A*33:03  7.987058e-01
#> 72      A*33:05  9.747476e-01
#> 73     A*33:125  9.873786e-01
#> 74     A*33:163  9.873786e-01
#> 75      A*33:51  9.873786e-01
#> 76      A*34:01  9.873786e-01
#> 77      A*34:02  9.241927e-01
#> 78      A*66:01  8.737410e-01
#> 79      A*68:01  7.151258e-01
#> 80      A*68:02  7.372008e-01
#> 81      A*68:04  9.873786e-01
#> 82      A*68:16  9.873786e-01
#> 83     A*68:164  9.873786e-01
#> 84      A*68:17  9.873786e-01
#> 85      A*68:23  9.873786e-01
#> 86      A*68:66  9.873786e-01
#> 87      A*69:01  9.241927e-01
#> 88      A*74:01  9.747476e-01
#> 89      B*07:02  5.715522e-02
#> 90      B*07:04  9.873786e-01
#> 91      B*07:05  9.368297e-01
#> 92      B*07:06  9.368297e-01
#> 93      B*07:10  9.873786e-01
#> 94      B*07:14  9.873786e-01
#> 95      B*07:22  9.873786e-01
#> 96      B*07:47  9.873786e-01
#> 97      B*08:01  3.148466e-01
#> 98     B*08:156  9.747476e-01
#> 99      B*08:94  9.873786e-01
#> 100     B*13:01  9.368297e-01
#> 101     B*13:02  1.954175e-02
#> 102    B*13:117  9.873786e-01
#> 103     B*14:01  7.250487e-01
#> 104     B*14:02  5.186921e-01
#> 105     B*14:28  9.873786e-01
#> 106     B*15:01  1.423391e-03
#> 107     B*15:02  9.241927e-01
#> 108     B*15:03  8.989411e-01
#> 109     B*15:07  9.747476e-01
#> 110     B*15:09  9.747476e-01
#> 111     B*15:10  9.494700e-01
#> 112     B*15:11  9.621104e-01
#> 113     B*15:12  9.873786e-01
#> 114     B*15:15  9.873786e-01
#> 115     B*15:17  8.737410e-01
#> 116     B*15:18  9.621104e-01
#> 117     B*15:20  9.873786e-01
#> 118    B*15:212  9.873786e-01
#> 119    B*15:220  9.873786e-01
#> 120     B*15:25  9.621104e-01
#> 121    B*15:256  9.873786e-01
#> 122     B*15:27  9.873786e-01
#> 123     B*15:39  9.873786e-01
#> 124     B*15:63  9.873786e-01
#> 125     B*18:01  1.366843e-01
#> 126     B*18:02  9.873786e-01
#> 127     B*18:03  9.873786e-01
#> 128     B*18:07  9.873786e-01
#> 129     B*27:02  8.863330e-01
#> 130     B*27:03  9.621104e-01
#> 131     B*27:04  9.747476e-01
#> 132     B*27:05  1.865273e-02
#> 133     B*27:06  9.873786e-01
#> 134    B*27:127  9.747476e-01
#> 135     B*27:14  9.873786e-01
#> 136     B*35:01  7.885904e-02
#> 137     B*35:02  1.515212e-05
#> 138     B*35:03  5.294094e-01
#> 139     B*35:05 1.795833e-219
#> 140     B*35:08  8.111342e-01
#> 141     B*35:09  9.873786e-01
#> 142     B*35:17  9.873786e-01
#> 143    B*35:194  9.873786e-01
#> 144     B*35:41  9.873786e-01
#> 145     B*35:42  9.873786e-01
#> 146     B*35:43  9.873786e-01
#> 147     B*35:60  9.873786e-01
#> 148     B*35:67  9.873786e-01
#> 149     B*35:71  9.873786e-01
#> 150     B*37:01  6.889442e-01
#> 151     B*37:68  9.873786e-01
#> 152     B*38:01  1.116598e-01
#> 153     B*38:02  9.368297e-01
#> 154     B*39:01  6.417333e-01
#> 155     B*39:05  9.873786e-01
#> 156     B*39:06  8.737410e-01
#> 157     B*39:09  9.873786e-01
#> 158     B*40:01  8.582846e-01
#> 159     B*40:02  6.534281e-01
#> 160     B*40:06  4.993169e-04
#> 161    B*40:117  9.873786e-01
#> 162     B*40:21  9.873786e-01
#> 163     B*40:32  9.873786e-01
#> 164    B*40:353  9.873786e-01
#> 165     B*40:49  9.873786e-01
#> 166     B*41:01  8.863330e-01
#> 167     B*41:02  8.863330e-01
#> 168     B*42:01  9.368297e-01
#> 169     B*42:02  9.747476e-01
#> 170     B*44:02  7.252547e-01
#> 171     B*44:03  9.347324e-01
#> 172     B*44:04  9.873786e-01
#> 173     B*44:05  9.368297e-01
#> 174     B*44:06  9.873786e-01
#> 175     B*44:27  8.737410e-01
#> 176    B*44:270  9.747476e-01
#> 177     B*44:37  9.873786e-01
#> 178     B*44:55  9.873786e-01
#> 179     B*45:01  8.111342e-01
#> 180     B*46:01  8.611683e-01
#> 181     B*47:01  9.241927e-01
#> 182     B*47:04  9.747476e-01
#> 183     B*48:01  2.064515e-08
#> 184     B*49:01  4.189087e-01
#> 185     B*50:01  6.185757e-01
#> 186     B*50:02  9.873786e-01
#> 187     B*51:01  8.550926e-01
#> 188     B*51:02  9.873786e-01
#> 189     B*51:05  9.873786e-01
#> 190     B*51:07  9.494700e-01
#> 191     B*51:08  9.241927e-01
#> 192     B*51:09  9.873786e-01
#> 193   B*51:173Q  9.873786e-01
#> 194    B*51:203  9.873786e-01
#> 195    B*51:230  9.873786e-01
#> 196     B*52:01  4.189087e-01
#> 197     B*53:01  8.989411e-01
#> 198     B*54:01  8.486181e-01
#> 199     B*55:01  1.350792e-01
#> 200     B*55:02  8.989411e-01
#> 201     B*56:01  7.863157e-01
#> 202     B*56:03  9.873786e-01
#> 203     B*56:09  9.873786e-01
#> 204     B*57:01  7.697112e-01
#> 205     B*57:02  9.115621e-01
#> 206     B*57:03  9.241927e-01
#> 207     B*58:01  7.250487e-01
#> 208     B*58:02  9.368297e-01
#> 209     B*59:01  9.747476e-01
#> 210     B*67:01  9.873786e-01
#> 211     B*73:01  9.747476e-01
#> 212     C*01:02  7.392314e-01
#> 213     C*01:03  9.747476e-01
#> 214    C*01:103  9.873786e-01
#> 215     C*01:22  9.873786e-01
#> 216     C*01:51  9.873786e-01
#> 217     C*02:02  9.616256e-01
#> 218     C*02:06  9.873786e-01
#> 219     C*02:10  8.989411e-01
#> 220     C*02:20  9.873786e-01
#> 221     C*03:02  8.360935e-01
#> 222     C*03:03  2.887710e-01
#> 223     C*03:04  1.546051e-02
#> 224     C*03:05  9.873786e-01
#> 225     C*03:06  9.873786e-01
#> 226    C*03:344  9.494700e-01
#> 227     C*03:39  9.621104e-01
#> 228     C*03:85  9.873786e-01
#> 229     C*04:01  4.869612e-02
#> 230     C*04:03  9.621104e-01
#> 231    C*04:277  9.873786e-01
#> 232     C*05:01  1.118098e-01
#> 233     C*05:05  9.873786e-01
#> 234     C*05:37  9.873786e-01
#> 235     C*06:02  3.701502e-01
#> 236     C*06:09  9.873786e-01
#> 237     C*07:01  5.073501e-01
#> 238     C*07:02  3.290282e-01
#> 239     C*07:04  5.844583e-01
#> 240     C*07:18  8.737410e-01
#> 241    C*07:204  9.873786e-01
#> 242     C*07:51  9.747476e-01
#> 243    C*07:578  9.115621e-01
#> 244    C*07:615  9.873786e-01
#> 245    C*07:696  9.747476e-01
#> 246     C*08:01  7.250487e-01
#> 247     C*08:02  7.873761e-01
#> 248     C*08:03  4.642348e-12
#> 249     C*12:02  4.189087e-01
#> 250     C*12:03  5.768731e-02
#> 251    C*12:132  9.873786e-01
#> 252    C*12:167  9.873786e-01
#> 253     C*12:28  9.873786e-01
#> 254     C*14:02  6.417333e-01
#> 255     C*14:03  9.368297e-01
#> 256     C*15:02  4.466456e-01
#> 257     C*15:05  8.486181e-01
#> 258     C*15:07  9.873786e-01
#> 259     C*15:09  9.873786e-01
#> 260     C*16:01  4.270695e-01
#> 261     C*16:02  7.987058e-01
#> 262     C*16:04  9.494700e-01
#> 263     C*17:01  8.611683e-01
#> 264     C*17:03  8.737410e-01
#> 265     C*18:02  9.621104e-01
#> 266  DPA1*01:03  1.029139e-03
#> 267  DPA1*01:04  8.486181e-01
#> 268  DPA1*01:05  9.873786e-01
#> 269  DPA1*01:11  9.873786e-01
#> 270  DPA1*02:01  1.996524e-01
#> 271  DPA1*02:02  6.702482e-25
#> 272  DPA1*02:06  8.863330e-01
#> 273  DPA1*02:07  8.360935e-01
#> 274  DPA1*03:01  8.989411e-01
#> 275  DPA1*04:01  9.873786e-01
#> 276  DPB1*01:01  5.754677e-01
#> 277  DPB1*02:01  1.480374e-01
#> 278  DPB1*02:02  9.717136e-02
#> 279  DPB1*03:01  8.787774e-01
#> 280  DPB1*04:01  5.594016e-05
#> 281  DPB1*04:02  9.465083e-02
#> 282  DPB1*05:01  3.458069e-14
#> 283  DPB1*06:01  2.233096e-03
#> 284  DPB1*09:01  7.494065e-01
#> 285  DPB1*10:01  6.889442e-01
#> 286 DPB1*104:01  5.455257e-02
#> 287 DPB1*105:01  8.235979e-01
#> 288 DPB1*107:01  9.873786e-01
#> 289  DPB1*11:01  5.186921e-01
#> 290 DPB1*113:01  9.747476e-01
#> 291 DPB1*124:01  8.486181e-01
#> 292 DPB1*126:01  9.115621e-01
#> 293  DPB1*13:01  8.586290e-02
#> 294 DPB1*131:01  9.747476e-01
#> 295 DPB1*133:01  9.873786e-01
#> 296 DPB1*135:01  9.241927e-01
#> 297 DPB1*138:01  9.873786e-01
#> 298  DPB1*14:01  3.053515e-01
#> 299  DPB1*15:01  8.360935e-01
#> 300 DPB1*150:01  9.873786e-01
#> 301 DPB1*158:01  9.873786e-01
#> 302  DPB1*16:01  8.737410e-01
#> 303 DPB1*162:01  9.368297e-01
#> 304  DPB1*17:01  6.401594e-02
#> 305  DPB1*18:01  9.873786e-01
#> 306  DPB1*19:01  9.115621e-01
#> 307  DPB1*20:01  9.115621e-01
#> 308  DPB1*21:01  9.494700e-01
#> 309  DPB1*22:01  9.873786e-01
#> 310  DPB1*23:01  8.111342e-01
#> 311 DPB1*234:01  9.873786e-01
#> 312 DPB1*259:01  9.873786e-01
#> 313 DPB1*271:01  9.494700e-01
#> 314 DPB1*294:01  9.873786e-01
#> 315  DPB1*33:01  9.873786e-01
#> 316 DPB1*350:01  9.115621e-01
#> 317 DPB1*351:01  9.621104e-01
#> 318  DPB1*36:01  9.873786e-01
#> 319  DPB1*38:01  9.873786e-01
#> 320 DPB1*398:01  9.873786e-01
#> 321  DPB1*40:01  9.873786e-01
#> 322  DPB1*41:01  9.873786e-01
#> 323 DPB1*414:01  9.115621e-01
#> 324  DPB1*45:01  9.873786e-01
#> 325 DPB1*461:01  9.873786e-01
#> 326 DPB1*462:01  9.873786e-01
#> 327 DPB1*463:01  9.494700e-01
#> 328  DPB1*47:01  9.873786e-01
#> 329  DPB1*51:01  9.873786e-01
#> 330 DPB1*519:01  9.873786e-01
#> 331 DPB1*547:01  9.873786e-01
#> 332  DPB1*57:01  9.621104e-01
#> 333 DPB1*651:01  9.494700e-01
#> 334 DPB1*702:01  9.115621e-01
#> 335 DPB1*744:01  9.873786e-01
#> 336  DPB1*80:01  9.873786e-01
#> 337  DPB1*88:01  9.873786e-01
#> 338  DPB1*92:01  9.873786e-01
#> 339  DQA1*01:01  1.679464e-01
#> 340  DQA1*01:02  8.227547e-01
#> 341  DQA1*01:03  5.308181e-01
#> 342  DQA1*01:04  3.985913e-01
#> 343  DQA1*01:05  6.889442e-01
#> 344  DQA1*01:06  9.873786e-01
#> 345  DQA1*01:10  9.747476e-01
#> 346  DQA1*02:01  1.909432e-01
#> 347  DQA1*03:01  4.878395e-01
#> 348  DQA1*03:02  2.655992e-04
#> 349  DQA1*03:03  4.866700e-01
#> 350  DQA1*04:01  8.582846e-01
#> 351  DQA1*04:02  9.747476e-01
#> 352  DQA1*05:01  5.414476e-03
#> 353  DQA1*05:03  7.494065e-01
#> 354  DQA1*05:05  3.596649e-01
#> 355  DQA1*05:06  9.873786e-01
#> 356  DQA1*05:07  9.873786e-01
#> 357  DQA1*05:08  9.873786e-01
#> 358  DQA1*05:09  9.241927e-01
#> 359  DQA1*06:01  1.061717e-02
#> 360  DQB1*02:01  2.985525e-03
#> 361  DQB1*02:02  3.903724e-02
#> 362  DQB1*02:10  9.873786e-01
#> 363  DQB1*03:01  3.718184e-02
#> 364  DQB1*03:02  3.816382e-01
#> 365  DQB1*03:03  6.860990e-01
#> 366  DQB1*03:04  9.115621e-01
#> 367  DQB1*03:05  9.621104e-01
#> 368  DQB1*03:19  8.611683e-01
#> 369  DQB1*03:22  7.250487e-01
#> 370 DQB1*03:251  9.747476e-01
#> 371  DQB1*03:96  9.368297e-01
#> 372  DQB1*04:01  6.403459e-06
#> 373  DQB1*04:02  3.449400e-01
#> 374  DQB1*04:23  9.621104e-01
#> 375  DQB1*05:01  2.457877e-01
#> 376  DQB1*05:02  7.582602e-02
#> 377  DQB1*05:03  4.367994e-01
#> 378  DQB1*05:04  9.494700e-01
#> 379 DQB1*05:107  9.873786e-01
#> 380  DQB1*06:01  6.401594e-02
#> 381  DQB1*06:02  7.277424e-02
#> 382  DQB1*06:03  2.835501e-01
#> 383  DQB1*06:04  8.331507e-01
#> 384  DQB1*06:07  9.873786e-01
#> 385  DQB1*06:09  7.739671e-01
#> 386   DRA*01:01  1.087711e-01
#> 387   DRA*01:02  1.087711e-01
#> 388  DRB1*01:01  4.950361e-02
#> 389  DRB1*01:02  5.423947e-03
#> 390  DRB1*01:03  8.863330e-01
#> 391  DRB1*03:01  1.854270e-03
#> 392  DRB1*03:02  9.494700e-01
#> 393  DRB1*03:07  9.873786e-01
#> 394  DRB1*04:01  3.390179e-01
#> 395  DRB1*04:02  4.042862e-03
#> 396  DRB1*04:03  2.265170e-01
#> 397  DRB1*04:04  5.670464e-01
#> 398  DRB1*04:05  6.401594e-02
#> 399  DRB1*04:06  8.989411e-01
#> 400  DRB1*04:07  1.917060e-04
#> 401  DRB1*04:08  9.115621e-01
#> 402  DRB1*04:10  9.747476e-01
#> 403  DRB1*04:13  9.873786e-01
#> 404  DRB1*04:92  9.873786e-01
#> 405  DRB1*07:01  1.105835e-01
#> 406  DRB1*08:01  1.350792e-01
#> 407  DRB1*08:02  8.235979e-01
#> 408  DRB1*08:03  7.739671e-01
#> 409  DRB1*08:04  8.863330e-01
#> 410  DRB1*08:06  9.747476e-01
#> 411  DRB1*08:11  9.873786e-01
#> 412  DRB1*08:18  9.873786e-01
#> 413  DRB1*08:40  9.873786e-01
#> 414  DRB1*08:77  9.115621e-01
#> 415  DRB1*09:01  5.475507e-14
#> 416  DRB1*10:01  7.250487e-01
#> 417  DRB1*11:01  5.169113e-01
#> 418  DRB1*11:02  8.989411e-01
#> 419  DRB1*11:03  8.486181e-01
#> 420  DRB1*11:04  1.899495e-01
#> 421  DRB1*11:14  9.873786e-01
#> 422  DRB1*12:01  5.402271e-01
#> 423  DRB1*12:02  1.120573e-03
#> 424  DRB1*12:10  9.747476e-01
#> 425  DRB1*13:01  1.537580e-01
#> 426  DRB1*13:02  9.882807e-01
#> 427  DRB1*13:03  6.770363e-01
#> 428  DRB1*13:05  9.494700e-01
#> 429  DRB1*13:12  9.747476e-01
#> 430  DRB1*13:27  9.873786e-01
#> 431  DRB1*14:01  9.115621e-01
#> 432  DRB1*14:02  9.621104e-01
#> 433  DRB1*14:03  9.747476e-01
#> 434  DRB1*14:04  9.621104e-01
#> 435  DRB1*14:05  9.747476e-01
#> 436  DRB1*14:06  9.241927e-01
#> 437  DRB1*14:07  9.873786e-01
#> 438  DRB1*14:18  9.873786e-01
#> 439  DRB1*14:54  5.186921e-01
#> 440  DRB1*15:01  4.817173e-03
#> 441  DRB1*15:02  5.511430e-01
#> 442  DRB1*15:03  4.292086e-56
#> 443  DRB1*16:01  7.385183e-03
#> 444  DRB1*16:02  9.368297e-01
#> 445  DRB1*16:15  8.737410e-01
#> 446  DRB1*16:32  9.873786e-01
#> 447  DRB1*16:42  9.873786e-01

# get HWE in groups defined by disease status
# grouping by `disease == 1` will divide samples into two groups:
# `disease == 1` and `not disease == 1`
HWETest(midas, experiment = "hla_alleles", HWE_group = disease == 1)

#>             var  disease == 1 not disease == 1
#> 1       A*01:01  4.458834e-01     4.510815e-01
#> 2       A*01:02  9.642580e-01              NaN
#> 3      A*01:234           NaN     9.821424e-01
#> 4      A*01:237           NaN     9.821424e-01
#> 5      A*01:274  9.642580e-01     9.642580e-01
#> 6       A*01:43  9.821424e-01              NaN
#> 7       A*01:98           NaN     9.821424e-01
#> 8       A*02:01  8.667941e-01     8.010079e-01
#> 9       A*02:02  9.821424e-01     9.821424e-01
#> 10      A*02:03  9.463557e-01     9.284446e-01
#> 11      A*02:05  6.877874e-19     7.859394e-01
#> 12      A*02:06  1.418316e-03     1.501066e-06
#> 13      A*02:07  9.463557e-01     9.284446e-01
#> 14      A*02:08           NaN     9.642580e-01
#> 15      A*02:10  9.821424e-01     9.642580e-01
#> 16     A*02:211  9.821424e-01              NaN
#> 17      A*02:22           NaN     9.821424e-01
#> 18     A*02:324           NaN     9.821424e-01
#> 19      A*02:60           NaN     9.821424e-01
#> 20     A*02:614           NaN     9.821424e-01
#> 21     A*02:642           NaN     9.821424e-01
#> 22     A*02:686           NaN     9.821424e-01
#> 23     A*02:753  9.821424e-01              NaN
#> 24     A*02:763           NaN     9.821424e-01
#> 25      A*03:01  7.345401e-01     5.095385e-01
#> 26      A*03:02  8.926325e-01     8.747498e-01
#> 27     A*03:205           NaN     9.821424e-01
#> 28      A*03:24  9.821424e-01              NaN
#> 29     A*03:312  9.821424e-01     9.642580e-01
#> 30      A*03:72  9.821424e-01     9.821424e-01
#> 31      A*11:01  8.474123e-01     3.360577e-02
#> 32      A*11:02           NaN     9.821424e-01
#> 33      A*11:03           NaN     9.821424e-01
#> 34      A*11:12           NaN     9.821424e-01
#> 35     A*11:306  9.642580e-01     9.821424e-01
#> 36      A*11:43           NaN     9.821424e-01
#> 37     A*11:50Q           NaN     9.821424e-01
#> 38      A*11:88  9.642580e-01     9.821424e-01
#> 39      A*23:01  6.314785e-01     5.505762e-01
#> 40      A*23:17           NaN     9.821424e-01
#> 41      A*23:18  9.821424e-01              NaN
#> 42      A*24:02  2.687114e-01     5.060263e-01
#> 43      A*24:03  9.821424e-01     9.821424e-01
#> 44      A*24:04           NaN     9.821424e-01
#> 45      A*24:07 9.505398e-111              NaN
#> 46     A*24:104  9.821424e-01     9.821424e-01
#> 47      A*24:20  9.642580e-01     9.642580e-01
#> 48      A*24:50  9.463557e-01     9.821424e-01
#> 49      A*24:95  9.821424e-01              NaN
#> 50      A*25:01  6.149636e-01     4.892090e-01
#> 51      A*25:09           NaN     9.642580e-01
#> 52      A*25:25           NaN     9.821424e-01
#> 53      A*26:01  3.034080e-01     3.370681e-01
#> 54      A*26:02  9.821424e-01     9.463557e-01
#> 55      A*26:03  9.821424e-01     9.821424e-01
#> 56      A*26:08  9.821424e-01     9.642580e-01
#> 57      A*26:16  9.821424e-01              NaN
#> 58      A*29:01  9.821424e-01     9.105338e-01
#> 59      A*29:02  6.481452e-01     2.071750e-01
#> 60     A*29:119  9.821424e-01     9.642580e-01
#> 61      A*30:01  6.989742e-01     6.481452e-01
#> 62      A*30:02  1.151375e-01     6.149636e-01
#> 63      A*30:04  9.821424e-01     9.642580e-01
#> 64      A*31:01  4.036917e-01     6.149636e-01
#> 65      A*31:08  9.821424e-01              NaN
#> 66      A*31:21           NaN     9.821424e-01
#> 67      A*32:01  5.505762e-01     4.597880e-01
#> 68      A*32:02           NaN     9.821424e-01
#> 69     A*32:116  9.463557e-01     9.821424e-01
#> 70      A*33:01  7.508685e-01     8.926325e-01
#> 71      A*33:03  8.747498e-01     8.390774e-01
#> 72      A*33:05           NaN     9.642580e-01
#> 73     A*33:125           NaN     9.821424e-01
#> 74     A*33:163           NaN     9.821424e-01
#> 75      A*33:51           NaN     9.821424e-01
#> 76      A*34:01  9.821424e-01              NaN
#> 77      A*34:02  9.105338e-01     9.821424e-01
#> 78      A*66:01  9.105338e-01     9.105338e-01
#> 79      A*68:01  2.071750e-01     5.194880e-01
#> 80      A*68:02  7.859394e-01     8.390774e-01
#> 81      A*68:04           NaN     9.821424e-01
#> 82      A*68:16  9.821424e-01              NaN
#> 83     A*68:164           NaN     9.821424e-01
#> 84      A*68:17  9.821424e-01              NaN
#> 85      A*68:23  9.821424e-01              NaN
#> 86      A*68:66  9.821424e-01              NaN
#> 87      A*69:01  9.463557e-01     9.463557e-01
#> 88      A*74:01           NaN     9.642580e-01
#> 89      B*07:02  2.143342e-02     9.714144e-01
#> 90      B*07:04  9.821424e-01              NaN
#> 91      B*07:05  9.821424e-01     9.284446e-01
#> 92      B*07:06  9.642580e-01     9.463557e-01
#> 93      B*07:10  9.821424e-01              NaN
#> 94      B*07:14           NaN     9.821424e-01
#> 95      B*07:22  9.821424e-01              NaN
#> 96      B*07:47  9.821424e-01              NaN
#> 97      B*08:01  7.185338e-01     1.585004e-01
#> 98     B*08:156  9.821424e-01     9.821424e-01
#> 99      B*08:94           NaN     9.821424e-01
#> 100     B*13:01  9.642580e-01     9.463557e-01
#> 101     B*13:02  2.438194e-03     7.337845e-01
#> 102    B*13:117           NaN     9.821424e-01
#> 103     B*14:01  7.859394e-01     8.213061e-01
#> 104     B*14:02  5.824204e-01     7.161655e-01
#> 105     B*14:28           NaN     9.821424e-01
#> 106     B*15:01  1.404105e-02     3.861508e-02
#> 107     B*15:02  9.821424e-01     9.105338e-01
#> 108     B*15:03  9.463557e-01     9.105338e-01
#> 109     B*15:07  9.642580e-01              NaN
#> 110     B*15:09  9.642580e-01              NaN
#> 111     B*15:10  9.284446e-01              NaN
#> 112     B*15:11  9.821424e-01     9.642580e-01
#> 113     B*15:12  9.821424e-01              NaN
#> 114     B*15:15  9.821424e-01              NaN
#> 115     B*15:17  9.105338e-01     9.105338e-01
#> 116     B*15:18           NaN     9.463557e-01
#> 117     B*15:20  9.821424e-01              NaN
#> 118    B*15:212           NaN     9.821424e-01
#> 119    B*15:220  9.821424e-01              NaN
#> 120     B*15:25  9.642580e-01     9.821424e-01
#> 121    B*15:256  9.821424e-01              NaN
#> 122     B*15:27           NaN     9.821424e-01
#> 123     B*15:39           NaN     9.821424e-01
#> 124     B*15:63           NaN     9.821424e-01
#> 125     B*18:01  3.041245e-01     3.922438e-01
#> 126     B*18:02  9.821424e-01              NaN
#> 127     B*18:03           NaN     9.821424e-01
#> 128     B*18:07           NaN     9.821424e-01
#> 129     B*27:02  9.284446e-01     9.105338e-01
#> 130     B*27:03  9.821424e-01     9.642580e-01
#> 131     B*27:04  9.821424e-01     9.821424e-01
#> 132     B*27:05  3.335070e-03     5.232419e-01
#> 133     B*27:06  9.821424e-01              NaN
#> 134    B*27:127  9.642580e-01              NaN
#> 135     B*27:14  9.821424e-01              NaN
#> 136     B*35:01  4.454132e-01     1.261788e-02
#> 137     B*35:02  9.463557e-01     1.124194e-04
#> 138     B*35:03  7.508685e-01     5.664072e-01
#> 139     B*35:05 9.505398e-111              NaN
#> 140     B*35:08  9.284446e-01     8.035904e-01
#> 141     B*35:09  9.821424e-01              NaN
#> 142     B*35:17           NaN     9.821424e-01
#> 143    B*35:194  9.821424e-01              NaN
#> 144     B*35:41  9.821424e-01              NaN
#> 145     B*35:42  9.821424e-01              NaN
#> 146     B*35:43  9.821424e-01              NaN
#> 147     B*35:60           NaN     9.821424e-01
#> 148     B*35:67  9.821424e-01              NaN
#> 149     B*35:71           NaN     9.821424e-01
#> 150     B*37:01  7.683624e-01     7.859394e-01
#> 151     B*37:68           NaN     9.821424e-01
#> 152     B*38:01  2.071750e-01     3.177626e-01
#> 153     B*38:02  9.463557e-01     9.642580e-01
#> 154     B*39:01  8.035904e-01     6.819014e-01
#> 155     B*39:05  9.821424e-01              NaN
#> 156     B*39:06  9.284446e-01     8.926325e-01
#> 157     B*39:09  9.821424e-01              NaN
#> 158     B*40:01  4.743884e-01     2.793337e-01
#> 159     B*40:02  8.390774e-01     6.649556e-01
#> 160     B*40:06  1.501066e-06     8.568950e-01
#> 161    B*40:117  9.821424e-01              NaN
#> 162     B*40:21           NaN     9.821424e-01
#> 163     B*40:32  9.821424e-01              NaN
#> 164    B*40:353  9.821424e-01              NaN
#> 165     B*40:49  9.821424e-01              NaN
#> 166     B*41:01  9.642580e-01     8.747498e-01
#> 167     B*41:02  9.105338e-01     9.284446e-01
#> 168     B*42:01  9.284446e-01     9.821424e-01
#> 169     B*42:02  9.642580e-01              NaN
#> 170     B*44:02  3.494980e-01     3.545424e-01
#> 171     B*44:03  2.140040e-01     2.920457e-01
#> 172     B*44:04           NaN     9.821424e-01
#> 173     B*44:05  9.463557e-01     9.642580e-01
#> 174     B*44:06  9.821424e-01              NaN
#> 175     B*44:27  9.463557e-01     8.747498e-01
#> 176    B*44:270  9.642580e-01              NaN
#> 177     B*44:37           NaN     9.821424e-01
#> 178     B*44:55  9.821424e-01              NaN
#> 179     B*45:01  8.747498e-01     8.568950e-01
#> 180     B*46:01  9.284446e-01     8.747498e-01
#> 181     B*47:01  9.105338e-01     9.821424e-01
#> 182     B*47:04  9.642580e-01              NaN
#> 183     B*48:01  8.568950e-01     4.143911e-50
#> 184     B*49:01  6.314785e-01     1.151375e-01
#> 185     B*50:01  7.334665e-01     7.161655e-01
#> 186     B*50:02           NaN     9.821424e-01
#> 187     B*51:01  4.417170e-01     5.180541e-01
#> 188     B*51:02  9.821424e-01              NaN
#> 189     B*51:05  9.821424e-01              NaN
#> 190     B*51:07           NaN     9.284446e-01
#> 191     B*51:08  9.821424e-01     9.105338e-01
#> 192     B*51:09  9.821424e-01              NaN
#> 193   B*51:173Q  9.821424e-01              NaN
#> 194    B*51:203           NaN     9.821424e-01
#> 195    B*51:230  9.821424e-01              NaN
#> 196     B*52:01  6.314785e-01     1.151375e-01
#> 197     B*53:01  9.284446e-01     9.284446e-01
#> 198     B*54:01  8.747498e-01     9.105338e-01
#> 199     B*55:01  7.683624e-01     4.926837e-02
#> 200     B*55:02  9.105338e-01     9.463557e-01
#> 201     B*56:01  8.568950e-01     8.390774e-01
#> 202     B*56:03  9.821424e-01              NaN
#> 203     B*56:09           NaN     9.821424e-01
#> 204     B*57:01  5.508523e-01     1.151375e-01
#> 205     B*57:02  8.926325e-01     9.821424e-01
#> 206     B*57:03  8.926325e-01              NaN
#> 207     B*58:01  8.747498e-01     7.334665e-01
#> 208     B*58:02  9.463557e-01     9.642580e-01
#> 209     B*59:01           NaN     9.642580e-01
#> 210     B*67:01  9.821424e-01              NaN
#> 211     B*73:01           NaN     9.642580e-01
#> 212     C*01:02  5.566223e-01     2.595610e-01
#> 213     C*01:03           NaN     9.642580e-01
#> 214    C*01:103           NaN     9.821424e-01
#> 215     C*01:22           NaN     9.821424e-01
#> 216     C*01:51  9.821424e-01              NaN
#> 217     C*02:02  9.896214e-01     9.346065e-01
#> 218     C*02:06           NaN     9.821424e-01
#> 219     C*02:10  9.463557e-01     9.105338e-01
#> 220     C*02:20           NaN     9.821424e-01
#> 221     C*03:02  9.284446e-01     8.390774e-01
#> 222     C*03:03  5.655567e-01     4.065879e-01
#> 223     C*03:04  9.528671e-01     1.939416e-04
#> 224     C*03:05           NaN     9.821424e-01
#> 225     C*03:06  9.821424e-01              NaN
#> 226    C*03:344  9.463557e-01     9.821424e-01
#> 227     C*03:39  9.821424e-01     9.642580e-01
#> 228     C*03:85           NaN     9.821424e-01
#> 229     C*04:01  7.750978e-01     1.382501e-02
#> 230     C*04:03  9.821424e-01     9.642580e-01
#> 231    C*04:277           NaN     9.821424e-01
#> 232     C*05:01  5.146674e-01     1.170605e-01
#> 233     C*05:05  9.821424e-01              NaN
#> 234     C*05:37  9.821424e-01              NaN
#> 235     C*06:02  2.661323e-01     8.800755e-01
#> 236     C*06:09           NaN     9.821424e-01
#> 237     C*07:01  6.643130e-01     6.540518e-01
#> 238     C*07:02  7.747196e-01     3.607264e-01
#> 239     C*07:04  7.508685e-01     6.481452e-01
#> 240     C*07:18  9.463557e-01     8.747498e-01
#> 241    C*07:204  9.821424e-01              NaN
#> 242     C*07:51  9.642580e-01              NaN
#> 243    C*07:578  9.284446e-01     9.463557e-01
#> 244    C*07:615           NaN     9.821424e-01
#> 245    C*07:696  9.821424e-01     9.821424e-01
#> 246     C*08:01  8.213061e-01     7.859394e-01
#> 247     C*08:02  2.920457e-01     2.793337e-01
#> 248     C*08:03  8.926325e-01     4.143911e-50
#> 249     C*12:02  6.481452e-01     1.432546e-01
#> 250     C*12:03  5.785555e-02     4.101504e-01
#> 251    C*12:132  9.821424e-01              NaN
#> 252    C*12:167           NaN     9.821424e-01
#> 253     C*12:28  9.821424e-01              NaN
#> 254     C*14:02  7.508685e-01     7.334665e-01
#> 255     C*14:03  9.463557e-01     9.642580e-01
#> 256     C*15:02  5.986083e-01     5.824204e-01
#> 257     C*15:05  9.284446e-01     8.568950e-01
#> 258     C*15:07  9.821424e-01              NaN
#> 259     C*15:09  9.821424e-01              NaN
#> 260     C*16:01  5.824204e-01     5.664072e-01
#> 261     C*16:02  9.284446e-01     7.859394e-01
#> 262     C*16:04  9.821424e-01     9.463557e-01
#> 263     C*17:01  9.105338e-01     8.926325e-01
#> 264     C*17:03  8.926325e-01     9.284446e-01
#> 265     C*18:02  9.463557e-01              NaN
#> 266  DPA1*01:03  3.031189e-01     2.951598e-04
#> 267  DPA1*01:04  8.390774e-01     9.463557e-01
#> 268  DPA1*01:05           NaN     9.821424e-01
#> 269  DPA1*01:11           NaN     9.821424e-01
#> 270  DPA1*02:01  5.853393e-01     2.771441e-02
#> 271  DPA1*02:02  1.435600e-13     9.804817e-13
#> 272  DPA1*02:06  9.284446e-01     9.105338e-01
#> 273  DPA1*02:07  8.568950e-01     9.105338e-01
#> 274  DPA1*03:01  9.284446e-01     9.284446e-01
#> 275  DPA1*04:01  9.821424e-01              NaN
#> 276  DPB1*01:01  6.501129e-01     7.337845e-01
#> 277  DPB1*02:01  8.533282e-01     3.342930e-02
#> 278  DPB1*02:02  8.035904e-01     4.926837e-02
#> 279  DPB1*03:01  5.888546e-01     7.572197e-01
#> 280  DPB1*04:01  2.970852e-02     4.586521e-04
#> 281  DPB1*04:02  3.743689e-01     1.588829e-01
#> 282  DPB1*05:01  3.024316e-17     8.416915e-02
#> 283  DPB1*06:01  8.568950e-01     1.124194e-04
#> 284  DPB1*09:01  8.035904e-01     8.390774e-01
#> 285  DPB1*10:01  8.035904e-01     7.508685e-01
#> 286 DPB1*104:01  6.819014e-01     1.721583e-02
#> 287 DPB1*105:01  8.568950e-01     8.926325e-01
#> 288 DPB1*107:01  9.821424e-01              NaN
#> 289  DPB1*11:01  6.481452e-01     6.481452e-01
#> 290 DPB1*113:01           NaN     9.642580e-01
#> 291 DPB1*124:01  8.747498e-01     9.105338e-01
#> 292 DPB1*126:01  9.105338e-01     9.642580e-01
#> 293  DPB1*13:01  6.314785e-01     1.721583e-02
#> 294 DPB1*131:01  9.821424e-01     9.821424e-01
#> 295 DPB1*133:01           NaN     9.821424e-01
#> 296 DPB1*135:01  9.642580e-01     9.284446e-01
#> 297 DPB1*138:01  9.821424e-01              NaN
#> 298  DPB1*14:01  8.995742e-02     6.819014e-01
#> 299  DPB1*15:01  8.390774e-01     9.284446e-01
#> 300 DPB1*150:01  9.821424e-01              NaN
#> 301 DPB1*158:01  9.821424e-01              NaN
#> 302  DPB1*16:01  9.284446e-01     8.926325e-01
#> 303 DPB1*162:01  9.463557e-01     9.642580e-01
#> 304  DPB1*17:01  8.995742e-02     3.177626e-01
#> 305  DPB1*18:01  9.821424e-01              NaN
#> 306  DPB1*19:01  9.284446e-01     9.463557e-01
#> 307  DPB1*20:01  9.284446e-01     9.463557e-01
#> 308  DPB1*21:01  9.821424e-01     9.463557e-01
#> 309  DPB1*22:01           NaN     9.821424e-01
#> 310  DPB1*23:01  8.390774e-01     8.926325e-01
#> 311 DPB1*234:01           NaN     9.821424e-01
#> 312 DPB1*259:01  9.821424e-01              NaN
#> 313 DPB1*271:01  9.642580e-01     9.642580e-01
#> 314 DPB1*294:01  9.821424e-01              NaN
#> 315  DPB1*33:01  9.821424e-01              NaN
#> 316 DPB1*350:01  8.926325e-01     9.821424e-01
#> 317 DPB1*351:01  9.821424e-01     9.642580e-01
#> 318  DPB1*36:01           NaN     9.821424e-01
#> 319  DPB1*38:01           NaN     9.821424e-01
#> 320 DPB1*398:01  9.821424e-01              NaN
#> 321  DPB1*40:01  9.821424e-01              NaN
#> 322  DPB1*41:01           NaN     9.821424e-01
#> 323 DPB1*414:01  9.284446e-01     9.463557e-01
#> 324  DPB1*45:01  9.821424e-01              NaN
#> 325 DPB1*461:01  9.821424e-01              NaN
#> 326 DPB1*462:01           NaN     9.821424e-01
#> 327 DPB1*463:01  9.821424e-01     9.463557e-01
#> 328  DPB1*47:01  9.821424e-01              NaN
#> 329  DPB1*51:01  9.821424e-01              NaN
#> 330 DPB1*519:01  9.821424e-01              NaN
#> 331 DPB1*547:01  9.821424e-01              NaN
#> 332  DPB1*57:01  9.821424e-01     9.642580e-01
#> 333 DPB1*651:01  9.284446e-01              NaN
#> 334 DPB1*702:01  9.284446e-01     9.463557e-01
#> 335 DPB1*744:01  9.821424e-01              NaN
#> 336  DPB1*80:01  9.821424e-01              NaN
#> 337  DPB1*88:01           NaN     9.821424e-01
#> 338  DPB1*92:01  9.821424e-01              NaN
#> 339  DQA1*01:01  7.763217e-01     8.132882e-02
#> 340  DQA1*01:02  1.561554e-01     2.362403e-01
#> 341  DQA1*01:03  4.910020e-01     9.273485e-02
#> 342  DQA1*01:04  5.986083e-01     5.042442e-01
#> 343  DQA1*01:05  8.035904e-01     7.508685e-01
#> 344  DQA1*01:06  9.821424e-01              NaN
#> 345  DQA1*01:10  9.821424e-01     9.821424e-01
#> 346  DQA1*02:01  1.403811e-02     6.348000e-01
#> 347  DQA1*03:01  6.270325e-01     6.004429e-01
#> 348  DQA1*03:02  3.398993e-02     1.418316e-03
#> 349  DQA1*03:03  8.972508e-01     3.927654e-01
#> 350  DQA1*04:01  7.337845e-01     6.481452e-01
#> 351  DQA1*04:02           NaN     9.642580e-01
#> 352  DQA1*05:01  5.814609e-01     2.051846e-03
#> 353  DQA1*05:03  8.926325e-01     7.508685e-01
#> 354  DQA1*05:05  6.956462e-01     4.051686e-01
#> 355  DQA1*05:06           NaN     9.821424e-01
#> 356  DQA1*05:07  9.821424e-01              NaN
#> 357  DQA1*05:08  9.821424e-01              NaN
#> 358  DQA1*05:09  9.463557e-01     9.463557e-01
#> 359  DQA1*06:01  1.124194e-04     8.035904e-01
#> 360  DQB1*02:01  4.390796e-01     2.051846e-03
#> 361  DQB1*02:02  3.260894e-05     4.786006e-01
#> 362  DQB1*02:10  9.821424e-01              NaN
#> 363  DQB1*03:01  5.645883e-02     2.530537e-01
#> 364  DQB1*03:02  6.270325e-01     4.417170e-01
#> 365  DQB1*03:03  3.041245e-01     4.454132e-01
#> 366  DQB1*03:04  9.105338e-01     9.642580e-01
#> 367  DQB1*03:05  9.463557e-01              NaN
#> 368  DQB1*03:19  8.747498e-01     9.284446e-01
#> 369  DQB1*03:22  8.213061e-01     7.859394e-01
#> 370 DQB1*03:251  9.642580e-01              NaN
#> 371  DQB1*03:96  9.821424e-01     9.284446e-01
#> 372  DQB1*04:01  2.450294e-11     7.683624e-01
#> 373  DQB1*04:02  4.173604e-01     5.986083e-01
#> 374  DQB1*04:23  9.642580e-01     9.821424e-01
#> 375  DQB1*05:01  4.448428e-01     3.800165e-01
#> 376  DQB1*05:02  3.335070e-03     8.956533e-01
#> 377  DQB1*05:03  6.314785e-01     5.349342e-01
#> 378  DQB1*05:04  9.821424e-01     9.463557e-01
#> 379 DQB1*05:107  9.821424e-01              NaN
#> 380  DQB1*06:01  5.505762e-01     5.524207e-04
#> 381  DQB1*06:02  8.444182e-01     1.531837e-03
#> 382  DQB1*06:03  7.547216e-01     2.257310e-01
#> 383  DQB1*06:04  4.036917e-01     5.655567e-01
#> 384  DQB1*06:07  9.821424e-01              NaN
#> 385  DQB1*06:09  8.390774e-01     8.390774e-01
#> 386   DRA*01:01  2.789248e-01     3.601145e-01
#> 387   DRA*01:02  2.789248e-01     3.601145e-01
#> 388  DRB1*01:01  8.739948e-01     1.473929e-02
#> 389  DRB1*01:02  1.151375e-01     1.325122e-02
#> 390  DRB1*01:03  8.568950e-01     9.821424e-01
#> 391  DRB1*03:01  3.743689e-01     1.602258e-03
#> 392  DRB1*03:02  9.284446e-01              NaN
#> 393  DRB1*03:07           NaN     9.821424e-01
#> 394  DRB1*04:01  6.884533e-01     5.841735e-02
#> 395  DRB1*04:02  1.501066e-06     8.035904e-01
#> 396  DRB1*04:03  1.151375e-01     7.508685e-01
#> 397  DRB1*04:04  6.149636e-01     2.423752e-01
#> 398  DRB1*04:05  3.335070e-03     5.986083e-01
#> 399  DRB1*04:06  9.642580e-01     8.926325e-01
#> 400  DRB1*04:07  8.213061e-01     1.790460e-13
#> 401  DRB1*04:08  9.642580e-01     9.105338e-01
#> 402  DRB1*04:10  9.642580e-01              NaN
#> 403  DRB1*04:13  9.821424e-01              NaN
#> 404  DRB1*04:92  9.821424e-01              NaN
#> 405  DRB1*07:01  3.160062e-03     6.348000e-01
#> 406  DRB1*08:01  4.926837e-02     7.683624e-01
#> 407  DRB1*08:02  8.213061e-01     9.284446e-01
#> 408  DRB1*08:03  8.390774e-01     8.390774e-01
#> 409  DRB1*08:04  8.747498e-01     9.642580e-01
#> 410  DRB1*08:06  9.821424e-01     9.821424e-01
#> 411  DRB1*08:11           NaN     9.821424e-01
#> 412  DRB1*08:18           NaN     9.821424e-01
#> 413  DRB1*08:40           NaN     9.821424e-01
#> 414  DRB1*08:77  9.284446e-01     9.463557e-01
#> 415  DRB1*09:01  1.672301e-10     4.644607e-04
#> 416  DRB1*10:01  8.390774e-01     7.683624e-01
#> 417  DRB1*11:01  7.147688e-01     6.655523e-01
#> 418  DRB1*11:02  9.284446e-01     9.284446e-01
#> 419  DRB1*11:03  8.926325e-01     8.926325e-01
#> 420  DRB1*11:04  2.725558e-01     4.429119e-01
#> 421  DRB1*11:14  9.821424e-01              NaN
#> 422  DRB1*12:01  6.819014e-01     6.481452e-01
#> 423  DRB1*12:02  1.124194e-04     8.747498e-01
#> 424  DRB1*12:10           NaN     9.642580e-01
#> 425  DRB1*13:01  6.748768e-01     9.942882e-02
#> 426  DRB1*13:02  2.809499e-01     2.424790e-01
#> 427  DRB1*13:03  7.508685e-01     7.859394e-01
#> 428  DRB1*13:05  9.642580e-01     9.642580e-01
#> 429  DRB1*13:12           NaN     9.642580e-01
#> 430  DRB1*13:27           NaN     9.821424e-01
#> 431  DRB1*14:01  9.284446e-01     9.463557e-01
#> 432  DRB1*14:02  9.642580e-01     9.821424e-01
#> 433  DRB1*14:03  9.821424e-01     9.821424e-01
#> 434  DRB1*14:04  9.463557e-01              NaN
#> 435  DRB1*14:05  9.642580e-01              NaN
#> 436  DRB1*14:06  9.642580e-01     9.284446e-01
#> 437  DRB1*14:07           NaN     9.821424e-01
#> 438  DRB1*14:18           NaN     9.821424e-01
#> 439  DRB1*14:54  7.334665e-01     5.664072e-01
#> 440  DRB1*15:01  3.695519e-01     1.097330e-03
#> 441  DRB1*15:02  6.819014e-01     6.649556e-01
#> 442  DRB1*15:03  9.642580e-01    9.505398e-111
#> 443  DRB1*16:01  3.475184e-03     2.071750e-01
#> 444  DRB1*16:02  9.642580e-01     9.463557e-01
#> 445  DRB1*16:15  9.821424e-01     8.390774e-01
#> 446  DRB1*16:32  9.821424e-01              NaN
#> 447  DRB1*16:42  9.821424e-01              NaN

# filter MiDAS object by HWE test p-value
HWETest(midas, experiment = "hla_alleles", HWE_cutoff = 0.05, as.MiDAS = TRUE)

#> A MiDAS object of 1 listed
#>  experiment with a user-defined name and respective class.
#>  Containing an ExperimentList class object of length 1:
#>  [1] hla_alleles: SummarizedExperiment with 412 rows and 1000 columns
#> Functionality:
#>  experiments() - obtain the ExperimentList instance
#>  colData() - the primary/phenotype DataFrame
#>  sampleMap() - the sample coordination DataFrame
#>  `$`, `[`, `[[` - extract colData columns, subset, or experiment
#>  *Format() - convert into a long or wide DataFrame
#>  assays() - convert ExperimentList to a SimpleList of matrices
#>  exportClass() - save all data to files