MiDAS class

The MiDAS class is a MultiAssayExperiment object containing data and metadata required for MiDAS analysis.

Valid MiDAS object must have unique features names across all experiments and colData. It's metadata list needs to have a placeholder element, which is a string specifying name of column in colData used when defining statistical model for downstream analyses (see runMiDAS for more details). Optionally the object's metadata can also store 'hla_calls' and 'kir_calls' data frames (see prepareMiDAS for more details).

# S4 method for MiDAS
getExperiments(object)

# S4 method for MiDAS
getHlaCalls(object)

# S4 method for MiDAS
getKirCalls(object)

# S4 method for MiDAS
getPlaceholder(object)

# S4 method for MiDAS
getOmnibusGroups(object, experiment)

# S4 method for MiDAS
getFrequencies(
  object,
  experiment,
  carrier_frequency = FALSE,
  compare = FALSE,
  ref_pop = list(hla_alleles = c("USA NMDP African American pop 2", "USA NMDP Chinese",
    "USA NMDP European Caucasian", "USA NMDP Hispanic South or Central American",
    "USA NMDP Japanese", "USA NMDP North American Amerindian",
    "USA NMDP South Asian Indian"), kir_genes = c("USA California African American KIR",
    "USA California Asian American KIR", "USA California Caucasians KIR",
    "USA California Hispanic KIR")),
  ref = list(hla_alleles = allele_frequencies, kir_genes = kir_frequencies)
)

# S4 method for MiDAS
filterByFrequency(
  object,
  experiment,
  lower_frequency_cutoff = NULL,
  upper_frequency_cutoff = NULL,
  carrier_frequency = FALSE
)

# S4 method for MiDAS
filterByOmnibusGroups(object, experiment, groups)

# S4 method for MiDAS
filterByVariables(object, experiment, variables)

# S4 method for MiDAS
getAllelesForAA(object, aa_pos)

Arguments

object
MiDAS object.

experiment
String specifying experiment.

carrier_frequency
Logical flag indicating if carrier frequency should be returned.

compare
Logical flag indicating if hla_calls frequencies should be compared to reference frequencies given in ref.

ref_pop
Named list of character vectors giving names of reference populations in ref to compare with. Optionally vectors can be named, then those names will be used as population names. Each vector should correspond to a specific experiment.

ref
Named list of reference frequencies data frames. Each element should give reference for a specific experiment. See allele_frequencies for an example on how reference frequency data frame should be formatted.

lower_frequency_cutoff
Number giving lower frequency threshold. Numbers greater than 1 are interpreted as the number of feature occurrences, numbers between 0 and 1 as fractions.

upper_frequency_cutoff
Number giving upper frequency threshold. Numbers greater than 1 are interpreted as the number of feature occurrences, numbers between 0 and 1 as fractions.

groups
Character vector specifying omnibus groups to select. See getOmnibusGroups for more details.

variables
Character vector specifying features to select.

aa_pos
String specifying gene and amino acid position, example "A_9".

Value

Instance of class MiDAS