Calculate amino acid frequencies

getAAFrequencies calculates amino acid frequencies in amino acid data frame.

getAAFrequencies(aa_variation)

Arguments

aa_variation

Amino acid variation data frame as returned by hlaToAAVariation.

Value

Data frame with each row holding specific amino acid position, it's count and frequency.

Details

Both gene copies are taken into consideration for frequencies calculation, frequency = n / (2 * j) where n is the number of amino acid occurrences and j is the number of samples in aa_variation.

Examples

aa_variation <- hlaToAAVariation(MiDAS_tut_HLA)
getAAFrequencies(aa_variation)
#>          aa_pos Counts   Freq
#> 1       A_-21_*   1931 0.9655
#> 2       A_-21_V      2 0.0010
#> 3       A_-21_M     67 0.0335
#> 4       A_-18_*   1931 0.9655
#> 5       A_-18_R     68 0.0340
#> 6       A_-18_P      1 0.0005
#> 7       A_-16_*   1931 0.9655
#> 8       A_-16_L     68 0.0340
#> 9       A_-16_I      1 0.0005
#> 10      A_-15_*   1931 0.9655
#> 11      A_-15_V     38 0.0190
#> 12      A_-15_L     31 0.0155
#> 13      A_-11_*   1931 0.9655
#> 14      A_-11_S     57 0.0285
#> 15      A_-11_L     12 0.0060
#> 16       A_-5_*   1931 0.9655
#> 17       A_-5_T     68 0.0340
#> 18       A_-5_P      1 0.0005
#> 19        A_3_H   1996 0.9980
#> 20        A_3_Q      4 0.0020
#> 21        A_6_R   1999 0.9995
#> 22        A_6_G      1 0.0005
#> 23        A_9_Y    429 0.2145
#> 24        A_9_F   1037 0.5185
#> 25        A_9_S    379 0.1895
#> 26        A_9_T    155 0.0775
#> 27       A_12_V   1979 0.9895
#> 28       A_12_M     21 0.0105
#> 29       A_17_R   1914 0.9570
#> 30       A_17_S     86 0.0430
#> 31       A_19_E   1999 0.9995
#> 32       A_19_K      1 0.0005
#> 33       A_29_D   1999 0.9995
#> 34       A_29_A      1 0.0005
#> 35       A_43_Q   1978 0.9890
#> 36       A_43_R     22 0.0110
#> 37       A_44_R   1754 0.8770
#> 38       A_44_K    246 0.1230
#> 39       A_54_Q   1998 0.9990
#> 40       A_54_R      2 0.0010
#> 41       A_56_G   1857 0.9285
#> 42       A_56_R    143 0.0715
#> 43       A_62_R    299 0.1495
#> 44       A_62_G    555 0.2775
#> 45       A_62_E    294 0.1470
#> 46       A_62_Q    798 0.3990
#> 47       A_62_L     54 0.0270
#> 48       A_63_N    299 0.1495
#> 49       A_63_E   1647 0.8235
#> 50       A_63_Q     54 0.0270
#> 51       A_65_R   1706 0.8530
#> 52       A_65_G    294 0.1470
#> 53       A_66_N   1152 0.5760
#> 54       A_66_K    848 0.4240
#> 55       A_67_V   1754 0.8770
#> 56       A_67_M    246 0.1230
#> 57       A_70_H   1460 0.7300
#> 58       A_70_Q    540 0.2700
#> 59       A_73_T   1900 0.9500
#> 60       A_73_I    100 0.0500
#> 61       A_74_D   1443 0.7215
#> 62       A_74_H    557 0.2785
#> 63       A_76_E    458 0.2290
#> 64       A_76_V   1143 0.5715
#> 65       A_76_A    399 0.1995
#> 66       A_77_S    118 0.0590
#> 67       A_77_D   1143 0.5715
#> 68       A_77_N    739 0.3695
#> 69       A_79_R    411 0.2055
#> 70       A_79_G   1589 0.7945
#> 71       A_80_I    411 0.2055
#> 72       A_80_T   1589 0.7945
#> 73       A_81_A    411 0.2055
#> 74       A_81_L   1589 0.7945
#> 75       A_82_L    411 0.2055
#> 76       A_82_R   1589 0.7945
#> 77       A_83_R    411 0.2055
#> 78       A_83_G   1589 0.7945
#> 79       A_90_D    542 0.2710
#> 80       A_90_A   1458 0.7290
#> 81       A_95_I   1146 0.5730
#> 82       A_95_V    539 0.2695
#> 83       A_95_L    315 0.1575
#> 84       A_97_R    749 0.3745
#> 85       A_97_M    571 0.2855
#> 86       A_97_I    680 0.3400
#> 87       A_99_Y   1697 0.8485
#> 88       A_99_F    296 0.1480
#> 89       A_99_C      7 0.0035
#> 90      A_102_D   1994 0.9970
#> 91      A_102_H      6 0.0030
#> 92      A_103_V   1999 0.9995
#> 93      A_103_L      1 0.0005
#> 94      A_105_P    623 0.3115
#> 95      A_105_S   1377 0.6885
#> 96      A_107_G   1442 0.7210
#> 97      A_107_W    558 0.2790
#> 98      A_109_F   1935 0.9675
#> 99      A_109_L     65 0.0325
#> 100     A_114_Q    332 0.1660
#> 101     A_114_H    875 0.4375
#> 102     A_114_R    709 0.3545
#> 103     A_114_E     84 0.0420
#> 104     A_116_D   1036 0.5180
#> 105     A_116_Y    874 0.4370
#> 106     A_116_H     84 0.0420
#> 107     A_116_N      4 0.0020
#> 108     A_116_F      1 0.0005
#> 109     A_116_V      1 0.0005
#> 110     A_127_N   1066 0.5330
#> 111     A_127_K    934 0.4670
#> 112     A_142_I   1359 0.6795
#> 113     A_142_T    641 0.3205
#> 114     A_144_Q    525 0.2625
#> 115     A_144_K   1475 0.7375
#> 116     A_145_R   1359 0.6795
#> 117     A_145_H    641 0.3205
#> 118     A_149_T    177 0.0885
#> 119     A_149_A   1823 0.9115
#> 120     A_150_A   1754 0.8770
#> 121     A_150_V    246 0.1230
#> 122     A_151_H   1648 0.8240
#> 123     A_151_R    351 0.1755
#> 124     A_151_L      1 0.0005
#> 125     A_152_E    386 0.1930
#> 126     A_152_V   1163 0.5815
#> 127     A_152_A    369 0.1845
#> 128     A_152_R     45 0.0225
#> 129     A_152_W     37 0.0185
#> 130     A_156_W    276 0.1380
#> 131     A_156_L   1092 0.5460
#> 132     A_156_Q    385 0.1925
#> 133     A_156_R    247 0.1235
#> 134     A_158_A   1754 0.8770
#> 135     A_158_V    246 0.1230
#> 136     A_161_E   1781 0.8905
#> 137     A_161_D    219 0.1095
#> 138     A_163_R    535 0.2675
#> 139     A_163_T   1464 0.7320
#> 140     A_163_E      1 0.0005
#> 141     A_164_C   1999 0.9995
#> 142     A_164_F      1 0.0005
#> 143     A_166_E   1463 0.7315
#> 144     A_166_D    537 0.2685
#> 145     A_167_W   1463 0.7315
#> 146     A_167_G    537 0.2685
#> 147     A_171_Y   1977 0.9885
#> 148     A_171_H     23 0.0115
#> 149     A_182_T   1999 0.9995
#> 150     A_182_A      1 0.0005
#> 151     A_184_*   1910 0.9550
#> 152     A_184_A     44 0.0220
#> 153     A_184_P     46 0.0230
#> 154     A_186_*   1910 0.9550
#> 155     A_186_K     87 0.0435
#> 156     A_186_R      3 0.0015
#> 157     A_193_*   1910 0.9550
#> 158     A_193_A     50 0.0250
#> 159     A_193_P     40 0.0200
#> 160     A_194_*   1910 0.9550
#> 161     A_194_V     50 0.0250
#> 162     A_194_I     40 0.0200
#> 163     A_207_*   1910 0.9550
#> 164     A_207_S     50 0.0250
#> 165     A_207_G     40 0.0200
#> 166     A_219_*   1910 0.9550
#> 167     A_219_R     89 0.0445
#> 168     A_219_Q      1 0.0005
#> 169     A_231_*   1910 0.9550
#> 170     A_231_V     89 0.0445
#> 171     A_231_M      1 0.0005
#> 172     A_245_*   1910 0.9550
#> 173     A_245_A     85 0.0425
#> 174     A_245_V      5 0.0025
#> 175     A_246_*   1910 0.9550
#> 176     A_246_A     64 0.0320
#> 177     A_246_S     26 0.0130
#> 178     A_253_*   1910 0.9550
#> 179     A_253_Q     50 0.0250
#> 180     A_253_E     40 0.0200
#> 181     A_276_*   1931 0.9655
#> 182     A_276_P     48 0.0240
#> 183     A_276_L     21 0.0105
#> 184     A_282_*   1931 0.9655
#> 185     A_282_I     55 0.0275
#> 186     A_282_V     14 0.0070
#> 187     A_283_*   1931 0.9655
#> 188     A_283_P     68 0.0340
#> 189     A_283_H      1 0.0005
#> 190     A_284_*   1931 0.9655
#> 191     A_284_I     68 0.0340
#> 192     A_284_T      1 0.0005
#> 193     A_294_*   1931 0.9655
#> 194     A_294_F     34 0.0170
#> 195     A_294_L     35 0.0175
#> 196     A_297_*   1931 0.9655
#> 197     A_297_V     64 0.0320
#> 198     A_297_M      5 0.0025
#> 199     A_298_*   1931 0.9655
#> 200     A_298_I     56 0.0280
#> 201     A_298_F     13 0.0065
#> 202     A_299_*   1931 0.9655
#> 203     A_299_T     49 0.0245
#> 204     A_299_A     20 0.0100
#> 205     A_307_*   1931 0.9655
#> 206     A_307_M     56 0.0280
#> 207     A_307_R     13 0.0065
#> 208     A_311_*   1931 0.9655
#> 209     A_311_K     55 0.0275
#> 210     A_311_N     14 0.0070
#> 211     A_321_*   1931 0.9655
#> 212     A_321_S     48 0.0240
#> 213     A_321_T     21 0.0105
#> 214     A_334_*   1931 0.9655
#> 215     A_334_V     49 0.0245
#> 216     A_334_M     20 0.0100
#> 217     B_-23_*   1928 0.9640
#> 218     B_-23_R     58 0.0290
#> 219     B_-23_L     14 0.0070
#> 220     B_-21_*   1927 0.9635
#> 221     B_-21_T     58 0.0290
#> 222     B_-21_M     15 0.0075
#> 223     B_-16_*   1926 0.9630
#> 224     B_-16_V     48 0.0240
#> 225     B_-16_L     26 0.0130
#> 226     B_-11_*   1926 0.9630
#> 227     B_-11_S     26 0.0130
#> 228     B_-11_W     48 0.0240
#> 229     B_-10_*   1926 0.9630
#> 230     B_-10_G     57 0.0285
#> 231     B_-10_A     17 0.0085
#> 232      B_-9_*   1926 0.9630
#> 233      B_-9_A     73 0.0365
#> 234      B_-9_V      1 0.0005
#> 235      B_-8_*   1926 0.9630
#> 236      B_-8_L     29 0.0145
#> 237      B_-8_V     45 0.0225
#> 238       B_9_Y   1414 0.7070
#> 239       B_9_H    455 0.2275
#> 240       B_9_D    131 0.0655
#> 241      B_11_S    590 0.2950
#> 242      B_11_A   1410 0.7050
#> 243      B_12_V    630 0.3150
#> 244      B_12_M   1370 0.6850
#> 245      B_24_S    655 0.3275
#> 246      B_24_T    659 0.3295
#> 247      B_24_A    686 0.3430
#> 248      B_30_D   1868 0.9340
#> 249      B_30_G    132 0.0660
#> 250      B_32_Q   1422 0.7110
#> 251      B_32_L    578 0.2890
#> 252      B_33_F   1999 0.9995
#> 253      B_33_S      1 0.0005
#> 254      B_41_A   1420 0.7100
#> 255      B_41_T    580 0.2900
#> 256      B_45_E    637 0.3185
#> 257      B_45_K    500 0.2500
#> 258      B_45_M    303 0.1515
#> 259      B_45_T    548 0.2740
#> 260      B_45_G     12 0.0060
#> 261      B_46_E   1697 0.8485
#> 262      B_46_A    303 0.1515
#> 263      B_52_I   1988 0.9940
#> 264      B_52_V     12 0.0060
#> 265      B_59_Y   1997 0.9985
#> 266      B_59_H      3 0.0015
#> 267      B_62_R   1879 0.9395
#> 268      B_62_G    121 0.0605
#> 269      B_63_N   1014 0.5070
#> 270      B_63_E    986 0.4930
#> 271      B_65_Q   1869 0.9345
#> 272      B_65_R    131 0.0655
#> 273      B_66_I   1858 0.9290
#> 274      B_66_K     11 0.0055
#> 275      B_66_N    131 0.0655
#> 276      B_67_Y    259 0.1295
#> 277      B_67_S    909 0.4545
#> 278      B_67_F    450 0.2250
#> 279      B_67_C    251 0.1255
#> 280      B_67_M    131 0.0655
#> 281      B_69_A    456 0.2280
#> 282      B_69_T   1533 0.7665
#> 283      B_69_R     11 0.0055
#> 284      B_70_Q    257 0.1285
#> 285      B_70_N   1533 0.7665
#> 286      B_70_K     79 0.0395
#> 287      B_70_S    131 0.0655
#> 288      B_71_A    467 0.2335
#> 289      B_71_T   1533 0.7665
#> 290      B_72_Q   1999 0.9995
#> 291      B_72_R      1 0.0005
#> 292      B_74_D    571 0.2855
#> 293      B_74_Y   1429 0.7145
#> 294      B_76_E   1987 0.9935
#> 295      B_76_V     13 0.0065
#> 296      B_77_S   1098 0.5490
#> 297      B_77_N    803 0.4015
#> 298      B_77_D     97 0.0485
#> 299      B_77_G      2 0.0010
#> 300      B_80_N   1097 0.5485
#> 301      B_80_T    464 0.2320
#> 302      B_80_I    439 0.2195
#> 303      B_81_L   1197 0.5985
#> 304      B_81_A    803 0.4015
#> 305      B_82_R   1097 0.5485
#> 306      B_82_L    903 0.4515
#> 307      B_83_G   1097 0.5485
#> 308      B_83_R    903 0.4515
#> 309      B_90_A   1998 0.9990
#> 310      B_90_D      2 0.0010
#> 311      B_94_T   1422 0.7110
#> 312      B_94_I    578 0.2890
#> 313      B_95_L    929 0.4645
#> 314      B_95_W    467 0.2335
#> 315      B_95_I    604 0.3020
#> 316      B_97_S    358 0.1790
#> 317      B_97_R   1035 0.5175
#> 318      B_97_T    368 0.1840
#> 319      B_97_N     77 0.0385
#> 320      B_97_V     94 0.0470
#> 321      B_97_W     68 0.0340
#> 322      B_99_Y   1966 0.9830
#> 323      B_99_S     27 0.0135
#> 324      B_99_F      7 0.0035
#> 325     B_101_C   1999 0.9995
#> 326     B_101_Y      1 0.0005
#> 327     B_103_V   1559 0.7795
#> 328     B_103_L    439 0.2195
#> 329     B_103_M      2 0.0010
#> 330     B_109_L   1986 0.9930
#> 331     B_109_F     14 0.0070
#> 332     B_113_H   1437 0.7185
#> 333     B_113_Y    563 0.2815
#> 334     B_114_D    972 0.4860
#> 335     B_114_N    936 0.4680
#> 336     B_114_H     92 0.0460
#> 337     B_116_Y    674 0.3370
#> 338     B_116_L    242 0.1210
#> 339     B_116_S    488 0.2440
#> 340     B_116_F    230 0.1150
#> 341     B_116_D    366 0.1830
#> 342     B_118_Y   1999 0.9995
#> 343     B_118_C      1 0.0005
#> 344     B_131_R    445 0.2225
#> 345     B_131_S   1555 0.7775
#> 346     B_137_D   1999 0.9995
#> 347     B_137_N      1 0.0005
#> 348     B_143_T   1927 0.9635
#> 349     B_143_S     73 0.0365
#> 350     B_145_R   1920 0.9600
#> 351     B_145_L     80 0.0400
#> 352     B_147_W   1927 0.9635
#> 353     B_147_L     73 0.0365
#> 354     B_152_E    675 0.3375
#> 355     B_152_V   1325 0.6625
#> 356     B_156_R    193 0.0965
#> 357     B_156_L   1304 0.6520
#> 358     B_156_W    111 0.0555
#> 359     B_156_D    392 0.1960
#> 360     B_158_A   1900 0.9500
#> 361     B_158_T    100 0.0500
#> 362     B_163_E    453 0.2265
#> 363     B_163_L   1016 0.5080
#> 364     B_163_T    531 0.2655
#> 365     B_166_E   1999 0.9995
#> 366     B_166_D      1 0.0005
#> 367     B_167_W   1704 0.8520
#> 368     B_167_S    295 0.1475
#> 369     B_167_G      1 0.0005
#> 370     B_171_Y   1614 0.8070
#> 371     B_171_H    386 0.1930
#> 372     B_177_D    397 0.1985
#> 373     B_177_E   1603 0.8015
#> 374     B_178_K    243 0.1215
#> 375     B_178_T   1757 0.8785
#> 376     B_180_E    397 0.1985
#> 377     B_180_Q   1603 0.8015
#> 378     B_194_*   1907 0.9535
#> 379     B_194_I     76 0.0380
#> 380     B_194_V     17 0.0085
#> 381     B_199_*   1907 0.9535
#> 382     B_199_A     88 0.0440
#> 383     B_199_V      5 0.0025
#> 384     B_211_*   1907 0.9535
#> 385     B_211_A     92 0.0460
#> 386     B_211_G      1 0.0005
#> 387     B_239_*   1907 0.9535
#> 388     B_239_R     91 0.0455
#> 389     B_239_G      2 0.0010
#> 390     B_253_*   1907 0.9535
#> 391     B_253_E     91 0.0455
#> 392     B_253_Q      2 0.0010
#> 393     B_267_*   1907 0.9535
#> 394     B_267_P     91 0.0455
#> 395     B_267_Q      2 0.0010
#> 396     B_268_*   1907 0.9535
#> 397     B_268_K     91 0.0455
#> 398     B_268_E      2 0.0010
#> 399     B_270_*   1907 0.9535
#> 400     B_270_L     91 0.0455
#> 401     B_270_C      2 0.0010
#> 402     B_275_*   1937 0.9685
#> 403     B_275_E     61 0.0305
#> 404     B_275_K      2 0.0010
#> 405     B_282_*   1937 0.9685
#> 406     B_282_I     28 0.0140
#> 407     B_282_V     35 0.0175
#> 408     B_295_*   1937 0.9685
#> 409     B_295_.     61 0.0305
#> 410     B_295_V      2 0.0010
#> 411     B_296_*   1937 0.9685
#> 412     B_296_A     61 0.0305
#> 413     B_296_V      2 0.0010
#> 414     B_297_*   1937 0.9685
#> 415     B_297_V     61 0.0305
#> 416     B_297_T      2 0.0010
#> 417     B_299_*   1937 0.9685
#> 418     B_299_V     61 0.0305
#> 419     B_299_A      2 0.0010
#> 420     B_300_*   1937 0.9685
#> 421     B_300_I     61 0.0305
#> 422     B_300_V      2 0.0010
#> 423     B_301_*   1937 0.9685
#> 424     B_301_G     61 0.0305
#> 425     B_301_V      2 0.0010
#> 426     B_306_*   1937 0.9685
#> 427     B_306_T     26 0.0130
#> 428     B_306_A     37 0.0185
#> 429     B_326_*   1939 0.9695
#> 430     B_326_S     25 0.0125
#> 431     B_326_C     36 0.0180
#> 432     B_339_*   1939 0.9695
#> 433     B_339_A     60 0.0300
#> 434     B_339_D      1 0.0005
#> 435     C_-18_*   1944 0.9720
#> 436     C_-18_R     46 0.0230
#> 437     C_-18_Q     10 0.0050
#> 438     C_-17_*   1944 0.9720
#> 439     C_-17_T     28 0.0140
#> 440     C_-17_A     28 0.0140
#> 441     C_-15_*   1944 0.9720
#> 442     C_-15_I     15 0.0075
#> 443     C_-15_L     41 0.0205
#> 444      C_-9_*   1944 0.9720
#> 445      C_-9_A     40 0.0200
#> 446      C_-9_G     16 0.0080
#> 447      C_-5_*   1944 0.9720
#> 448      C_-5_T     46 0.0230
#> 449      C_-5_I     10 0.0050
#> 450      C_-1_*   1944 0.9720
#> 451      C_-1_A     46 0.0230
#> 452      C_-1_T     10 0.0050
#> 453       C_1_*   1944 0.9720
#> 454       C_1_C     41 0.0205
#> 455       C_1_G     15 0.0075
#> 456       C_6_R   1895 0.9475
#> 457       C_6_K    105 0.0525
#> 458       C_9_Y    895 0.4475
#> 459       C_9_D    718 0.3590
#> 460       C_9_F    105 0.0525
#> 461       C_9_S    282 0.1410
#> 462      C_11_A   1613 0.8065
#> 463      C_11_S    387 0.1935
#> 464      C_14_R   1752 0.8760
#> 465      C_14_W    248 0.1240
#> 466      C_16_G   1899 0.9495
#> 467      C_16_S    101 0.0505
#> 468      C_21_R   1642 0.8210
#> 469      C_21_H    358 0.1790
#> 470      C_24_A   1177 0.5885
#> 471      C_24_S    823 0.4115
#> 472      C_35_R   1736 0.8680
#> 473      C_35_Q    264 0.1320
#> 474      C_49_A   1752 0.8760
#> 475      C_49_E    248 0.1240
#> 476      C_66_K   1682 0.8410
#> 477      C_66_N    318 0.1590
#> 478      C_68_K   1999 0.9995
#> 479      C_68_N      1 0.0005
#> 480      C_73_A   1178 0.5890
#> 481      C_73_T    822 0.4110
#> 482      C_77_S   1194 0.5970
#> 483      C_77_N    806 0.4030
#> 484      C_80_N   1194 0.5970
#> 485      C_80_K    806 0.4030
#> 486      C_90_A   1031 0.5155
#> 487      C_90_D    969 0.4845
#> 488      C_91_G   1913 0.9565
#> 489      C_91_R     87 0.0435
#> 490      C_94_T   1749 0.8745
#> 491      C_94_I    251 0.1255
#> 492      C_95_L   1707 0.8535
#> 493      C_95_I    259 0.1295
#> 494      C_95_F     34 0.0170
#> 495      C_97_W    548 0.2740
#> 496      C_97_R   1451 0.7255
#> 497      C_97_S      1 0.0005
#> 498      C_99_Y   1379 0.6895
#> 499      C_99_C    105 0.0525
#> 500      C_99_F    288 0.1440
#> 501      C_99_S    228 0.1140
#> 502     C_103_L   1808 0.9040
#> 503     C_103_V    192 0.0960
#> 504     C_113_Y   1938 0.9690
#> 505     C_113_H     62 0.0310
#> 506     C_114_D   1457 0.7285
#> 507     C_114_N    542 0.2710
#> 508     C_114_V      1 0.0005
#> 509     C_116_S   1082 0.5410
#> 510     C_116_F    588 0.2940
#> 511     C_116_Y    280 0.1400
#> 512     C_116_L     49 0.0245
#> 513     C_116_H      1 0.0005
#> 514     C_127_N   1999 0.9995
#> 515     C_127_K      1 0.0005
#> 516     C_138_T   1767 0.8835
#> 517     C_138_K    233 0.1165
#> 518     C_143_T   1979 0.9895
#> 519     C_143_S     21 0.0105
#> 520     C_147_W   1460 0.7300
#> 521     C_147_L    540 0.2700
#> 522     C_152_E   1381 0.6905
#> 523     C_152_A    588 0.2940
#> 524     C_152_T     31 0.0155
#> 525     C_156_W    486 0.2430
#> 526     C_156_R    625 0.3125
#> 527     C_156_L    790 0.3950
#> 528     C_156_D     34 0.0170
#> 529     C_156_Q     65 0.0325
#> 530     C_163_T   1690 0.8450
#> 531     C_163_L    191 0.0955
#> 532     C_163_E    119 0.0595
#> 533     C_169_R   1999 0.9995
#> 534     C_169_H      1 0.0005
#> 535     C_170_R   1979 0.9895
#> 536     C_170_G     21 0.0105
#> 537     C_173_E   1809 0.9045
#> 538     C_173_K    191 0.0955
#> 539     C_175_G   1991 0.9955
#> 540     C_175_R      9 0.0045
#> 541     C_177_E   1702 0.8510
#> 542     C_177_K    298 0.1490
#> 543     C_184_*   1941 0.9705
#> 544     C_184_H     34 0.0170
#> 545     C_184_P     15 0.0075
#> 546     C_184_R     10 0.0050
#> 547     C_186_*   1941 0.9705
#> 548     C_186_K     58 0.0290
#> 549     C_186_R      1 0.0005
#> 550     C_194_*   1941 0.9705
#> 551     C_194_V     44 0.0220
#> 552     C_194_L     15 0.0075
#> 553     C_211_*   1941 0.9705
#> 554     C_211_A     57 0.0285
#> 555     C_211_T      2 0.0010
#> 556     C_219_*   1941 0.9705
#> 557     C_219_W     19 0.0095
#> 558     C_219_R     40 0.0200
#> 559     C_248_*   1941 0.9705
#> 560     C_248_M      4 0.0020
#> 561     C_248_V     55 0.0275
#> 562     C_253_*   1941 0.9705
#> 563     C_253_E     34 0.0170
#> 564     C_253_Q     25 0.0125
#> 565     C_261_*   1941 0.9705
#> 566     C_261_V     44 0.0220
#> 567     C_261_M     15 0.0075
#> 568     C_267_*   1941 0.9705
#> 569     C_267_P     34 0.0170
#> 570     C_267_Q     25 0.0125
#> 571     C_270_*   1941 0.9705
#> 572     C_270_L     49 0.0245
#> 573     C_270_C     10 0.0050
#> 574     C_273_*   1941 0.9705
#> 575     C_273_R     44 0.0220
#> 576     C_273_S     15 0.0075
#> 577     C_275_*   1944 0.9720
#> 578     C_275_E     40 0.0200
#> 579     C_275_K     14 0.0070
#> 580     C_275_G      2 0.0010
#> 581     C_284_*   1944 0.9720
#> 582     C_284_I     46 0.0230
#> 583     C_284_N     10 0.0050
#> 584     C_285_*   1944 0.9720
#> 585     C_285_V     31 0.0155
#> 586     C_285_M     15 0.0075
#> 587     C_285_L     10 0.0050
#> 588     C_289_*   1944 0.9720
#> 589     C_289_A     46 0.0230
#> 590     C_289_S     10 0.0050
#> 591     C_291_*   1944 0.9720
#> 592     C_291_L     46 0.0230
#> 593     C_291_P     10 0.0050
#> 594     C_295_*   1944 0.9720
#> 595     C_295_A     41 0.0205
#> 596     C_295_V     15 0.0075
#> 597     C_301_.   1979 0.9895
#> 598     C_301_A     10 0.0050
#> 599     C_301_*     11 0.0055
#> 600     C_302_.   1979 0.9895
#> 601     C_302_V     10 0.0050
#> 602     C_302_*     11 0.0055
#> 603     C_303_.   1979 0.9895
#> 604     C_303_L     10 0.0050
#> 605     C_303_*     11 0.0055
#> 606     C_304_.   1979 0.9895
#> 607     C_304_A     10 0.0050
#> 608     C_304_*     11 0.0055
#> 609     C_305_.   1979 0.9895
#> 610     C_305_V     10 0.0050
#> 611     C_305_*     11 0.0055
#> 612     C_306_.   1979 0.9895
#> 613     C_306_L     10 0.0050
#> 614     C_306_*     11 0.0055
#> 615     C_309_*   1944 0.9720
#> 616     C_309_V     55 0.0275
#> 617     C_309_M      1 0.0005
#> 618     C_310_*   1944 0.9720
#> 619     C_310_V     52 0.0260
#> 620     C_310_M      4 0.0020
#> 621     C_311_*   1944 0.9720
#> 622     C_311_A     41 0.0205
#> 623     C_311_T     15 0.0075
#> 624     C_312_*   1944 0.9720
#> 625     C_312_V     31 0.0155
#> 626     C_312_A     25 0.0125
#> 627     C_313_*   1944 0.9720
#> 628     C_313_V     41 0.0205
#> 629     C_313_M     15 0.0075
#> 630     C_314_*   1944 0.9720
#> 631     C_314_M     46 0.0230
#> 632     C_314_I     10 0.0050
#> 633     C_315_*   1944 0.9720
#> 634     C_315_C     46 0.0230
#> 635     C_315_H     10 0.0050
#> 636     C_327_*   1944 0.9720
#> 637     C_327_C     55 0.0275
#> 638     C_327_Y      1 0.0005
#> 639     C_330_*   1944 0.9720
#> 640     C_330_A     46 0.0230
#> 641     C_330_V     10 0.0050
#> 642     C_332_*   1945 0.9725
#> 643     C_332_S     41 0.0205
#> 644     C_332_C     14 0.0070
#> 645     C_345_*   1945 0.9725
#> 646     C_345_A     40 0.0200
#> 647     C_345_T     15 0.0075
#> 648   DPA1_11_A   1845 0.9225
#> 649   DPA1_11_M    155 0.0775
#> 650   DPA1_18_P   1999 0.9995
#> 651   DPA1_18_T      1 0.0005
#> 652   DPA1_28_E   1987 0.9935
#> 653   DPA1_28_D     13 0.0065
#> 654   DPA1_31_M   1545 0.7725
#> 655   DPA1_31_Q    455 0.2275
#> 656   DPA1_50_Q   1543 0.7715
#> 657   DPA1_50_R    457 0.2285
#> 658   DPA1_66_L   1992 0.9960
#> 659   DPA1_66_S      8 0.0040
#> 660   DPA1_72_T   1999 0.9995
#> 661   DPA1_72_I      1 0.0005
#> 662   DPA1_73_L   1999 0.9995
#> 663   DPA1_73_A      1 0.0005
#> 664   DPA1_83_T   1543 0.7715
#> 665   DPA1_83_A    457 0.2285
#> 666   DPA1_91_*   1957 0.9785
#> 667   DPA1_91_V     30 0.0150
#> 668   DPA1_91_M     13 0.0065
#> 669   DPA1_96_*   1957 0.9785
#> 670   DPA1_96_P     42 0.0210
#> 671   DPA1_96_A      1 0.0005
#> 672  DPA1_111_*   1957 0.9785
#> 673  DPA1_111_R     22 0.0110
#> 674  DPA1_111_K     21 0.0105
#> 675  DPA1_127_*   1957 0.9785
#> 676  DPA1_127_P     23 0.0115
#> 677  DPA1_127_L     20 0.0100
#> 678  DPA1_160_*   1957 0.9785
#> 679  DPA1_160_V     23 0.0115
#> 680  DPA1_160_F     20 0.0100
#> 681  DPA1_190_*   1957 0.9785
#> 682  DPA1_190_M      9 0.0045
#> 683  DPA1_190_T     33 0.0165
#> 684  DPA1_190_A      1 0.0005
#> 685  DPA1_228_*   1957 0.9785
#> 686  DPA1_228_P     23 0.0115
#> 687  DPA1_228_T     20 0.0100
#> 688  DPB1_-14_*   1750 0.8750
#> 689  DPB1_-14_T    249 0.1245
#> 690  DPB1_-14_M      1 0.0005
#> 691    DPB1_8_L   1431 0.7155
#> 692    DPB1_8_V    569 0.2845
#> 693    DPB1_9_F   1431 0.7155
#> 694    DPB1_9_Y    427 0.2135
#> 695    DPB1_9_H    142 0.0710
#> 696   DPB1_11_G   1526 0.7630
#> 697   DPB1_11_L    474 0.2370
#> 698   DPB1_17_A   1999 0.9995
#> 699   DPB1_17_P      1 0.0005
#> 700   DPB1_33_E   1947 0.9735
#> 701   DPB1_33_Q     53 0.0265
#> 702   DPB1_35_F   1659 0.8295
#> 703   DPB1_35_L    152 0.0760
#> 704   DPB1_35_Y    189 0.0945
#> 705   DPB1_36_V   1099 0.5495
#> 706   DPB1_36_A    901 0.4505
#> 707   DPB1_55_D    925 0.4625
#> 708   DPB1_55_E    160 0.0800
#> 709   DPB1_55_A    915 0.4575
#> 710   DPB1_56_E    925 0.4625
#> 711   DPB1_56_A   1075 0.5375
#> 712   DPB1_57_E   1652 0.8260
#> 713   DPB1_57_D    348 0.1740
#> 714   DPB1_65_I   1671 0.8355
#> 715   DPB1_65_L    328 0.1640
#> 716   DPB1_65_F      1 0.0005
#> 717   DPB1_69_E    541 0.2705
#> 718   DPB1_69_K   1406 0.7030
#> 719   DPB1_69_R     53 0.0265
#> 720   DPB1_76_M   1563 0.7815
#> 721   DPB1_76_V    376 0.1880
#> 722   DPB1_76_I     61 0.0305
#> 723   DPB1_84_G   1298 0.6490
#> 724   DPB1_84_D    687 0.3435
#> 725   DPB1_84_V     15 0.0075
#> 726   DPB1_85_G   1313 0.6565
#> 727   DPB1_85_E    687 0.3435
#> 728   DPB1_86_P   1313 0.6565
#> 729   DPB1_86_A    687 0.3435
#> 730   DPB1_87_M   1313 0.6565
#> 731   DPB1_87_V    687 0.3435
#> 732   DPB1_96_*   1710 0.8550
#> 733   DPB1_96_K    156 0.0780
#> 734   DPB1_96_R    134 0.0670
#> 735  DPB1_170_*   1710 0.8550
#> 736  DPB1_170_I    156 0.0780
#> 737  DPB1_170_T    134 0.0670
#> 738  DPB1_178_*   1710 0.8550
#> 739  DPB1_178_L    277 0.1385
#> 740  DPB1_178_M     13 0.0065
#> 741  DPB1_194_*   1714 0.8570
#> 742  DPB1_194_R    285 0.1425
#> 743  DPB1_194_Q      1 0.0005
#> 744  DPB1_205_*   1714 0.8570
#> 745  DPB1_205_V    213 0.1065
#> 746  DPB1_205_M     73 0.0365
#> 747  DPB1_215_*   1714 0.8570
#> 748  DPB1_215_I    285 0.1425
#> 749  DPB1_215_T      1 0.0005
#> 750  DQA1_-16_M    497 0.2485
#> 751  DQA1_-16_*    664 0.3320
#> 752  DQA1_-16_L    839 0.4195
#> 753  DQA1_-13_T    312 0.1560
#> 754  DQA1_-13_*    664 0.3320
#> 755  DQA1_-13_A   1024 0.5120
#> 756   DQA1_-7_V   1259 0.6295
#> 757   DQA1_-7_*    664 0.3320
#> 758   DQA1_-7_M     77 0.0385
#> 759   DQA1_-6_M   1305 0.6525
#> 760   DQA1_-6_*    664 0.3320
#> 761   DQA1_-6_T     31 0.0155
#> 762    DQA1_1_E   1552 0.7760
#> 763    DQA1_1_*    442 0.2210
#> 764    DQA1_1_K      6 0.0030
#> 765    DQA1_2_D   1481 0.7405
#> 766    DQA1_2_*    442 0.2210
#> 767    DQA1_2_G     77 0.0385
#> 768   DQA1_11_Y   1198 0.5990
#> 769   DQA1_11_C    780 0.3900
#> 770   DQA1_11_*     22 0.0110
#> 771   DQA1_18_S   1220 0.6100
#> 772   DQA1_18_F    780 0.3900
#> 773   DQA1_25_Y   1570 0.7850
#> 774   DQA1_25_F    430 0.2150
#> 775   DQA1_26_T   1674 0.8370
#> 776   DQA1_26_S    326 0.1630
#> 777   DQA1_34_Q   1149 0.5745
#> 778   DQA1_34_E    851 0.4255
#> 779   DQA1_40_G    638 0.3190
#> 780   DQA1_40_E   1362 0.6810
#> 781   DQA1_41_R   1848 0.9240
#> 782   DQA1_41_K    152 0.0760
#> 783   DQA1_44_T   1999 0.9995
#> 784   DQA1_44_A      1 0.0005
#> 785   DQA1_45_V   1220 0.6100
#> 786   DQA1_45_A    780 0.3900
#> 787   DQA1_47_C    638 0.3190
#> 788   DQA1_47_K    256 0.1280
#> 789   DQA1_47_R    780 0.3900
#> 790   DQA1_47_Q    326 0.1630
#> 791   DQA1_48_L   1220 0.6100
#> 792   DQA1_48_W    780 0.3900
#> 793   DQA1_50_V    638 0.3190
#> 794   DQA1_50_L    582 0.2910
#> 795   DQA1_50_E    780 0.3900
#> 796   DQA1_51_L    638 0.3190
#> 797   DQA1_51_F   1362 0.6810
#> 798   DQA1_52_R    964 0.4820
#> 799   DQA1_52_H    256 0.1280
#> 800   DQA1_52_S    780 0.3900
#> 801   DQA1_53_Q    638 0.3190
#> 802   DQA1_53_R    582 0.2910
#> 803   DQA1_53_K    780 0.3900
#> 804   DQA1_54_F   1744 0.8720
#> 805   DQA1_54_L    256 0.1280
#> 806   DQA1_55_R   1220 0.6100
#> 807   DQA1_55_G    780 0.3900
#> 808   DQA1_56_.    894 0.4470
#> 809   DQA1_56_G    780 0.3900
#> 810   DQA1_56_R    326 0.1630
#> 811   DQA1_61_F   1220 0.6100
#> 812   DQA1_61_G    780 0.3900
#> 813   DQA1_64_T   1220 0.6100
#> 814   DQA1_64_R    780 0.3900
#> 815   DQA1_66_I   1220 0.6100
#> 816   DQA1_66_M    780 0.3900
#> 817   DQA1_69_L   1138 0.5690
#> 818   DQA1_69_A    780 0.3900
#> 819   DQA1_69_T     82 0.0410
#> 820   DQA1_75_S    556 0.2780
#> 821   DQA1_75_I   1444 0.7220
#> 822   DQA1_76_L    894 0.4470
#> 823   DQA1_76_M    780 0.3900
#> 824   DQA1_76_V    326 0.1630
#> 825   DQA1_80_S   1220 0.6100
#> 826   DQA1_80_Y    780 0.3900
#> 827   DQA1_84_A   1998 0.9990
#> 828   DQA1_84_T      2 0.0010
#> 829  DQA1_102_L   1498 0.7490
#> 830  DQA1_102_*    501 0.2505
#> 831  DQA1_102_V      1 0.0005
#> 832  DQA1_107_I    334 0.1670
#> 833  DQA1_107_*    501 0.2505
#> 834  DQA1_107_T   1165 0.5825
#> 835  DQA1_129_H    872 0.4360
#> 836  DQA1_129_*    501 0.2505
#> 837  DQA1_129_Q    627 0.3135
#> 838  DQA1_130_S   1347 0.6735
#> 839  DQA1_130_*    501 0.2505
#> 840  DQA1_130_A    152 0.0760
#> 841  DQA1_138_T   1497 0.7485
#> 842  DQA1_138_*    501 0.2505
#> 843  DQA1_138_I      2 0.0010
#> 844  DQA1_139_S   1498 0.7490
#> 845  DQA1_139_*    501 0.2505
#> 846  DQA1_139_I      1 0.0005
#> 847  DQA1_156_L    334 0.1670
#> 848  DQA1_156_*    501 0.2505
#> 849  DQA1_156_F   1165 0.5825
#> 850  DQA1_160_A   1314 0.6570
#> 851  DQA1_160_*    501 0.2505
#> 852  DQA1_160_D    163 0.0815
#> 853  DQA1_160_S     22 0.0110
#> 854  DQA1_161_E    334 0.1670
#> 855  DQA1_161_*    501 0.2505
#> 856  DQA1_161_D   1165 0.5825
#> 857  DQA1_163_S    334 0.1670
#> 858  DQA1_163_*    501 0.2505
#> 859  DQA1_163_I   1165 0.5825
#> 860  DQA1_175_K    334 0.1670
#> 861  DQA1_175_*    501 0.2505
#> 862  DQA1_175_Q    779 0.3895
#> 863  DQA1_175_E    386 0.1930
#> 864  DQA1_187_A   1173 0.5865
#> 865  DQA1_187_*    664 0.3320
#> 866  DQA1_187_T    163 0.0815
#> 867  DQA1_199_A   1284 0.6420
#> 868  DQA1_199_*    664 0.3320
#> 869  DQA1_199_T     52 0.0260
#> 870  DQA1_207_V    978 0.4890
#> 871  DQA1_207_*    664 0.3320
#> 872  DQA1_207_M    358 0.1790
#> 873  DQA1_208_G   1335 0.6675
#> 874  DQA1_208_*    664 0.3320
#> 875  DQA1_208_C      1 0.0005
#> 876  DQA1_215_F   1173 0.5865
#> 877  DQA1_215_*    664 0.3320
#> 878  DQA1_215_L    163 0.0815
#> 879  DQA1_218_R    557 0.2785
#> 880  DQA1_218_*    664 0.3320
#> 881  DQA1_218_Q    779 0.3895
#> 882  DQB1_-27_*   1988 0.9940
#> 883  DQB1_-27_S      4 0.0020
#> 884  DQB1_-27_A      8 0.0040
#> 885  DQB1_-21_*   1988 0.9940
#> 886  DQB1_-21_D      5 0.0025
#> 887  DQB1_-21_G      7 0.0035
#> 888   DQB1_-9_*   1988 0.9940
#> 889   DQB1_-9_I      4 0.0020
#> 890   DQB1_-9_M      8 0.0040
#> 891   DQB1_-6_*   1988 0.9940
#> 892   DQB1_-6_S      5 0.0025
#> 893   DQB1_-6_T      7 0.0035
#> 894   DQB1_-5_*   1988 0.9940
#> 895   DQB1_-5_S      4 0.0020
#> 896   DQB1_-5_P      7 0.0035
#> 897   DQB1_-5_L      1 0.0005
#> 898   DQB1_-4_*   1988 0.9940
#> 899   DQB1_-4_L      5 0.0025
#> 900   DQB1_-4_V      7 0.0035
#> 901    DQB1_9_F    302 0.1510
#> 902    DQB1_9_Y   1601 0.8005
#> 903    DQB1_9_*     97 0.0485
#> 904   DQB1_13_G   1460 0.7300
#> 905   DQB1_13_A    492 0.2460
#> 906   DQB1_13_*     48 0.0240
#> 907   DQB1_14_M   1673 0.8365
#> 908   DQB1_14_L    279 0.1395
#> 909   DQB1_14_*     48 0.0240
#> 910   DQB1_23_R   1974 0.9870
#> 911   DQB1_23_L     26 0.0130
#> 912   DQB1_26_L   1091 0.5455
#> 913   DQB1_26_Y    492 0.2460
#> 914   DQB1_26_G    417 0.2085
#> 915   DQB1_28_T   1580 0.7900
#> 916   DQB1_28_S    420 0.2100
#> 917   DQB1_30_Y   1073 0.5365
#> 918   DQB1_30_S    420 0.2100
#> 919   DQB1_30_H    507 0.2535
#> 920   DQB1_37_Y   1531 0.7655
#> 921   DQB1_37_I    420 0.2100
#> 922   DQB1_37_D     49 0.0245
#> 923   DQB1_38_A   1204 0.6020
#> 924   DQB1_38_V    796 0.3980
#> 925   DQB1_45_G   1557 0.7785
#> 926   DQB1_45_E    443 0.2215
#> 927   DQB1_46_V   1580 0.7900
#> 928   DQB1_46_E    420 0.2100
#> 929   DQB1_47_Y   1580 0.7900
#> 930   DQB1_47_F    420 0.2100
#> 931   DQB1_52_P   1580 0.7900
#> 932   DQB1_52_L    420 0.2100
#> 933   DQB1_53_Q    793 0.3965
#> 934   DQB1_53_L   1207 0.6035
#> 935   DQB1_55_R    880 0.4400
#> 936   DQB1_55_L    420 0.2100
#> 937   DQB1_55_P    700 0.3500
#> 938   DQB1_56_P   1913 0.9565
#> 939   DQB1_56_L     87 0.0435
#> 940   DQB1_57_D   1032 0.5160
#> 941   DQB1_57_A    601 0.3005
#> 942   DQB1_57_V    295 0.1475
#> 943   DQB1_57_S     72 0.0360
#> 944   DQB1_66_E   1440 0.7200
#> 945   DQB1_66_D    560 0.2800
#> 946   DQB1_67_V   1440 0.7200
#> 947   DQB1_67_I    560 0.2800
#> 948   DQB1_70_G    603 0.3015
#> 949   DQB1_70_R   1258 0.6290
#> 950   DQB1_70_E     91 0.0455
#> 951   DQB1_70_*     48 0.0240
#> 952   DQB1_71_T   1166 0.5830
#> 953   DQB1_71_K    420 0.2100
#> 954   DQB1_71_D     91 0.0455
#> 955   DQB1_71_A    275 0.1375
#> 956   DQB1_71_*     48 0.0240
#> 957   DQB1_74_E   1166 0.5830
#> 958   DQB1_74_A    420 0.2100
#> 959   DQB1_74_S    366 0.1830
#> 960   DQB1_74_*     48 0.0240
#> 961   DQB1_75_L   1166 0.5830
#> 962   DQB1_75_V    786 0.3930
#> 963   DQB1_75_*     48 0.0240
#> 964   DQB1_77_T   1253 0.6265
#> 965   DQB1_77_R    699 0.3495
#> 966   DQB1_77_*     48 0.0240
#> 967   DQB1_84_E    696 0.3480
#> 968   DQB1_84_Q   1207 0.6035
#> 969   DQB1_84_*     97 0.0485
#> 970   DQB1_85_V    696 0.3480
#> 971   DQB1_85_L   1207 0.6035
#> 972   DQB1_85_*     97 0.0485
#> 973   DQB1_86_A    607 0.3035
#> 974   DQB1_86_E   1207 0.6035
#> 975   DQB1_86_*     97 0.0485
#> 976   DQB1_86_G     89 0.0445
#> 977   DQB1_87_F    328 0.1640
#> 978   DQB1_87_L   1207 0.6035
#> 979   DQB1_87_*     97 0.0485
#> 980   DQB1_87_Y    368 0.1840
#> 981   DQB1_89_G    696 0.3480
#> 982   DQB1_89_T   1207 0.6035
#> 983   DQB1_89_*     97 0.0485
#> 984   DQB1_90_I    696 0.3480
#> 985   DQB1_90_T   1207 0.6035
#> 986   DQB1_90_*     97 0.0485
#> 987  DQB1_116_*   1749 0.8745
#> 988  DQB1_116_V    246 0.1230
#> 989  DQB1_116_I      5 0.0025
#> 990  DQB1_125_*   1749 0.8745
#> 991  DQB1_125_A    245 0.1225
#> 992  DQB1_125_S      5 0.0025
#> 993  DQB1_125_G      1 0.0005
#> 994  DQB1_126_*   1749 0.8745
#> 995  DQB1_126_Q    247 0.1235
#> 996  DQB1_126_H      4 0.0020
#> 997  DQB1_130_*   1749 0.8745
#> 998  DQB1_130_R    218 0.1090
#> 999  DQB1_130_Q     33 0.0165
#> 1000 DQB1_135_*   1749 0.8745
#> 1001 DQB1_135_G    202 0.1010
#> 1002 DQB1_135_D     49 0.0245
#> 1003 DQB1_140_*   1749 0.8745
#> 1004 DQB1_140_A    208 0.1040
#> 1005 DQB1_140_T     43 0.0215
#> 1006 DQB1_142_*   1749 0.8745
#> 1007 DQB1_142_V    250 0.1250
#> 1008 DQB1_142_I      1 0.0005
#> 1009 DQB1_167_*   1749 0.8745
#> 1010 DQB1_167_R    218 0.1090
#> 1011 DQB1_167_H     33 0.0165
#> 1012 DQB1_177_*   1749 0.8745
#> 1013 DQB1_177_H    250 0.1250
#> 1014 DQB1_177_Q      1 0.0005
#> 1015 DQB1_182_*   1749 0.8745
#> 1016 DQB1_182_S    208 0.1040
#> 1017 DQB1_182_N     43 0.0215
#> 1018 DQB1_185_*   1749 0.8745
#> 1019 DQB1_185_T    230 0.1150
#> 1020 DQB1_185_I     21 0.0105
#> 1021 DQB1_203_*   1988 0.9940
#> 1022 DQB1_203_V      5 0.0025
#> 1023 DQB1_203_I      7 0.0035
#> 1024 DQB1_220_*   1988 0.9940
#> 1025 DQB1_220_R      5 0.0025
#> 1026 DQB1_220_H      7 0.0035
#> 1027 DQB1_221_*   1988 0.9940
#> 1028 DQB1_221_Q      5 0.0025
#> 1029 DQB1_221_H      7 0.0035
#> 1030 DQB1_224_*   1988 0.9940
#> 1031 DQB1_224_R      4 0.0020
#> 1032 DQB1_224_Q      8 0.0040
#> 1033  DRA_217_L    711 0.3555
#> 1034  DRA_217_V   1289 0.6445
#> 1035 DRB1_-16_*   1986 0.9930
#> 1036 DRB1_-16_V     12 0.0060
#> 1037 DRB1_-16_I      2 0.0010
#> 1038   DRB1_9_W    751 0.3755
#> 1039   DRB1_9_E   1208 0.6040
#> 1040   DRB1_9_K     32 0.0160
#> 1041   DRB1_9_*      9 0.0045
#> 1042  DRB1_10_Q   1072 0.5360
#> 1043  DRB1_10_Y    897 0.4485
#> 1044  DRB1_10_E     22 0.0110
#> 1045  DRB1_10_*      9 0.0045
#> 1046  DRB1_11_P    325 0.1625
#> 1047  DRB1_11_G    249 0.1245
#> 1048  DRB1_11_S    897 0.4485
#> 1049  DRB1_11_D     32 0.0160
#> 1050  DRB1_11_V    311 0.1555
#> 1051  DRB1_11_L    177 0.0885
#> 1052  DRB1_11_*      9 0.0045
#> 1053  DRB1_12_K   1094 0.5470
#> 1054  DRB1_12_T    897 0.4485
#> 1055  DRB1_12_*      9 0.0045
#> 1056  DRB1_13_R    325 0.1625
#> 1057  DRB1_13_Y    249 0.1245
#> 1058  DRB1_13_S    760 0.3800
#> 1059  DRB1_13_F    231 0.1155
#> 1060  DRB1_13_H    289 0.1445
#> 1061  DRB1_13_G    137 0.0685
#> 1062  DRB1_13_*      9 0.0045
#> 1063  DRB1_14_E   1742 0.8710
#> 1064  DRB1_14_K    249 0.1245
#> 1065  DRB1_14_*      9 0.0045
#> 1066  DRB1_16_H   1852 0.9260
#> 1067  DRB1_16_Y    137 0.0685
#> 1068  DRB1_16_Q      2 0.0010
#> 1069  DRB1_16_*      9 0.0045
#> 1070  DRB1_25_R   1751 0.8755
#> 1071  DRB1_25_Q    249 0.1245
#> 1072  DRB1_26_F   1495 0.7475
#> 1073  DRB1_26_Y    248 0.1240
#> 1074  DRB1_26_L    257 0.1285
#> 1075  DRB1_28_D   1446 0.7230
#> 1076  DRB1_28_E    522 0.2610
#> 1077  DRB1_28_H     32 0.0160
#> 1078  DRB1_30_Y   1458 0.7290
#> 1079  DRB1_30_L    249 0.1245
#> 1080  DRB1_30_G     32 0.0160
#> 1081  DRB1_30_R     22 0.0110
#> 1082  DRB1_30_C    177 0.0885
#> 1083  DRB1_30_H     62 0.0310
#> 1084  DRB1_31_F   1769 0.8845
#> 1085  DRB1_31_I    209 0.1045
#> 1086  DRB1_31_V     22 0.0110
#> 1087  DRB1_32_Y   1436 0.7180
#> 1088  DRB1_32_H    564 0.2820
#> 1089  DRB1_33_N   1711 0.8555
#> 1090  DRB1_33_H    289 0.1445
#> 1091  DRB1_37_S    510 0.2550
#> 1092  DRB1_37_F    302 0.1510
#> 1093  DRB1_37_N    463 0.2315
#> 1094  DRB1_37_Y    667 0.3335
#> 1095  DRB1_37_L     58 0.0290
#> 1096  DRB1_38_V   1920 0.9600
#> 1097  DRB1_38_A     22 0.0110
#> 1098  DRB1_38_L     58 0.0290
#> 1099  DRB1_40_F   1978 0.9890
#> 1100  DRB1_40_Y     22 0.0110
#> 1101  DRB1_47_F    993 0.4965
#> 1102  DRB1_47_Y   1007 0.5035
#> 1103  DRB1_57_D   1469 0.7345
#> 1104  DRB1_57_V    339 0.1695
#> 1105  DRB1_57_A     52 0.0260
#> 1106  DRB1_57_S    140 0.0700
#> 1107  DRB1_58_A   1749 0.8745
#> 1108  DRB1_58_E    251 0.1255
#> 1109  DRB1_60_Y   1610 0.8050
#> 1110  DRB1_60_S    339 0.1695
#> 1111  DRB1_60_H     51 0.0255
#> 1112  DRB1_67_I    839 0.4195
#> 1113  DRB1_67_L    749 0.3745
#> 1114  DRB1_67_F    412 0.2060
#> 1115  DRB1_70_Q    935 0.4675
#> 1116  DRB1_70_D    957 0.4785
#> 1117  DRB1_70_R    108 0.0540
#> 1118  DRB1_71_A    269 0.1345
#> 1119  DRB1_71_R   1137 0.5685
#> 1120  DRB1_71_E    245 0.1225
#> 1121  DRB1_71_K    349 0.1745
#> 1122  DRB1_73_A   1531 0.7655
#> 1123  DRB1_73_G    469 0.2345
#> 1124  DRB1_74_A   1299 0.6495
#> 1125  DRB1_74_Q    249 0.1245
#> 1126  DRB1_74_E    147 0.0735
#> 1127  DRB1_74_R    220 0.1100
#> 1128  DRB1_74_L     85 0.0425
#> 1129  DRB1_77_T   1780 0.8900
#> 1130  DRB1_77_N    220 0.1100
#> 1131  DRB1_78_Y   1718 0.8590
#> 1132  DRB1_78_V    282 0.1410
#> 1133  DRB1_82_N   1999 0.9995
#> 1134  DRB1_82_S      1 0.0005
#> 1135  DRB1_85_V   1904 0.9520
#> 1136  DRB1_85_A     96 0.0480
#> 1137  DRB1_86_V    958 0.4790
#> 1138  DRB1_86_G   1042 0.5210
#> 1139  DRB1_96_*   1943 0.9715
#> 1140  DRB1_96_H     56 0.0280
#> 1141  DRB1_96_Y      1 0.0005
#> 1142  DRB1_98_*   1943 0.9715
#> 1143  DRB1_98_K     56 0.0280
#> 1144  DRB1_98_E      1 0.0005
#> 1145 DRB1_104_*   1943 0.9715
#> 1146 DRB1_104_S     56 0.0280
#> 1147 DRB1_104_A      1 0.0005
#> 1148 DRB1_120_*   1943 0.9715
#> 1149 DRB1_120_S     56 0.0280
#> 1150 DRB1_120_N      1 0.0005
#> 1151 DRB1_149_*   1943 0.9715
#> 1152 DRB1_149_H     56 0.0280
#> 1153 DRB1_149_Q      1 0.0005
#> 1154 DRB1_180_*   1943 0.9715
#> 1155 DRB1_180_V     56 0.0280
#> 1156 DRB1_180_L      1 0.0005
#> 1157 DRB1_189_*   1985 0.9925
#> 1158 DRB1_189_R     11 0.0055
#> 1159 DRB1_189_S      4 0.0020
#> 1160 DRB1_207_*   1985 0.9925
#> 1161 DRB1_207_V     14 0.0070
#> 1162 DRB1_207_M      1 0.0005
#> 1163 DRB1_233_*   1986 0.9930
#> 1164 DRB1_233_R     10 0.0050
#> 1165 DRB1_233_T      4 0.0020