Find variable positions in sequence alignment

getVariableAAPos finds variable amino acid positions in protein sequence alignment.

getVariableAAPos(alignment, varchar = "[A-Z]")

Arguments

alignment

Matrix containing amino acid level alignment, as returned by readHlaAlignments,

varchar

Regex matching characters that should be considered when looking for variable amino acid positions. See details for further explanations.

Value

Integer vector specifying which alignment columns are variable.

Details

The variable amino acid positions in the alignment are those at which different amino acids can be found. As the alignments can also contain indels and unknown characters, the user choice might be to consider those positions as variable or not. This can be achieved by passing appropriate regular expression in varchar. Eg. when varchar = "[A-Z]" occurence of deletion/insertion (".") will not be treated as variability. In order to detect this kind of variability varchar = "[A-Z\\.]" should be used.

Examples

alignment <- readHlaAlignments(gene = "TAP1")
getVariableAAPos(alignment)
#>  80 131 201 202 203 204 205 206 207 208 209 210 211 212 213 214 215 216 217 218 
#>  80 131 201 202 203 204 205 206 207 208 209 210 211 212 213 214 215 216 217 218 
#> 219 220 221 222 223 224 225 226 227 228 333 370 458 518 637 648 
#> 219 220 221 222 223 224 225 226 227 228 333 370 458 518 637 648