Convert counts table to variables

countsToVariables converts counts table to additional variables.

countsToVariables(counts, dictionary, na.value = NA, nacols.rm = TRUE)

Arguments

counts
Data frame with counts, such as returned by hlaCallsToCounts function. First column should contain samples IDs, following columns should contain counts (natural numbers including zero).

dictionary
Path to file containing variables dictionary or data frame. See details for further explanations.

na.value
Vector of length one speciyfing value for variables with no matching entry in dictionary. Default is to use 0.

nacols.rm
Logical indicating if result columns that contain only NA should be removed.

Value

Data frame with variable number of columns. First column named "ID" corresponds to "ID" column in counts, further columns hold indicators for converted variables. 1 and 0 code presence and absence of a variable respectively.

Details

dictionary file should be a tsv format with header and two columns. First column should be named "Name" and hold variable name, second should be named "Expression" and hold expression used to identify variable (eg. "KIR2DL3 & ! KIR2DL2" will match all samples with KIR2DL3 and without KIR2DL2). Optionally a data frame formatted in the same manner can be passed instead.

Dictionaries shipped with the package:

kir_haplotypes

KIR genes to KIR haplotypes dictionary.

Examples

countsToVariables(MiDAS_tut_KIR, "kir_haplotypes")

#>        ID cenAA cenBB cenAB telAA telBB telAB
#> 1    C001     1     0     0     1     0     0
#> 2    C002     1     0     0     1     0     0
#> 3    C003     1     0     0     0     1     0
#> 4    C004     1     0     0     0     1     0
#> 5    C005     1     0     0     0     0     1
#> 6    C006     1     0     0     1     0     0
#> 7    C007     1     0     0     1     0     0
#> 8    C008     1     0     0     1     0     0
#> 9    C009     0     1     0     0     1     0
#> 10   C010     0     0     1     0     1     0
#> 11   C011    NA    NA    NA    NA    NA    NA
#> 12   C012     0     0     1     0     1     0
#> 13   C013     0     0     1     0     1     0
#> 14   C014     0     0     1     0     1     0
#> 15   C015    NA    NA    NA    NA    NA    NA
#> 16   C016     1     0     0     0     1     0
#> 17   C017     0     0     1     0     1     0
#> 18   C018     1     0     0     0     1     0
#> 19   C019     1     0     0     0     1     0
#> 20   C020     0     0     1     0     1     0
#> 21   C021     0     0     1     0     1     0
#> 22   C022     0     0     1     0     1     0
#> 23   C023     0     0     1     0     1     0
#> 24   C024     0     0     1     0     1     0
#> 25   C025     0     0     1     0     1     0
#> 26   C026    NA    NA    NA    NA    NA    NA
#> 27   C027     0     1     0     0     1     0
#> 28   C028     0     1     0     0     1     0
#> 29   C029     1     0     0     0     1     0
#> 30   C030     0     0     1     0     1     0
#> 31   C031     0     0     1     0     1     0
#> 32   C032     0     0     1     1     0     0
#> 33   C033     1     0     0     0     0     1
#> 34   C034     1     0     0     0     0     1
#> 35   C035     0     0     1     1     0     0
#> 36   C036     1     0     0     1     0     0
#> 37   C037     0     0     1     0     0     1
#> 38   C038     0     0     1     1     0     0
#> 39   C039     0     0     1     0     0     1
#> 40   C040     0     0     1     1     0     0
#> 41   C041     1     0     0     1     0     0
#> 42   C042     0     0     1     1     0     0
#> 43   C043     1     0     0     1     0     0
#> 44   C044     1     0     0     1     0     0
#> 45   C045     1     0     0     1     0     0
#> 46   C046     1     0     0     0     0     1
#> 47   C047     0     0     1     0     1     0
#> 48   C048     0     1     0     0     1     0
#> 49   C049     0     0     1     0     1     0
#> 50   C050     0     0     1     0     1     0
#> 51   C051     0     0     1     0     1     0
#> 52   C052     0     0     1     0     1     0
#> 53   C053     1     0     0     0     1     0
#> 54   C054     0     0     1     0     1     0
#> 55   C055     1     0     0     0     1     0
#> 56   C056     0     0     1     0     1     0
#> 57   C057     0     0     1     0     1     0
#> 58   C058     0     0     1     0     1     0
#> 59   C059     0     0     1     0     1     0
#> 60   C060     0     0     1     0     1     0
#> 61   C061     0     0     1     0     1     0
#> 62   C062     0     1     0     0     1     0
#> 63   C063     0     1     0     0     1     0
#> 64   C064     1     0     0     0     1     0
#> 65   C065     0     0     1     0     1     0
#> 66   C066     0     0     1     0     1     0
#> 67   C067     1     0     0     1     0     0
#> 68   C068     0     0     1     1     0     0
#> 69   C069     1     0     0     1     0     0
#> 70   C070     0     0     1     0     0     1
#> 71   C071     0     0     1     0     0     1
#> 72   C072     0     1     0     0     0     1
#> 73   C073     1     0     0     1     0     0
#> 74   C074     1     0     0     1     0     0
#> 75   C075     1     0     0     1     0     0
#> 76   C076     0     0     1     0     0     1
#> 77   C077     0     0     1     1     0     0
#> 78   C078    NA    NA    NA    NA    NA    NA
#> 79   C079     1     0     0     1     0     0
#> 80   C080     1     0     0     0     0     1
#> 81   C081     1     0     0     0     0     1
#> 82   C082     1     0     0     1     0     0
#> 83   C083     1     0     0     1     0     0
#> 84   C084     1     0     0     1     0     0
#> 85   C085     1     0     0     0     0     1
#> 86   C086    NA    NA    NA    NA    NA    NA
#> 87   C087     1     0     0     1     0     0
#> 88   C088     0     0     1     0     0     1
#> 89   C089     0     1     0     0     0     1
#> 90   C090     0     0     1     0     0     1
#> 91   C091     1     0     0     1     0     0
#> 92   C092     0     0     1     0     0     1
#> 93   C093     1     0     0     1     0     0
#> 94   C094     0     0     1     0     0     1
#> 95   C095     1     0     0     1     0     0
#> 96   C096     1     0     0     1     0     0
#> 97   C097     0     0     1     0     0     1
#> 98   C098     1     0     0     1     0     0
#> 99   C099     1     0     0     1     0     0
#> 100  C100    NA    NA    NA    NA    NA    NA
#> 101  C101     0     1     0     0     0     1
#> 102  C102     0     0     1     1     0     0
#> 103  C103     1     0     0     1     0     0
#> 104  C104     1     0     0     1     0     0
#> 105  C105     1     0     0     1     0     0
#> 106  C106     0     0     1     1     0     0
#> 107  C107     0     1     0     0     0     1
#> 108  C108     1     0     0     1     0     0
#> 109  C109     0     0     1     0     0     1
#> 110  C110     1     0     0     1     0     0
#> 111  C111     0     0     1     1     0     0
#> 112  C112     0     0     1     0     0     1
#> 113  C113     1     0     0     1     0     0
#> 114  C114     0     0     1     0     0     1
#> 115  C115     1     0     0     0     0     1
#> 116  C116     1     0     0     1     0     0
#> 117  C117     1     0     0     1     0     0
#> 118  C118     0     0     1     0     1     0
#> 119  C119     0     1     0     0     0     1
#> 120  C120     0     0     1     0     1     0
#> 121  C121     0     0     1     1     0     0
#> 122  C122     1     0     0     1     0     0
#> 123  C123     1     0     0     1     0     0
#> 124  C124     1     0     0     1     0     0
#> 125  C125     1     0     0     1     0     0
#> 126  C126     1     0     0     0     0     1
#> 127  C127     0     1     0     1     0     0
#> 128  C128     1     0     0     1     0     0
#> 129  C129     1     0     0     1     0     0
#> 130  C130     0     0     1     1     0     0
#> 131  C131     1     0     0     1     0     0
#> 132  C132     0     0     1     1     0     0
#> 133  C133     0     0     1     1     0     0
#> 134  C134     0     1     0     0     1     0
#> 135  C135    NA    NA    NA    NA    NA    NA
#> 136  C136     1     0     0     1     0     0
#> 137  C137     1     0     0     0     0     1
#> 138  C138     1     0     0     0     0     1
#> 139  C139     0     1     0     1     0     0
#> 140  C140     0     0     1     1     0     0
#> 141  C141     1     0     0     0     0     1
#> 142  C142     1     0     0     1     0     0
#> 143  C143     1     0     0     1     0     0
#> 144  C144     0     0     1     1     0     0
#> 145  C145     0     1     0     0     0     1
#> 146  C146    NA    NA    NA    NA    NA    NA
#> 147  C147     1     0     0     1     0     0
#> 148  C148     0     0     1     1     0     0
#> 149  C149     1     0     0     0     0     1
#> 150  C150     0     0     1     0     0     1
#> 151  C151     0     1     0     0     0     1
#> 152  C152     1     0     0     1     0     0
#> 153  C153     1     0     0     1     0     0
#> 154  C154     0     0     1     1     0     0
#> 155  C155     1     0     0     1     0     0
#> 156  C156     1     0     0     1     0     0
#> 157  C157     1     0     0     1     0     0
#> 158  C158     0     0     1     1     0     0
#> 159  C159     1     0     0     1     0     0
#> 160  C160     1     0     0     0     1     0
#> 161  C161     0     0     1     1     0     0
#> 162  C162     1     0     0     0     0     1
#> 163  C163     0     0     1     1     0     0
#> 164  C164     1     0     0     0     0     1
#> 165  C165     1     0     0     1     0     0
#> 166  C166     0     0     1     0     0     1
#> 167  C167     0     0     1     1     0     0
#> 168  C168     1     0     0     0     1     0
#> 169  C169     1     0     0     1     0     0
#> 170  C170     0     0     1     1     0     0
#> 171  C171     1     0     0     0     0     1
#> 172  C172     0     1     0     1     0     0
#> 173  C173     0     0     1     1     0     0
#> 174  C174     1     0     0     0     0     1
#> 175  C175     0     0     1     1     0     0
#> 176  C176     1     0     0     1     0     0
#> 177  C177     1     0     0     1     0     0
#> 178  C178     1     0     0     0     0     1
#> 179  C179     0     0     1     1     0     0
#> 180  C180     0     0     1     0     0     1
#> 181  C181     1     0     0     1     0     0
#> 182  C182     1     0     0     0     0     1
#> 183  C183    NA    NA    NA    NA    NA    NA
#> 184  C184     0     0     1     1     0     0
#> 185  C185     0     0     1     1     0     0
#> 186  C186     1     0     0     1     0     0
#> 187  C187     1     0     0     0     0     1
#> 188  C188     1     0     0     1     0     0
#> 189  C189     0     0     1     1     0     0
#> 190  C190     0     0     1     0     0     1
#> 191  C191     0     0     1     1     0     0
#> 192  C192     0     0     1     0     0     1
#> 193  C193     1     0     0     1     0     0
#> 194  C194    NA    NA    NA    NA    NA    NA
#> 195  C195     1     0     0     1     0     0
#> 196  C196    NA    NA    NA    NA    NA    NA
#> 197  C197     1     0     0     0     0     1
#> 198  C198     0     0     1     0     1     0
#> 199  C199     0     0     1     0     0     1
#> 200  C200     0     0     1     0     0     1
#> 201  C201     0     0     1     0     0     1
#> 202  C202     1     0     0     1     0     0
#> 203  C203     1     0     0     0     0     1
#> 204  C204     0     0     1     1     0     0
#> 205  C205     0     1     0     1     0     0
#> 206  C206     0     0     1     0     0     1
#> 207  C207     1     0     0     1     0     0
#> 208  C208     1     0     0     1     0     0
#> 209  C209    NA    NA    NA    NA    NA    NA
#> 210  C210     1     0     0     0     0     1
#> 211  C211     0     0     1     0     0     1
#> 212  C212     1     0     0     1     0     0
#> 213  C213     0     0     1     0     1     0
#> 214  C214     0     0     1     1     0     0
#> 215  C215     0     0     1     0     1     0
#> 216  C216     0     0     1     1     0     0
#> 217  C217     0     0     1     0     0     1
#> 218  C218     1     0     0     1     0     0
#> 219  C219     1     0     0     1     0     0
#> 220  C220     1     0     0     0     0     1
#> 221  C221     0     0     1     1     0     0
#> 222  C222     1     0     0     1     0     0
#> 223  C223     1     0     0     0     0     1
#> 224  C224     0     0     1     1     0     0
#> 225  C225     1     0     0     1     0     0
#> 226  C226     1     0     0     1     0     0
#> 227  C227     0     0     1     1     0     0
#> 228  C228     0     0     1     0     0     1
#> 229  C229     0     0     1     0     0     1
#> 230  C230     1     0     0     0     0     1
#> 231  C231     0     0     1     1     0     0
#> 232  C232     1     0     0     1     0     0
#> 233  C233     0     0     1     1     0     0
#> 234  C234     1     0     0     1     0     0
#> 235  C235     0     0     1     0     0     1
#> 236  C236     1     0     0     1     0     0
#> 237  C237     1     0     0     0     0     1
#> 238  C238     0     0     1     1     0     0
#> 239  C239     0     0     1     1     0     0
#> 240  C240     1     0     0     1     0     0
#> 241  C241     1     0     0     1     0     0
#> 242  C242     0     1     0     1     0     0
#> 243  C243     1     0     0     1     0     0
#> 244  C244     0     0     1     1     0     0
#> 245  C245     1     0     0     0     0     1
#> 246  C246     1     0     0     1     0     0
#> 247  C247     1     0     0     1     0     0
#> 248  C248    NA    NA    NA    NA    NA    NA
#> 249  C249     1     0     0     1     0     0
#> 250  C250     1     0     0     1     0     0
#> 251  C251     1     0     0     1     0     0
#> 252  C252     1     0     0     1     0     0
#> 253  C253     0     0     1     1     0     0
#> 254  C254     1     0     0     1     0     0
#> 255  C255     0     1     0     0     0     1
#> 256  C256    NA    NA    NA    NA    NA    NA
#> 257  C257     1     0     0     0     0     1
#> 258  C258     1     0     0     1     0     0
#> 259  C259     0     0     1     1     0     0
#> 260  C260     1     0     0     0     0     1
#> 261  C261     1     0     0     1     0     0
#> 262  C262     1     0     0     0     0     1
#> 263  C263     1     0     0     1     0     0
#> 264  C264     1     0     0     0     0     1
#> 265  C265     1     0     0     0     0     1
#> 266  C266    NA    NA    NA    NA    NA    NA
#> 267  C267     1     0     0     0     0     1
#> 268  C268    NA    NA    NA    NA    NA    NA
#> 269  C269     1     0     0     0     0     1
#> 270  C270     1     0     0     0     0     1
#> 271  C271     1     0     0     0     0     1
#> 272  C272     0     0     1     0     0     1
#> 273  C273     0     0     1     1     0     0
#> 274  C274     0     0     1     1     0     0
#> 275  C275    NA    NA    NA    NA    NA    NA
#> 276  C276     1     0     0     1     0     0
#> 277  C277     0     0     1     1     0     0
#> 278  C278     1     0     0     1     0     0
#> 279  C279     1     0     0     1     0     0
#> 280  C280     1     0     0     1     0     0
#> 281  C281     1     0     0     1     0     0
#> 282  C282     1     0     0     1     0     0
#> 283  C283     1     0     0     0     0     1
#> 284  C284     0     0     1     0     1     0
#> 285  C285     1     0     0     0     0     1
#> 286  C286     1     0     0     1     0     0
#> 287  C287    NA    NA    NA    NA    NA    NA
#> 288  C288    NA    NA    NA    NA    NA    NA
#> 289  C289     0     0     1     0     1     0
#> 290  C290     0     0     1     0     0     1
#> 291  C291     0     0     1     1     0     0
#> 292  C292     1     0     0     0     0     1
#> 293  C293     0     1     0     0     0     1
#> 294  C294     0     1     0     1     0     0
#> 295  C295     1     0     0     1     0     0
#> 296  C296     0     1     0     0     0     1
#> 297  C297     1     0     0     0     0     1
#> 298  C298     0     0     1     0     0     1
#> 299  C299     1     0     0     0     0     1
#> 300  C300     1     0     0     1     0     0
#> 301  C301     1     0     0     1     0     0
#> 302  C302     0     0     1     0     0     1
#> 303  C303     0     1     0     0     0     1
#> 304  C304     1     0     0     1     0     0
#> 305  C305     1     0     0     0     0     1
#> 306  C306     1     0     0     0     0     1
#> 307  C307     0     0     1     0     0     1
#> 308  C308     0     1     0     1     0     0
#> 309  C309     0     0     1     1     0     0
#> 310  C310     1     0     0     1     0     0
#> 311  C311     1     0     0     1     0     0
#> 312  C312     1     0     0     1     0     0
#> 313  C313     1     0     0     1     0     0
#> 314  C314     0     0     1     0     0     1
#> 315  C315     0     0     1     0     0     1
#> 316  C316     1     0     0     0     0     1
#> 317  C317     1     0     0     1     0     0
#> 318  C318     0     0     1     1     0     0
#> 319  C319     1     0     0     1     0     0
#> 320  C320     0     0     1     1     0     0
#> 321  C321     0     0     1     0     0     1
#> 322  C322     1     0     0     0     1     0
#> 323  C323     1     0     0     1     0     0
#> 324  C324     1     0     0     0     0     1
#> 325  C325     1     0     0     1     0     0
#> 326  C326     1     0     0     1     0     0
#> 327  C327     1     0     0     0     1     0
#> 328  C328     1     0     0     1     0     0
#> 329  C329     0     0     1     1     0     0
#> 330  C330     1     0     0     1     0     0
#> 331  C331     0     1     0     0     0     1
#> 332  C332     1     0     0     1     0     0
#> 333  C333     1     0     0     1     0     0
#> 334  C334     1     0     0     0     0     1
#> 335  C335     1     0     0     1     0     0
#> 336  C336     1     0     0     1     0     0
#> 337  C337     1     0     0     1     0     0
#> 338  C338     1     0     0     1     0     0
#> 339  C339     0     0     1     0     0     1
#> 340  C340     1     0     0     1     0     0
#> 341  C341     1     0     0     1     0     0
#> 342  C342     0     0     1     0     0     1
#> 343  C343     1     0     0     1     0     0
#> 344  C344     0     0     1     1     0     0
#> 345  C345    NA    NA    NA    NA    NA    NA
#> 346  C346    NA    NA    NA    NA    NA    NA
#> 347  C347     0     0     1     1     0     0
#> 348  C348     1     0     0     1     0     0
#> 349  C349     0     0     1     1     0     0
#> 350  C350     1     0     0     0     0     1
#> 351  C351     1     0     0     0     1     0
#> 352  C352     1     0     0     1     0     0
#> 353  C353     0     0     1     1     0     0
#> 354  C354     0     0     1     1     0     0
#> 355  C355     1     0     0     1     0     0
#> 356  C356     1     0     0     1     0     0
#> 357  C357     0     0     1     1     0     0
#> 358  C358     1     0     0     1     0     0
#> 359  C359    NA    NA    NA    NA    NA    NA
#> 360  C360     0     0     1     0     0     1
#> 361  C361     0     0     1     0     0     1
#> 362  C362     0     0     1     1     0     0
#> 363  C363    NA    NA    NA    NA    NA    NA
#> 364  C364    NA    NA    NA    NA    NA    NA
#> 365  C365     1     0     0     1     0     0
#> 366  C366     1     0     0     1     0     0
#> 367  C367     1     0     0     0     0     1
#> 368  C368     0     0     1     0     1     0
#> 369  C369     1     0     0     0     0     1
#> 370  C370     0     0     1     1     0     0
#> 371  C371     0     0     1     1     0     0
#> 372  C372     0     0     1     1     0     0
#> 373  C373     0     1     0     1     0     0
#> 374  C374     0     0     1     1     0     0
#> 375  C375     1     0     0     1     0     0
#> 376  C376     0     0     1     0     0     1
#> 377  C377     0     0     1     0     0     1
#> 378  C378     0     0     1     1     0     0
#> 379  C379     1     0     0     1     0     0
#> 380  C380     1     0     0     1     0     0
#> 381  C381     0     0     1     1     0     0
#> 382  C382     1     0     0     1     0     0
#> 383  C383     0     0     1     0     0     1
#> 384  C384     0     1     0     1     0     0
#> 385  C385     0     0     1     1     0     0
#> 386  C386     1     0     0     1     0     0
#> 387  C387    NA    NA    NA    NA    NA    NA
#> 388  C388    NA    NA    NA    NA    NA    NA
#> 389  C389     1     0     0     0     0     1
#> 390  C390     1     0     0     1     0     0
#> 391  C391    NA    NA    NA    NA    NA    NA
#> 392  C392     0     0     1     0     0     1
#> 393  C393     0     0     1     1     0     0
#> 394  C394     0     0     1     1     0     0
#> 395  C395     1     0     0     1     0     0
#> 396  C396     0     1     0     1     0     0
#> 397  C397     0     0     1     0     0     1
#> 398  C398     1     0     0     0     1     0
#> 399  C399     0     0     1     1     0     0
#> 400  C400     0     0     1     1     0     0
#> 401  C401     1     0     0     1     0     0
#> 402  C402     0     0     1     1     0     0
#> 403  C403     0     0     1     0     0     1
#> 404  C404     0     0     1     1     0     0
#> 405  C405     1     0     0     1     0     0
#> 406  C406     0     0     1     0     0     1
#> 407  C407     1     0     0     1     0     0
#> 408  C408     1     0     0     1     0     0
#> 409  C409     1     0     0     0     0     1
#> 410  C410     1     0     0     1     0     0
#> 411  C411     0     0     1     0     0     1
#> 412  C412    NA    NA    NA    NA    NA    NA
#> 413  C413     0     0     1     1     0     0
#> 414  C414     1     0     0     0     1     0
#> 415  C415     1     0     0     1     0     0
#> 416  C416     1     0     0     1     0     0
#> 417  C417     0     1     0     1     0     0
#> 418  C418     1     0     0     1     0     0
#> 419  C419     0     0     1     1     0     0
#> 420  C420     1     0     0     0     0     1
#> 421  C421     0     0     1     1     0     0
#> 422  C422     1     0     0     0     1     0
#> 423  C423     0     0     1     0     0     1
#> 424  C424     1     0     0     1     0     0
#> 425  C425     0     1     0     1     0     0
#> 426  C426     0     0     1     1     0     0
#> 427  C427     0     0     1     1     0     0
#> 428  C428     0     0     1     1     0     0
#> 429  C429     1     0     0     1     0     0
#> 430  C430     1     0     0     1     0     0
#> 431  C431     1     0     0     0     0     1
#> 432  C432     1     0     0     0     1     0
#> 433  C433     0     0     1     1     0     0
#> 434  C434     0     1     0     0     0     1
#> 435  C435     1     0     0     1     0     0
#> 436  C436    NA    NA    NA    NA    NA    NA
#> 437  C437    NA    NA    NA    NA    NA    NA
#> 438  C438     0     0     1     0     1     0
#> 439  C439     0     0     1     1     0     0
#> 440  C440     1     0     0     1     0     0
#> 441  C441     0     0     1     0     0     1
#> 442  C442     0     0     1     1     0     0
#> 443  C443     0     0     1     0     0     1
#> 444  C444     1     0     0     1     0     0
#> 445  C445     0     0     1     0     0     1
#> 446  C446     1     0     0     0     0     1
#> 447  C447     1     0     0     0     0     1
#> 448  C448     1     0     0     0     0     1
#> 449  C449     0     0     1     0     0     1
#> 450  C450     1     0     0     0     1     0
#> 451  C451     1     0     0     1     0     0
#> 452  C452     1     0     0     1     0     0
#> 453  C453     1     0     0     1     0     0
#> 454  C454     0     0     1     0     0     1
#> 455  C455     1     0     0     1     0     0
#> 456  C456     1     0     0     1     0     0
#> 457  C457     0     1     0     0     0     1
#> 458  C458     1     0     0     1     0     0
#> 459  C459     1     0     0     1     0     0
#> 460  C460     0     0     1     0     0     1
#> 461  C461    NA    NA    NA    NA    NA    NA
#> 462  C462     1     0     0     1     0     0
#> 463  C463     0     1     0     1     0     0
#> 464  C464     1     0     0     1     0     0
#> 465  C465     1     0     0     1     0     0
#> 466  C466     1     0     0     1     0     0
#> 467  C467     1     0     0     1     0     0
#> 468  C468     1     0     0     1     0     0
#> 469  C469     0     0     1     1     0     0
#> 470  C470     1     0     0     1     0     0
#> 471  C471     0     0     1     0     0     1
#> 472  C472     1     0     0     1     0     0
#> 473  C473     0     0     1     1     0     0
#> 474  C474     0     0     1     1     0     0
#> 475  C475     1     0     0     1     0     0
#> 476  C476     1     0     0     0     0     1
#> 477  C477     1     0     0     1     0     0
#> 478  C478     1     0     0     1     0     0
#> 479  C479     1     0     0     0     0     1
#> 480  C480     0     0     1     0     0     1
#> 481  C481     1     0     0     1     0     0
#> 482  C482     0     0     1     0     0     1
#> 483  C483     1     0     0     0     0     1
#> 484  C484     0     0     1     1     0     0
#> 485  C485     1     0     0     1     0     0
#> 486  C486     1     0     0     1     0     0
#> 487  C487     1     0     0     0     0     1
#> 488  C488     1     0     0     1     0     0
#> 489  C489     1     0     0     1     0     0
#> 490  C490     1     0     0     0     0     1
#> 491  C491    NA    NA    NA    NA    NA    NA
#> 492  C492     0     0     1     1     0     0
#> 493  C493     1     0     0     0     0     1
#> 494  C494     0     0     1     1     0     0
#> 495  C495     1     0     0     1     0     0
#> 496  C496     0     0     1     1     0     0
#> 497  C497     1     0     0     0     0     1
#> 498  C498     1     0     0     1     0     0
#> 499  C499     1     0     0     1     0     0
#> 500  C500    NA    NA    NA    NA    NA    NA
#> 501  P001     0     0     1     1     0     0
#> 502  P002     0     0     1     1     0     0
#> 503  P003     1     0     0     1     0     0
#> 504  P004     1     0     0     1     0     0
#> 505  P005     0     0     1     0     0     1
#> 506  P006     0     1     0     0     0     1
#> 507  P007     1     0     0     1     0     0
#> 508  P008     0     0     1     0     0     1
#> 509  P009     0     0     1     1     0     0
#> 510  P010     0     0     1     1     0     0
#> 511  P011     0     0     1     1     0     0
#> 512  P012    NA    NA    NA    NA    NA    NA
#> 513  P013     1     0     0     0     0     1
#> 514  P014     1     0     0     1     0     0
#> 515  P015     1     0     0     1     0     0
#> 516  P016     1     0     0     0     0     1
#> 517  P017     1     0     0     1     0     0
#> 518  P018    NA    NA    NA    NA    NA    NA
#> 519  P019     1     0     0     1     0     0
#> 520  P020     0     0     1     0     0     1
#> 521  P021     1     0     0     1     0     0
#> 522  P022     1     0     0     1     0     0
#> 523  P023     0     0     1     0     0     1
#> 524  P024     1     0     0     1     0     0
#> 525  P025     1     0     0     1     0     0
#> 526  P026     0     1     0     1     0     0
#> 527  P027     0     0     1     1     0     0
#> 528  P028     0     1     0     1     0     0
#> 529  P029     1     0     0     1     0     0
#> 530  P030     1     0     0     1     0     0
#> 531  P031     0     1     0     0     0     1
#> 532  P032     0     0     1     0     1     0
#> 533  P033     0     0     1     0     0     1
#> 534  P034     0     1     0     1     0     0
#> 535  P035     1     0     0     0     0     1
#> 536  P036     0     1     0     0     1     0
#> 537  P037     0     0     1     1     0     0
#> 538  P038     1     0     0     1     0     0
#> 539  P039     1     0     0     1     0     0
#> 540  P040     1     0     0     0     0     1
#> 541  P041    NA    NA    NA    NA    NA    NA
#> 542  P042     0     1     0     1     0     0
#> 543  P043     1     0     0     1     0     0
#> 544  P044     1     0     0     1     0     0
#> 545  P045     0     0     1     0     0     1
#> 546  P046     1     0     0     1     0     0
#> 547  P047     1     0     0     1     0     0
#> 548  P048     0     0     1     0     0     1
#> 549  P049     1     0     0     1     0     0
#> 550  P050     1     0     0     0     0     1
#> 551  P051     1     0     0     0     0     1
#> 552  P052     1     0     0     1     0     0
#> 553  P053     1     0     0     1     0     0
#> 554  P054     0     0     1     1     0     0
#> 555  P055     0     0     1     0     1     0
#> 556  P056     1     0     0     1     0     0
#> 557  P057     0     0     1     0     0     1
#> 558  P058    NA    NA    NA    NA    NA    NA
#> 559  P059     0     0     1     0     0     1
#> 560  P060     0     0     1     1     0     0
#> 561  P061     0     0     1     0     0     1
#> 562  P062    NA    NA    NA    NA    NA    NA
#> 563  P063     0     0     1     1     0     0
#> 564  P064     0     1     0     0     0     1
#> 565  P065     0     0     1     0     0     1
#> 566  P066     0     0     1     1     0     0
#> 567  P067     0     1     0     1     0     0
#> 568  P068     1     0     0     0     0     1
#> 569  P069     0     0     1     1     0     0
#> 570  P070     1     0     0     1     0     0
#> 571  P071     0     1     0     0     0     1
#> 572  P072     0     0     1     1     0     0
#> 573  P073     1     0     0     1     0     0
#> 574  P074     0     0     1     1     0     0
#> 575  P075     1     0     0     1     0     0
#> 576  P076     0     0     1     0     0     1
#> 577  P077     0     0     1     0     0     1
#> 578  P078     1     0     0     0     0     1
#> 579  P079     1     0     0     1     0     0
#> 580  P080    NA    NA    NA    NA    NA    NA
#> 581  P081     0     0     1     1     0     0
#> 582  P082     0     0     1     0     0     1
#> 583  P083     0     0     1     1     0     0
#> 584  P084     0     0     1     0     0     1
#> 585  P085    NA    NA    NA    NA    NA    NA
#> 586  P086     0     0     1     0     1     0
#> 587  P087     0     0     1     0     0     1
#> 588  P088     1     0     0     1     0     0
#> 589  P089     1     0     0     1     0     0
#> 590  P090     0     0     1     1     0     0
#> 591  P091     0     0     1     1     0     0
#> 592  P092     1     0     0     0     0     1
#> 593  P093     1     0     0     1     0     0
#> 594  P094     1     0     0     1     0     0
#> 595  P095     1     0     0     0     0     1
#> 596  P096     1     0     0     1     0     0
#> 597  P097     1     0     0     1     0     0
#> 598  P098     1     0     0     1     0     0
#> 599  P099     0     0     1     0     0     1
#> 600  P100     0     0     1     1     0     0
#> 601  P101     0     0     1     1     0     0
#> 602  P102     1     0     0     0     0     1
#> 603  P103     0     0     1     0     1     0
#> 604  P104     0     0     1     1     0     0
#> 605  P105     1     0     0     1     0     0
#> 606  P106     1     0     0     0     0     1
#> 607  P107     0     0     1     0     0     1
#> 608  P108     1     0     0     1     0     0
#> 609  P109     0     0     1     0     1     0
#> 610  P110     1     0     0     0     0     1
#> 611  P111     0     0     1     1     0     0
#> 612  P112     0     0     1     0     0     1
#> 613  P113     1     0     0     1     0     0
#> 614  P114     1     0     0     0     0     1
#> 615  P115     1     0     0     1     0     0
#> 616  P116    NA    NA    NA    NA    NA    NA
#> 617  P117     0     0     1     0     0     1
#> 618  P118     0     0     1     1     0     0
#> 619  P119     0     0     1     1     0     0
#> 620  P120     1     0     0     0     0     1
#> 621  P121     0     0     1     0     0     1
#> 622  P122     1     0     0     1     0     0
#> 623  P123     0     0     1     1     0     0
#> 624  P124     0     0     1     1     0     0
#> 625  P125     1     0     0     1     0     0
#> 626  P126     1     0     0     0     0     1
#> 627  P127     1     0     0     0     0     1
#> 628  P128     1     0     0     1     0     0
#> 629  P129     0     1     0     0     0     1
#> 630  P130     0     0     1     1     0     0
#> 631  P131     0     0     1     1     0     0
#> 632  P132     1     0     0     0     0     1
#> 633  P133     1     0     0     0     1     0
#> 634  P134     1     0     0     1     0     0
#> 635  P135     1     0     0     1     0     0
#> 636  P136    NA    NA    NA    NA    NA    NA
#> 637  P137     1     0     0     0     0     1
#> 638  P138     0     0     1     0     1     0
#> 639  P139     0     0     1     1     0     0
#> 640  P140     1     0     0     1     0     0
#> 641  P141     0     0     1     1     0     0
#> 642  P142     0     0     1     1     0     0
#> 643  P143     0     0     1     0     0     1
#> 644  P144     1     0     0     1     0     0
#> 645  P145     1     0     0     1     0     0
#> 646  P146     1     0     0     1     0     0
#> 647  P147     0     0     1     1     0     0
#> 648  P148     0     0     1     1     0     0
#> 649  P149     1     0     0     1     0     0
#> 650  P150     1     0     0     0     0     1
#> 651  P151     1     0     0     1     0     0
#> 652  P152     0     1     0     0     0     1
#> 653  P153     1     0     0     1     0     0
#> 654  P154     0     0     1     1     0     0
#> 655  P155     1     0     0     1     0     0
#> 656  P156     1     0     0     1     0     0
#> 657  P157     1     0     0     1     0     0
#> 658  P158     0     0     1     1     0     0
#> 659  P159     1     0     0     0     0     1
#> 660  P160     0     0     1     1     0     0
#> 661  P161     1     0     0     1     0     0
#> 662  P162     1     0     0     1     0     0
#> 663  P163     1     0     0     1     0     0
#> 664  P164     0     1     0     0     0     1
#> 665  P165     1     0     0     1     0     0
#> 666  P166     0     0     1     0     0     1
#> 667  P167     0     0     1     0     0     1
#> 668  P168     0     0     1     0     0     1
#> 669  P169     0     0     1     0     0     1
#> 670  P170     1     0     0     0     0     1
#> 671  P171     0     0     1     1     0     0
#> 672  P172     1     0     0     1     0     0
#> 673  P173    NA    NA    NA    NA    NA    NA
#> 674  P174     1     0     0     1     0     0
#> 675  P175     0     0     1     0     0     1
#> 676  P176     0     0     1     1     0     0
#> 677  P177     1     0     0     1     0     0
#> 678  P178    NA    NA    NA    NA    NA    NA
#> 679  P179     0     0     1     1     0     0
#> 680  P180     0     0     1     1     0     0
#> 681  P181     0     0     1     0     0     1
#> 682  P182     1     0     0     1     0     0
#> 683  P183     0     0     1     1     0     0
#> 684  P184     1     0     0     1     0     0
#> 685  P185     0     0     1     1     0     0
#> 686  P186     1     0     0     0     0     1
#> 687  P187     1     0     0     1     0     0
#> 688  P188     0     0     1     1     0     0
#> 689  P189     1     0     0     0     1     0
#> 690  P190    NA    NA    NA    NA    NA    NA
#> 691  P191    NA    NA    NA    NA    NA    NA
#> 692  P192     1     0     0     1     0     0
#> 693  P193     0     1     0     0     0     1
#> 694  P194     0     0     1     1     0     0
#> 695  P195     1     0     0     1     0     0
#> 696  P196     0     1     0     1     0     0
#> 697  P197     0     1     0     0     0     1
#> 698  P198     1     0     0     1     0     0
#> 699  P199     0     1     0     1     0     0
#> 700  P200     1     0     0     1     0     0
#> 701  P201     1     0     0     1     0     0
#> 702  P202     0     0     1     0     0     1
#> 703  P203     0     1     0     0     0     1
#> 704  P204     0     1     0     1     0     0
#> 705  P205     0     1     0     0     0     1
#> 706  P206     0     1     0     0     0     1
#> 707  P207     1     0     0     0     0     1
#> 708  P208     0     0     1     1     0     0
#> 709  P209     0     0     1     1     0     0
#> 710  P210     0     0     1     1     0     0
#> 711  P211     0     1     0     1     0     0
#> 712  P212     0     0     1     1     0     0
#> 713  P213     1     0     0     0     0     1
#> 714  P214     1     0     0     1     0     0
#> 715  P215     1     0     0     1     0     0
#> 716  P216     0     0     1     0     0     1
#> 717  P217     1     0     0     1     0     0
#> 718  P218     1     0     0     0     0     1
#> 719  P219     0     0     1     1     0     0
#> 720  P220     1     0     0     0     0     1
#> 721  P221    NA    NA    NA    NA    NA    NA
#> 722  P222     0     0     1     1     0     0
#> 723  P223     1     0     0     1     0     0
#> 724  P224     0     0     1     0     0     1
#> 725  P225     0     1     0     1     0     0
#> 726  P226     1     0     0     1     0     0
#> 727  P227     0     0     1     0     0     1
#> 728  P228     1     0     0     1     0     0
#> 729  P229     1     0     0     1     0     0
#> 730  P230     1     0     0     1     0     0
#> 731  P231     0     0     1     0     1     0
#> 732  P232     0     0     1     1     0     0
#> 733  P233     1     0     0     1     0     0
#> 734  P234     0     0     1     1     0     0
#> 735  P235     1     0     0     1     0     0
#> 736  P236     1     0     0     1     0     0
#> 737  P237    NA    NA    NA    NA    NA    NA
#> 738  P238     0     0     1     1     0     0
#> 739  P239     1     0     0     0     0     1
#> 740  P240     0     0     1     0     0     1
#> 741  P241     1     0     0     1     0     0
#> 742  P242     1     0     0     1     0     0
#> 743  P243     1     0     0     1     0     0
#> 744  P244     0     0     1     1     0     0
#> 745  P245     1     0     0     1     0     0
#> 746  P246     0     0     1     0     0     1
#> 747  P247    NA    NA    NA    NA    NA    NA
#> 748  P248     1     0     0     1     0     0
#> 749  P249     0     1     0     0     0     1
#> 750  P250     0     0     1     0     0     1
#> 751  P251     1     0     0     0     0     1
#> 752  P252     1     0     0     1     0     0
#> 753  P253     0     0     1     0     1     0
#> 754  P254     1     0     0     1     0     0
#> 755  P255     1     0     0     1     0     0
#> 756  P256     0     1     0     0     0     1
#> 757  P257     0     0     1     0     0     1
#> 758  P258     0     0     1     0     0     1
#> 759  P259     0     0     1     0     0     1
#> 760  P260     1     0     0     1     0     0
#> 761  P261     1     0     0     1     0     0
#> 762  P262     0     0     1     0     1     0
#> 763  P263     1     0     0     1     0     0
#> 764  P264     0     1     0     0     0     1
#> 765  P265    NA    NA    NA    NA    NA    NA
#> 766  P266     0     0     1     0     1     0
#> 767  P267     0     1     0     1     0     0
#> 768  P268     0     0     1     0     0     1
#> 769  P269     0     0     1     0     0     1
#> 770  P270     1     0     0     1     0     0
#> 771  P271     0     1     0     1     0     0
#> 772  P272     0     0     1     0     0     1
#> 773  P273    NA    NA    NA    NA    NA    NA
#> 774  P274    NA    NA    NA    NA    NA    NA
#> 775  P275     0     0     1     0     0     1
#> 776  P276     1     0     0     0     0     1
#> 777  P277     0     0     1     0     0     1
#> 778  P278     0     1     0     1     0     0
#> 779  P279     0     0     1     0     0     1
#> 780  P280     0     0     1     1     0     0
#> 781  P281     0     0     1     1     0     0
#> 782  P282     0     0     1     0     0     1
#> 783  P283     0     1     0     1     0     0
#> 784  P284    NA    NA    NA    NA    NA    NA
#> 785  P285     0     0     1     0     0     1
#> 786  P286     1     0     0     0     0     1
#> 787  P287     1     0     0     0     0     1
#> 788  P288     1     0     0     0     0     1
#> 789  P289     0     0     1     0     1     0
#> 790  P290     0     0     1     1     0     0
#> 791  P291     0     0     1     1     0     0
#> 792  P292     1     0     0     1     0     0
#> 793  P293    NA    NA    NA    NA    NA    NA
#> 794  P294     1     0     0     0     0     1
#> 795  P295     1     0     0     1     0     0
#> 796  P296     1     0     0     1     0     0
#> 797  P297     0     1     0     0     0     1
#> 798  P298     1     0     0     1     0     0
#> 799  P299     1     0     0     0     0     1
#> 800  P300     1     0     0     1     0     0
#> 801  P301     0     1     0     1     0     0
#> 802  P302     0     0     1     0     0     1
#> 803  P303     0     1     0     0     0     1
#> 804  P304     1     0     0     0     0     1
#> 805  P305     0     0     1     0     0     1
#> 806  P306     1     0     0     1     0     0
#> 807  P307     1     0     0     1     0     0
#> 808  P308     1     0     0     1     0     0
#> 809  P309     0     0     1     0     0     1
#> 810  P310     1     0     0     0     0     1
#> 811  P311     0     0     1     1     0     0
#> 812  P312     1     0     0     0     0     1
#> 813  P313     1     0     0     1     0     0
#> 814  P314     1     0     0     1     0     0
#> 815  P315     1     0     0     1     0     0
#> 816  P316     0     0     1     0     0     1
#> 817  P317     0     0     1     0     0     1
#> 818  P318     0     0     1     0     0     1
#> 819  P319    NA    NA    NA    NA    NA    NA
#> 820  P320     0     0     1     1     0     0
#> 821  P321     1     0     0     0     0     1
#> 822  P322     1     0     0     1     0     0
#> 823  P323     0     0     1     0     1     0
#> 824  P324     0     0     1     0     0     1
#> 825  P325    NA    NA    NA    NA    NA    NA
#> 826  P326     0     0     1     1     0     0
#> 827  P327     0     0     1     0     0     1
#> 828  P328     1     0     0     1     0     0
#> 829  P329     0     1     0     0     1     0
#> 830  P330     1     0     0     1     0     0
#> 831  P331     1     0     0     1     0     0
#> 832  P332     0     0     1     1     0     0
#> 833  P333     0     0     1     1     0     0
#> 834  P334     0     0     1     0     0     1
#> 835  P335     1     0     0     1     0     0
#> 836  P336     1     0     0     1     0     0
#> 837  P337    NA    NA    NA    NA    NA    NA
#> 838  P338     0     1     0     0     0     1
#> 839  P339     0     1     0     0     0     1
#> 840  P340     1     0     0     0     0     1
#> 841  P341     1     0     0     1     0     0
#> 842  P342     1     0     0     1     0     0
#> 843  P343     0     1     0     0     0     1
#> 844  P344     1     0     0     1     0     0
#> 845  P345     1     0     0     1     0     0
#> 846  P346     0     0     1     1     0     0
#> 847  P347     0     0     1     0     0     1
#> 848  P348    NA    NA    NA    NA    NA    NA
#> 849  P349     1     0     0     0     0     1
#> 850  P350     0     0     1     0     0     1
#> 851  P351     0     0     1     1     0     0
#> 852  P352     0     1     0     1     0     0
#> 853  P353     1     0     0     1     0     0
#> 854  P354     1     0     0     1     0     0
#> 855  P355     1     0     0     0     0     1
#> 856  P356     0     0     1     1     0     0
#> 857  P357     1     0     0     1     0     0
#> 858  P358     0     0     1     1     0     0
#> 859  P359    NA    NA    NA    NA    NA    NA
#> 860  P360     0     0     1     1     0     0
#> 861  P361     0     1     0     1     0     0
#> 862  P362     1     0     0     1     0     0
#> 863  P363     1     0     0     1     0     0
#> 864  P364     0     0     1     1     0     0
#> 865  P365     0     1     0     1     0     0
#> 866  P366    NA    NA    NA    NA    NA    NA
#> 867  P367     1     0     0     1     0     0
#> 868  P368    NA    NA    NA    NA    NA    NA
#> 869  P369     0     0     1     1     0     0
#> 870  P370     0     0     1     1     0     0
#> 871  P371     1     0     0     1     0     0
#> 872  P372     1     0     0     1     0     0
#> 873  P373     0     0     1     1     0     0
#> 874  P374     0     0     1     1     0     0
#> 875  P375     1     0     0     0     0     1
#> 876  P376     0     0     1     1     0     0
#> 877  P377     1     0     0     1     0     0
#> 878  P378     0     1     0     0     0     1
#> 879  P379     0     1     0     0     1     0
#> 880  P380     1     0     0     1     0     0
#> 881  P381     1     0     0     1     0     0
#> 882  P382     1     0     0     1     0     0
#> 883  P383     0     0     1     0     0     1
#> 884  P384     1     0     0     1     0     0
#> 885  P385     1     0     0     1     0     0
#> 886  P386     1     0     0     1     0     0
#> 887  P387     1     0     0     1     0     0
#> 888  P388     0     1     0     0     0     1
#> 889  P389     1     0     0     1     0     0
#> 890  P390     1     0     0     0     0     1
#> 891  P391     1     0     0     1     0     0
#> 892  P392     0     1     0     1     0     0
#> 893  P393     1     0     0     1     0     0
#> 894  P394     1     0     0     0     0     1
#> 895  P395     1     0     0     1     0     0
#> 896  P396     1     0     0     1     0     0
#> 897  P397    NA    NA    NA    NA    NA    NA
#> 898  P398     1     0     0     0     0     1
#> 899  P399     0     0     1     1     0     0
#> 900  P400     1     0     0     1     0     0
#> 901  P401     1     0     0     1     0     0
#> 902  P402     1     0     0     0     0     1
#> 903  P403     0     0     1     1     0     0
#> 904  P404    NA    NA    NA    NA    NA    NA
#> 905  P405     1     0     0     0     0     1
#> 906  P406     1     0     0     1     0     0
#> 907  P407     0     0     1     1     0     0
#> 908  P408     0     0     1     0     0     1
#> 909  P409     0     1     0     1     0     0
#> 910  P410     0     0     1     0     0     1
#> 911  P411     0     1     0     0     0     1
#> 912  P412     0     0     1     1     0     0
#> 913  P413     1     0     0     0     0     1
#> 914  P414     0     0     1     1     0     0
#> 915  P415     1     0     0     1     0     0
#> 916  P416     1     0     0     0     0     1
#> 917  P417     0     0     1     1     0     0
#> 918  P418     0     0     1     0     0     1
#> 919  P419     1     0     0     1     0     0
#> 920  P420     1     0     0     1     0     0
#> 921  P421     1     0     0     0     1     0
#> 922  P422     1     0     0     1     0     0
#> 923  P423     1     0     0     1     0     0
#> 924  P424     0     0     1     1     0     0
#> 925  P425     0     0     1     0     0     1
#> 926  P426     1     0     0     0     0     1
#> 927  P427     1     0     0     1     0     0
#> 928  P428     0     0     1     1     0     0
#> 929  P429     1     0     0     1     0     0
#> 930  P430     1     0     0     0     0     1
#> 931  P431     0     1     0     1     0     0
#> 932  P432     0     0     1     1     0     0
#> 933  P433     0     0     1     1     0     0
#> 934  P434     1     0     0     0     0     1
#> 935  P435     0     0     1     0     0     1
#> 936  P436     0     0     1     1     0     0
#> 937  P437     1     0     0     1     0     0
#> 938  P438     1     0     0     0     0     1
#> 939  P439     0     0     1     0     0     1
#> 940  P440     1     0     0     0     0     1
#> 941  P441     0     0     1     1     0     0
#> 942  P442     0     0     1     0     1     0
#> 943  P443     1     0     0     1     0     0
#> 944  P444     1     0     0     1     0     0
#> 945  P445     1     0     0     0     0     1
#> 946  P446     0     0     1     0     0     1
#> 947  P447     0     0     1     1     0     0
#> 948  P448     0     0     1     0     0     1
#> 949  P449     1     0     0     1     0     0
#> 950  P450     0     0     1     0     0     1
#> 951  P451     1     0     0     1     0     0
#> 952  P452     0     0     1     0     0     1
#> 953  P453     1     0     0     1     0     0
#> 954  P454     0     0     1     0     0     1
#> 955  P455     0     0     1     0     0     1
#> 956  P456     1     0     0     1     0     0
#> 957  P457     0     0     1     0     0     1
#> 958  P458     0     0     1     1     0     0
#> 959  P459     0     1     0     0     0     1
#> 960  P460     1     0     0     0     0     1
#> 961  P461     0     0     1     1     0     0
#> 962  P462     0     1     0     0     0     1
#> 963  P463     0     0     1     1     0     0
#> 964  P464     0     0     1     0     0     1
#> 965  P465     1     0     0     0     0     1
#> 966  P466    NA    NA    NA    NA    NA    NA
#> 967  P467     0     0     1     0     0     1
#> 968  P468     0     0     1     1     0     0
#> 969  P469     0     0     1     1     0     0
#> 970  P470     0     0     1     1     0     0
#> 971  P471     0     0     1     0     0     1
#> 972  P472     1     0     0     1     0     0
#> 973  P473     1     0     0     1     0     0
#> 974  P474     0     0     1     0     0     1
#> 975  P475     0     0     1     0     0     1
#> 976  P476     1     0     0     0     0     1
#> 977  P477     1     0     0     0     0     1
#> 978  P478     0     0     1     0     0     1
#> 979  P479     1     0     0     1     0     0
#> 980  P480     0     1     0     1     0     0
#> 981  P481     1     0     0     1     0     0
#> 982  P482    NA    NA    NA    NA    NA    NA
#> 983  P483     1     0     0     1     0     0
#> 984  P484     0     1     0     1     0     0
#> 985  P485     0     0     1     0     0     1
#> 986  P486     1     0     0     0     0     1
#> 987  P487     0     0     1     1     0     0
#> 988  P488     0     1     0     0     0     1
#> 989  P489     1     0     0     1     0     0
#> 990  P490     1     0     0     1     0     0
#> 991  P491     1     0     0     1     0     0
#> 992  P492     1     0     0     1     0     0
#> 993  P493     0     0     1     0     1     0
#> 994  P494     0     0     1     1     0     0
#> 995  P495     1     0     0     1     0     0
#> 996  P496     1     0     0     1     0     0
#> 997  P497     1     0     0     1     0     0
#> 998  P498     0     0     1     0     0     1
#> 999  P499     1     0     0     0     0     1
#> 1000 P500     0     0     1     0     0     1