Calculate amino acid frequencies

getAAFrequencies calculates amino acid frequencies in amino acid data frame.

getAAFrequencies(aa_variation)

Arguments

aa_variation
Amino acid variation data frame as returned by hlaToAAVariation.

Value

Data frame with each row holding specific amino acid position, it's count and frequency.

Details

Both gene copies are taken into consideration for frequencies calculation, frequency = n / (2 * j) where n is the number of amino acid occurrences and j is the number of samples in aa_variation.

Examples

aa_variation <- hlaToAAVariation(MiDAS_tut_HLA)
getAAFrequencies(aa_variation)

#>          aa_pos Counts   Freq
#> 1       A_-22_V   1981 0.9905
#> 2       A_-22_*     18 0.0090
#> 3       A_-22_I      1 0.0005
#> 4       A_-21_M   1980 0.9900
#> 5       A_-21_V      2 0.0010
#> 6       A_-21_*     18 0.0090
#> 7       A_-18_R   1981 0.9905
#> 8       A_-18_*     18 0.0090
#> 9       A_-18_P      1 0.0005
#> 10      A_-16_L   1981 0.9905
#> 11      A_-16_*     18 0.0090
#> 12      A_-16_I      1 0.0005
#> 13      A_-15_V   1102 0.5510
#> 14      A_-15_L    880 0.4400
#> 15      A_-15_*     18 0.0090
#> 16      A_-11_S   1765 0.8825
#> 17      A_-11_L    217 0.1085
#> 18      A_-11_*     18 0.0090
#> 19       A_-5_T   1981 0.9905
#> 20       A_-5_*     18 0.0090
#> 21       A_-5_P      1 0.0005
#> 22       A_-2_W   1980 0.9900
#> 23       A_-2_*     18 0.0090
#> 24       A_-2_R      2 0.0010
#> 25        A_3_H   1996 0.9980
#> 26        A_3_Q      4 0.0020
#> 27        A_6_R   1999 0.9995
#> 28        A_6_G      1 0.0005
#> 29        A_9_Y    429 0.2145
#> 30        A_9_F   1037 0.5185
#> 31        A_9_S    379 0.1895
#> 32        A_9_T    155 0.0775
#> 33       A_12_V   1979 0.9895
#> 34       A_12_M     21 0.0105
#> 35       A_17_R   1914 0.9570
#> 36       A_17_S     86 0.0430
#> 37       A_19_E   1999 0.9995
#> 38       A_19_K      1 0.0005
#> 39       A_29_D   1999 0.9995
#> 40       A_29_A      1 0.0005
#> 41       A_43_Q   1978 0.9890
#> 42       A_43_R     22 0.0110
#> 43       A_44_R   1754 0.8770
#> 44       A_44_K    246 0.1230
#> 45       A_54_Q   1998 0.9990
#> 46       A_54_R      2 0.0010
#> 47       A_56_G   1857 0.9285
#> 48       A_56_R    143 0.0715
#> 49       A_62_R    299 0.1495
#> 50       A_62_G    555 0.2775
#> 51       A_62_E    294 0.1470
#> 52       A_62_Q    798 0.3990
#> 53       A_62_L     54 0.0270
#> 54       A_63_N    299 0.1495
#> 55       A_63_E   1647 0.8235
#> 56       A_63_Q     54 0.0270
#> 57       A_65_R   1706 0.8530
#> 58       A_65_G    294 0.1470
#> 59       A_66_N   1152 0.5760
#> 60       A_66_K    848 0.4240
#> 61       A_67_V   1754 0.8770
#> 62       A_67_M    246 0.1230
#> 63       A_70_H   1460 0.7300
#> 64       A_70_Q    540 0.2700
#> 65       A_73_T   1900 0.9500
#> 66       A_73_I    100 0.0500
#> 67       A_74_D   1443 0.7215
#> 68       A_74_H    557 0.2785
#> 69       A_76_E    458 0.2290
#> 70       A_76_V   1143 0.5715
#> 71       A_76_A    399 0.1995
#> 72       A_77_S    118 0.0590
#> 73       A_77_D   1143 0.5715
#> 74       A_77_N    739 0.3695
#> 75       A_79_R    411 0.2055
#> 76       A_79_G   1589 0.7945
#> 77       A_80_I    411 0.2055
#> 78       A_80_T   1589 0.7945
#> 79       A_81_A    411 0.2055
#> 80       A_81_L   1589 0.7945
#> 81       A_82_L    411 0.2055
#> 82       A_82_R   1589 0.7945
#> 83       A_83_R    411 0.2055
#> 84       A_83_G   1589 0.7945
#> 85       A_90_D    542 0.2710
#> 86       A_90_A   1458 0.7290
#> 87       A_95_I   1146 0.5730
#> 88       A_95_V    539 0.2695
#> 89       A_95_L    315 0.1575
#> 90       A_97_R    749 0.3745
#> 91       A_97_M    571 0.2855
#> 92       A_97_I    680 0.3400
#> 93       A_99_Y   1697 0.8485
#> 94       A_99_F    296 0.1480
#> 95       A_99_C      7 0.0035
#> 96      A_102_D   1994 0.9970
#> 97      A_102_H      6 0.0030
#> 98      A_103_V   1999 0.9995
#> 99      A_103_L      1 0.0005
#> 100     A_105_P    623 0.3115
#> 101     A_105_S   1377 0.6885
#> 102     A_107_G   1442 0.7210
#> 103     A_107_W    558 0.2790
#> 104     A_109_F   1935 0.9675
#> 105     A_109_L     65 0.0325
#> 106     A_114_Q    332 0.1660
#> 107     A_114_H    875 0.4375
#> 108     A_114_R    709 0.3545
#> 109     A_114_E     84 0.0420
#> 110     A_116_D   1036 0.5180
#> 111     A_116_Y    874 0.4370
#> 112     A_116_H     84 0.0420
#> 113     A_116_N      4 0.0020
#> 114     A_116_F      1 0.0005
#> 115     A_116_V      1 0.0005
#> 116     A_127_N   1066 0.5330
#> 117     A_127_K    934 0.4670
#> 118     A_142_I   1359 0.6795
#> 119     A_142_T    641 0.3205
#> 120     A_144_Q    525 0.2625
#> 121     A_144_K   1475 0.7375
#> 122     A_145_R   1359 0.6795
#> 123     A_145_H    641 0.3205
#> 124     A_149_T    177 0.0885
#> 125     A_149_A   1823 0.9115
#> 126     A_150_A   1754 0.8770
#> 127     A_150_V    246 0.1230
#> 128     A_151_H   1648 0.8240
#> 129     A_151_R    351 0.1755
#> 130     A_151_L      1 0.0005
#> 131     A_152_E    386 0.1930
#> 132     A_152_V   1163 0.5815
#> 133     A_152_A    369 0.1845
#> 134     A_152_R     45 0.0225
#> 135     A_152_W     37 0.0185
#> 136     A_156_W    276 0.1380
#> 137     A_156_L   1092 0.5460
#> 138     A_156_Q    385 0.1925
#> 139     A_156_R    247 0.1235
#> 140     A_158_A   1754 0.8770
#> 141     A_158_V    246 0.1230
#> 142     A_161_E   1781 0.8905
#> 143     A_161_D    219 0.1095
#> 144     A_163_R    535 0.2675
#> 145     A_163_T   1464 0.7320
#> 146     A_163_E      1 0.0005
#> 147     A_164_C   1999 0.9995
#> 148     A_164_F      1 0.0005
#> 149     A_166_E   1463 0.7315
#> 150     A_166_D    537 0.2685
#> 151     A_167_W   1463 0.7315
#> 152     A_167_G    537 0.2685
#> 153     A_171_Y   1977 0.9885
#> 154     A_171_H     23 0.0115
#> 155     A_182_T   1999 0.9995
#> 156     A_182_A      1 0.0005
#> 157     A_184_A    929 0.4645
#> 158     A_184_P   1061 0.5305
#> 159     A_184_*     10 0.0050
#> 160     A_186_K   1967 0.9835
#> 161     A_186_R     23 0.0115
#> 162     A_186_*     10 0.0050
#> 163     A_193_A   1029 0.5145
#> 164     A_193_P    961 0.4805
#> 165     A_193_*     10 0.0050
#> 166     A_194_V   1029 0.5145
#> 167     A_194_I    961 0.4805
#> 168     A_194_*     10 0.0050
#> 169     A_207_S   1029 0.5145
#> 170     A_207_G    961 0.4805
#> 171     A_207_*     10 0.0050
#> 172     A_219_R   1989 0.9945
#> 173     A_219_*     10 0.0050
#> 174     A_219_Q      1 0.0005
#> 175     A_231_V   1989 0.9945
#> 176     A_231_*     10 0.0050
#> 177     A_231_M      1 0.0005
#> 178     A_245_A   1911 0.9555
#> 179     A_245_V     79 0.0395
#> 180     A_245_*     10 0.0050
#> 181     A_246_S    389 0.1945
#> 182     A_246_A   1601 0.8005
#> 183     A_246_*     10 0.0050
#> 184     A_253_Q   1029 0.5145
#> 185     A_253_E    961 0.4805
#> 186     A_253_*     10 0.0050
#> 187     A_276_P   1317 0.6585
#> 188     A_276_L    665 0.3325
#> 189     A_276_*     18 0.0090
#> 190     A_282_I   1691 0.8455
#> 191     A_282_V    291 0.1455
#> 192     A_282_*     18 0.0090
#> 193     A_283_P   1934 0.9670
#> 194     A_283_H     48 0.0240
#> 195     A_283_*     18 0.0090
#> 196     A_284_I   1981 0.9905
#> 197     A_284_*     18 0.0090
#> 198     A_284_T      1 0.0005
#> 199     A_288_I   1981 0.9905
#> 200     A_288_*     18 0.0090
#> 201     A_288_L      1 0.0005
#> 202     A_294_F   1026 0.5130
#> 203     A_294_L    956 0.4780
#> 204     A_294_*     18 0.0090
#> 205     A_297_V   1917 0.9585
#> 206     A_297_M     65 0.0325
#> 207     A_297_*     18 0.0090
#> 208     A_298_I   1764 0.8820
#> 209     A_298_F    218 0.1090
#> 210     A_298_*     18 0.0090
#> 211     A_299_A    388 0.1940
#> 212     A_299_T   1594 0.7970
#> 213     A_299_*     18 0.0090
#> 214     A_307_M   1764 0.8820
#> 215     A_307_R    218 0.1090
#> 216     A_307_*     18 0.0090
#> 217     A_311_K   1691 0.8455
#> 218     A_311_N    291 0.1455
#> 219     A_311_*     18 0.0090
#> 220     A_321_S   1317 0.6585
#> 221     A_321_T    665 0.3325
#> 222     A_321_*     18 0.0090
#> 223     A_334_M    388 0.1940
#> 224     A_334_V   1594 0.7970
#> 225     A_334_*     18 0.0090
#> 226     B_-23_L    477 0.2385
#> 227     B_-23_R   1507 0.7535
#> 228     B_-23_*     16 0.0080
#> 229     B_-21_M    478 0.2390
#> 230     B_-21_T   1507 0.7535
#> 231     B_-21_*     15 0.0075
#> 232     B_-16_V   1272 0.6360
#> 233     B_-16_L    714 0.3570
#> 234     B_-16_*     14 0.0070
#> 235     B_-11_S    790 0.3950
#> 236     B_-11_W   1196 0.5980
#> 237     B_-11_*     14 0.0070
#> 238     B_-10_A    645 0.3225
#> 239     B_-10_G   1341 0.6705
#> 240     B_-10_*     14 0.0070
#> 241      B_-9_A   1985 0.9925
#> 242      B_-9_*     14 0.0070
#> 243      B_-9_V      1 0.0005
#> 244      B_-8_L    862 0.4310
#> 245      B_-8_V   1124 0.5620
#> 246      B_-8_*     14 0.0070
#> 247       B_9_Y   1414 0.7070
#> 248       B_9_H    455 0.2275
#> 249       B_9_D    131 0.0655
#> 250      B_11_S    590 0.2950
#> 251      B_11_A   1410 0.7050
#> 252      B_12_V    630 0.3150
#> 253      B_12_M   1370 0.6850
#> 254      B_24_S    655 0.3275
#> 255      B_24_T    659 0.3295
#> 256      B_24_A    686 0.3430
#> 257      B_30_D   1868 0.9340
#> 258      B_30_G    132 0.0660
#> 259      B_32_Q   1422 0.7110
#> 260      B_32_L    578 0.2890
#> 261      B_33_F   1999 0.9995
#> 262      B_33_S      1 0.0005
#> 263      B_41_A   1420 0.7100
#> 264      B_41_T    580 0.2900
#> 265      B_45_E    637 0.3185
#> 266      B_45_K    500 0.2500
#> 267      B_45_M    303 0.1515
#> 268      B_45_T    548 0.2740
#> 269      B_45_G     12 0.0060
#> 270      B_46_E   1697 0.8485
#> 271      B_46_A    303 0.1515
#> 272      B_52_I   1988 0.9940
#> 273      B_52_V     12 0.0060
#> 274      B_59_Y   1997 0.9985
#> 275      B_59_H      3 0.0015
#> 276      B_62_R   1879 0.9395
#> 277      B_62_G    121 0.0605
#> 278      B_63_N   1014 0.5070
#> 279      B_63_E    986 0.4930
#> 280      B_65_Q   1869 0.9345
#> 281      B_65_R    131 0.0655
#> 282      B_66_I   1858 0.9290
#> 283      B_66_K     11 0.0055
#> 284      B_66_N    131 0.0655
#> 285      B_67_Y    259 0.1295
#> 286      B_67_S    909 0.4545
#> 287      B_67_F    450 0.2250
#> 288      B_67_C    251 0.1255
#> 289      B_67_M    131 0.0655
#> 290      B_69_A    456 0.2280
#> 291      B_69_T   1533 0.7665
#> 292      B_69_R     11 0.0055
#> 293      B_70_Q    257 0.1285
#> 294      B_70_N   1533 0.7665
#> 295      B_70_K     79 0.0395
#> 296      B_70_S    131 0.0655
#> 297      B_71_A    467 0.2335
#> 298      B_71_T   1533 0.7665
#> 299      B_72_Q   1999 0.9995
#> 300      B_72_R      1 0.0005
#> 301      B_74_D    571 0.2855
#> 302      B_74_Y   1429 0.7145
#> 303      B_76_E   1987 0.9935
#> 304      B_76_V     13 0.0065
#> 305      B_77_S   1098 0.5490
#> 306      B_77_N    803 0.4015
#> 307      B_77_D     97 0.0485
#> 308      B_77_G      2 0.0010
#> 309      B_80_N   1097 0.5485
#> 310      B_80_T    464 0.2320
#> 311      B_80_I    439 0.2195
#> 312      B_81_L   1197 0.5985
#> 313      B_81_A    803 0.4015
#> 314      B_82_R   1097 0.5485
#> 315      B_82_L    903 0.4515
#> 316      B_83_G   1097 0.5485
#> 317      B_83_R    903 0.4515
#> 318      B_90_A   1998 0.9990
#> 319      B_90_D      2 0.0010
#> 320      B_94_T   1422 0.7110
#> 321      B_94_I    578 0.2890
#> 322      B_95_L    929 0.4645
#> 323      B_95_W    467 0.2335
#> 324      B_95_I    604 0.3020
#> 325      B_97_S    358 0.1790
#> 326      B_97_R   1035 0.5175
#> 327      B_97_T    368 0.1840
#> 328      B_97_N     77 0.0385
#> 329      B_97_V     94 0.0470
#> 330      B_97_W     68 0.0340
#> 331      B_99_Y   1966 0.9830
#> 332      B_99_S     27 0.0135
#> 333      B_99_F      7 0.0035
#> 334     B_101_C   1999 0.9995
#> 335     B_101_Y      1 0.0005
#> 336     B_103_V   1559 0.7795
#> 337     B_103_L    439 0.2195
#> 338     B_103_M      2 0.0010
#> 339     B_109_L   1986 0.9930
#> 340     B_109_F     14 0.0070
#> 341     B_113_H   1437 0.7185
#> 342     B_113_Y    563 0.2815
#> 343     B_114_D    972 0.4860
#> 344     B_114_N    936 0.4680
#> 345     B_114_H     92 0.0460
#> 346     B_116_Y    674 0.3370
#> 347     B_116_L    242 0.1210
#> 348     B_116_S    488 0.2440
#> 349     B_116_F    230 0.1150
#> 350     B_116_D    366 0.1830
#> 351     B_118_Y   1999 0.9995
#> 352     B_118_C      1 0.0005
#> 353     B_131_R    445 0.2225
#> 354     B_131_S   1555 0.7775
#> 355     B_137_D   1999 0.9995
#> 356     B_137_N      1 0.0005
#> 357     B_143_T   1927 0.9635
#> 358     B_143_S     73 0.0365
#> 359     B_145_R   1920 0.9600
#> 360     B_145_L     80 0.0400
#> 361     B_147_W   1927 0.9635
#> 362     B_147_L     73 0.0365
#> 363     B_152_E    675 0.3375
#> 364     B_152_V   1325 0.6625
#> 365     B_156_R    193 0.0965
#> 366     B_156_L   1304 0.6520
#> 367     B_156_W    111 0.0555
#> 368     B_156_D    392 0.1960
#> 369     B_158_A   1900 0.9500
#> 370     B_158_T    100 0.0500
#> 371     B_163_E    453 0.2265
#> 372     B_163_L   1016 0.5080
#> 373     B_163_T    531 0.2655
#> 374     B_166_E   1999 0.9995
#> 375     B_166_D      1 0.0005
#> 376     B_167_W   1704 0.8520
#> 377     B_167_S    295 0.1475
#> 378     B_167_G      1 0.0005
#> 379     B_171_Y   1614 0.8070
#> 380     B_171_H    386 0.1930
#> 381     B_177_D    397 0.1985
#> 382     B_177_E   1603 0.8015
#> 383     B_178_K    243 0.1215
#> 384     B_178_T   1757 0.8785
#> 385     B_180_E    397 0.1985
#> 386     B_180_Q   1603 0.8015
#> 387     B_194_I   1602 0.8010
#> 388     B_194_V    385 0.1925
#> 389     B_194_*     13 0.0065
#> 390     B_199_A   1720 0.8600
#> 391     B_199_V    267 0.1335
#> 392     B_199_*     13 0.0065
#> 393     B_211_A   1984 0.9920
#> 394     B_211_*     13 0.0065
#> 395     B_211_G      3 0.0015
#> 396     B_239_R   1985 0.9925
#> 397     B_239_*     13 0.0065
#> 398     B_239_G      2 0.0010
#> 399     B_245_A   1976 0.9880
#> 400     B_245_*     13 0.0065
#> 401     B_245_T     11 0.0055
#> 402     B_253_E   1985 0.9925
#> 403     B_253_*     13 0.0065
#> 404     B_253_Q      2 0.0010
#> 405     B_267_P   1985 0.9925
#> 406     B_267_*     13 0.0065
#> 407     B_267_Q      2 0.0010
#> 408     B_268_K   1985 0.9925
#> 409     B_268_*     13 0.0065
#> 410     B_268_E      2 0.0010
#> 411     B_270_L   1985 0.9925
#> 412     B_270_*     13 0.0065
#> 413     B_270_C      2 0.0010
#> 414     B_275_E   1982 0.9910
#> 415     B_275_*     16 0.0080
#> 416     B_275_K      2 0.0010
#> 417     B_282_V   1118 0.5590
#> 418     B_282_I    866 0.4330
#> 419     B_282_*     16 0.0080
#> 420     B_295_.   1982 0.9910
#> 421     B_295_*     16 0.0080
#> 422     B_295_V      2 0.0010
#> 423     B_296_A   1982 0.9910
#> 424     B_296_*     16 0.0080
#> 425     B_296_V      2 0.0010
#> 426     B_297_V   1982 0.9910
#> 427     B_297_*     16 0.0080
#> 428     B_297_T      2 0.0010
#> 429     B_299_V   1982 0.9910
#> 430     B_299_*     16 0.0080
#> 431     B_299_A      2 0.0010
#> 432     B_300_I   1982 0.9910
#> 433     B_300_*     16 0.0080
#> 434     B_300_V      2 0.0010
#> 435     B_301_G   1982 0.9910
#> 436     B_301_*     16 0.0080
#> 437     B_301_V      2 0.0010
#> 438     B_306_A   1125 0.5625
#> 439     B_306_T    859 0.4295
#> 440     B_306_*     16 0.0080
#> 441     B_308_M   1978 0.9890
#> 442     B_308_*     16 0.0080
#> 443     B_308_V      6 0.0030
#> 444     B_326_C    962 0.4810
#> 445     B_326_S   1022 0.5110
#> 446     B_326_*     16 0.0080
#> 447     B_339_A   1983 0.9915
#> 448     B_339_*     16 0.0080
#> 449     B_339_D      1 0.0005
#> 450     C_-18_R   1962 0.9810
#> 451     C_-18_*     17 0.0085
#> 452     C_-18_Q     21 0.0105
#> 453     C_-17_T   1448 0.7240
#> 454     C_-17_A    535 0.2675
#> 455     C_-17_*     17 0.0085
#> 456     C_-15_I   1288 0.6440
#> 457     C_-15_L    695 0.3475
#> 458     C_-15_*     17 0.0085
#> 459      C_-9_A   1471 0.7355
#> 460      C_-9_G    512 0.2560
#> 461      C_-9_*     17 0.0085
#> 462      C_-5_T   1962 0.9810
#> 463      C_-5_*     17 0.0085
#> 464      C_-5_I     21 0.0105
#> 465      C_-1_A   1973 0.9865
#> 466      C_-1_*     17 0.0085
#> 467      C_-1_T     10 0.0050
#> 468       C_1_C   1525 0.7625
#> 469       C_1_G    458 0.2290
#> 470       C_1_*     17 0.0085
#> 471       C_6_R   1895 0.9475
#> 472       C_6_K    105 0.0525
#> 473       C_9_Y    895 0.4475
#> 474       C_9_D    718 0.3590
#> 475       C_9_F    105 0.0525
#> 476       C_9_S    282 0.1410
#> 477      C_11_A   1613 0.8065
#> 478      C_11_S    387 0.1935
#> 479      C_14_R   1752 0.8760
#> 480      C_14_W    248 0.1240
#> 481      C_16_G   1899 0.9495
#> 482      C_16_S    101 0.0505
#> 483      C_21_R   1642 0.8210
#> 484      C_21_H    358 0.1790
#> 485      C_24_A   1177 0.5885
#> 486      C_24_S    823 0.4115
#> 487      C_35_R   1736 0.8680
#> 488      C_35_Q    264 0.1320
#> 489      C_49_A   1752 0.8760
#> 490      C_49_E    248 0.1240
#> 491      C_66_K   1682 0.8410
#> 492      C_66_N    318 0.1590
#> 493      C_68_K   1999 0.9995
#> 494      C_68_N      1 0.0005
#> 495      C_73_A   1178 0.5890
#> 496      C_73_T    822 0.4110
#> 497      C_77_S   1194 0.5970
#> 498      C_77_N    806 0.4030
#> 499      C_80_N   1194 0.5970
#> 500      C_80_K    806 0.4030
#> 501      C_90_A   1031 0.5155
#> 502      C_90_D    969 0.4845
#> 503      C_91_G   1913 0.9565
#> 504      C_91_R     87 0.0435
#> 505      C_94_T   1749 0.8745
#> 506      C_94_I    251 0.1255
#> 507      C_95_L   1707 0.8535
#> 508      C_95_I    259 0.1295
#> 509      C_95_F     34 0.0170
#> 510      C_97_W    548 0.2740
#> 511      C_97_R   1451 0.7255
#> 512      C_97_S      1 0.0005
#> 513      C_99_Y   1379 0.6895
#> 514      C_99_C    105 0.0525
#> 515      C_99_F    288 0.1440
#> 516      C_99_S    228 0.1140
#> 517     C_103_L   1808 0.9040
#> 518     C_103_V    192 0.0960
#> 519     C_113_Y   1938 0.9690
#> 520     C_113_H     62 0.0310
#> 521     C_114_D   1457 0.7285
#> 522     C_114_N    542 0.2710
#> 523     C_114_V      1 0.0005
#> 524     C_116_S   1082 0.5410
#> 525     C_116_F    588 0.2940
#> 526     C_116_Y    280 0.1400
#> 527     C_116_L     49 0.0245
#> 528     C_116_H      1 0.0005
#> 529     C_127_N   1999 0.9995
#> 530     C_127_K      1 0.0005
#> 531     C_138_T   1767 0.8835
#> 532     C_138_K    233 0.1165
#> 533     C_143_T   1979 0.9895
#> 534     C_143_S     21 0.0105
#> 535     C_147_W   1460 0.7300
#> 536     C_147_L    540 0.2700
#> 537     C_152_E   1381 0.6905
#> 538     C_152_A    588 0.2940
#> 539     C_152_T     31 0.0155
#> 540     C_156_W    486 0.2430
#> 541     C_156_R    625 0.3125
#> 542     C_156_L    790 0.3950
#> 543     C_156_D     34 0.0170
#> 544     C_156_Q     65 0.0325
#> 545     C_163_T   1690 0.8450
#> 546     C_163_L    191 0.0955
#> 547     C_163_E    119 0.0595
#> 548     C_169_R   1999 0.9995
#> 549     C_169_H      1 0.0005
#> 550     C_170_R   1979 0.9895
#> 551     C_170_G     21 0.0105
#> 552     C_173_E   1809 0.9045
#> 553     C_173_K    191 0.0955
#> 554     C_175_G   1991 0.9955
#> 555     C_175_R      9 0.0045
#> 556     C_177_E   1702 0.8510
#> 557     C_177_K    298 0.1490
#> 558     C_184_H   1454 0.7270
#> 559     C_184_P    511 0.2555
#> 560     C_184_*     14 0.0070
#> 561     C_184_R     21 0.0105
#> 562     C_186_K   1985 0.9925
#> 563     C_186_*     14 0.0070
#> 564     C_186_R      1 0.0005
#> 565     C_193_P   1917 0.9585
#> 566     C_193_L     69 0.0345
#> 567     C_193_*     14 0.0070
#> 568     C_194_V   1475 0.7375
#> 569     C_194_L    511 0.2555
#> 570     C_194_*     14 0.0070
#> 571     C_211_A   1896 0.9480
#> 572     C_211_T     90 0.0450
#> 573     C_211_*     14 0.0070
#> 574     C_219_R   1406 0.7030
#> 575     C_219_W    580 0.2900
#> 576     C_219_*     14 0.0070
#> 577     C_248_V   1882 0.9410
#> 578     C_248_M    104 0.0520
#> 579     C_248_*     14 0.0070
#> 580     C_253_E   1454 0.7270
#> 581     C_253_Q    532 0.2660
#> 582     C_253_*     14 0.0070
#> 583     C_261_V   1475 0.7375
#> 584     C_261_M    511 0.2555
#> 585     C_261_*     14 0.0070
#> 586     C_267_P   1454 0.7270
#> 587     C_267_Q    532 0.2660
#> 588     C_267_*     14 0.0070
#> 589     C_270_L   1965 0.9825
#> 590     C_270_*     14 0.0070
#> 591     C_270_C     21 0.0105
#> 592     C_273_R   1475 0.7375
#> 593     C_273_S    511 0.2555
#> 594     C_273_*     14 0.0070
#> 595     C_275_E   1444 0.7220
#> 596     C_275_K    275 0.1375
#> 597     C_275_G    264 0.1320
#> 598     C_275_*     17 0.0085
#> 599     C_284_I   1962 0.9810
#> 600     C_284_*     17 0.0085
#> 601     C_284_N     21 0.0105
#> 602     C_285_V   1451 0.7255
#> 603     C_285_M    511 0.2555
#> 604     C_285_*     17 0.0085
#> 605     C_285_L     21 0.0105
#> 606     C_289_A   1962 0.9810
#> 607     C_289_*     17 0.0085
#> 608     C_289_S     21 0.0105
#> 609     C_291_L   1962 0.9810
#> 610     C_291_*     17 0.0085
#> 611     C_291_P     21 0.0105
#> 612     C_295_A   1472 0.7360
#> 613     C_295_V    511 0.2555
#> 614     C_295_*     17 0.0085
#> 615     C_301_.   1979 0.9895
#> 616     C_301_A     21 0.0105
#> 617     C_302_.   1979 0.9895
#> 618     C_302_V     21 0.0105
#> 619     C_303_.   1979 0.9895
#> 620     C_303_L     21 0.0105
#> 621     C_304_.   1979 0.9895
#> 622     C_304_A     21 0.0105
#> 623     C_305_.   1979 0.9895
#> 624     C_305_V     21 0.0105
#> 625     C_306_.   1979 0.9895
#> 626     C_306_L     21 0.0105
#> 627     C_309_V   1735 0.8675
#> 628     C_309_M    248 0.1240
#> 629     C_309_*     17 0.0085
#> 630     C_310_M    706 0.3530
#> 631     C_310_V   1277 0.6385
#> 632     C_310_*     17 0.0085
#> 633     C_311_A   1472 0.7360
#> 634     C_311_T    511 0.2555
#> 635     C_311_*     17 0.0085
#> 636     C_312_V   1451 0.7255
#> 637     C_312_A    532 0.2660
#> 638     C_312_*     17 0.0085
#> 639     C_313_V   1472 0.7360
#> 640     C_313_M    511 0.2555
#> 641     C_313_*     17 0.0085
#> 642     C_314_M   1962 0.9810
#> 643     C_314_*     17 0.0085
#> 644     C_314_I     21 0.0105
#> 645     C_315_C   1962 0.9810
#> 646     C_315_*     17 0.0085
#> 647     C_315_H     21 0.0105
#> 648     C_327_C   1982 0.9910
#> 649     C_327_*     17 0.0085
#> 650     C_327_Y      1 0.0005
#> 651     C_330_A   1973 0.9865
#> 652     C_330_V     10 0.0050
#> 653     C_330_*     17 0.0085
#> 654     C_332_S   1472 0.7360
#> 655     C_332_C    510 0.2550
#> 656     C_332_*     18 0.0090
#> 657     C_345_A   1471 0.7355
#> 658     C_345_T    511 0.2555
#> 659     C_345_*     18 0.0090
#> 660   DPA1_11_A   1845 0.9225
#> 661   DPA1_11_M    155 0.0775
#> 662   DPA1_18_P   1999 0.9995
#> 663   DPA1_18_T      1 0.0005
#> 664   DPA1_28_E   1987 0.9935
#> 665   DPA1_28_D     13 0.0065
#> 666   DPA1_31_M   1545 0.7725
#> 667   DPA1_31_Q    455 0.2275
#> 668   DPA1_50_Q   1543 0.7715
#> 669   DPA1_50_R    457 0.2285
#> 670   DPA1_66_L   1992 0.9960
#> 671   DPA1_66_S      8 0.0040
#> 672   DPA1_72_T   1999 0.9995
#> 673   DPA1_72_I      1 0.0005
#> 674   DPA1_73_L   1999 0.9995
#> 675   DPA1_73_A      1 0.0005
#> 676   DPA1_83_T   1543 0.7715
#> 677   DPA1_83_A    457 0.2285
#> 678   DPA1_91_V   1985 0.9925
#> 679   DPA1_91_M     13 0.0065
#> 680   DPA1_91_*      2 0.0010
#> 681   DPA1_96_P   1997 0.9985
#> 682   DPA1_96_*      2 0.0010
#> 683   DPA1_96_A      1 0.0005
#> 684  DPA1_111_K   1543 0.7715
#> 685  DPA1_111_R    455 0.2275
#> 686  DPA1_111_*      2 0.0010
#> 687  DPA1_127_L   1542 0.7710
#> 688  DPA1_127_P    456 0.2280
#> 689  DPA1_127_*      2 0.0010
#> 690  DPA1_160_F   1542 0.7710
#> 691  DPA1_160_V    456 0.2280
#> 692  DPA1_160_*      2 0.0010
#> 693  DPA1_190_T   1988 0.9940
#> 694  DPA1_190_M      9 0.0045
#> 695  DPA1_190_*      2 0.0010
#> 696  DPA1_190_A      1 0.0005
#> 697  DPA1_228_T   1542 0.7710
#> 698  DPA1_228_P    456 0.2280
#> 699  DPA1_228_*      2 0.0010
#> 700  DPB1_-14_T   1961 0.9805
#> 701  DPB1_-14_*     38 0.0190
#> 702  DPB1_-14_M      1 0.0005
#> 703    DPB1_8_L   1431 0.7155
#> 704    DPB1_8_V    569 0.2845
#> 705    DPB1_9_F   1431 0.7155
#> 706    DPB1_9_Y    427 0.2135
#> 707    DPB1_9_H    142 0.0710
#> 708   DPB1_11_G   1526 0.7630
#> 709   DPB1_11_L    474 0.2370
#> 710   DPB1_17_A   1999 0.9995
#> 711   DPB1_17_P      1 0.0005
#> 712   DPB1_33_E   1947 0.9735
#> 713   DPB1_33_Q     53 0.0265
#> 714   DPB1_35_F   1659 0.8295
#> 715   DPB1_35_L    152 0.0760
#> 716   DPB1_35_Y    189 0.0945
#> 717   DPB1_36_V   1099 0.5495
#> 718   DPB1_36_A    901 0.4505
#> 719   DPB1_55_D    925 0.4625
#> 720   DPB1_55_E    160 0.0800
#> 721   DPB1_55_A    915 0.4575
#> 722   DPB1_56_E    925 0.4625
#> 723   DPB1_56_A   1075 0.5375
#> 724   DPB1_57_E   1652 0.8260
#> 725   DPB1_57_D    348 0.1740
#> 726   DPB1_65_I   1671 0.8355
#> 727   DPB1_65_L    328 0.1640
#> 728   DPB1_65_F      1 0.0005
#> 729   DPB1_69_E    541 0.2705
#> 730   DPB1_69_K   1406 0.7030
#> 731   DPB1_69_R     53 0.0265
#> 732   DPB1_76_M   1563 0.7815
#> 733   DPB1_76_V    376 0.1880
#> 734   DPB1_76_I     61 0.0305
#> 735   DPB1_84_G   1298 0.6490
#> 736   DPB1_84_D    687 0.3435
#> 737   DPB1_84_V     15 0.0075
#> 738   DPB1_85_G   1313 0.6565
#> 739   DPB1_85_E    687 0.3435
#> 740   DPB1_86_P   1313 0.6565
#> 741   DPB1_86_A    687 0.3435
#> 742   DPB1_87_M   1313 0.6565
#> 743   DPB1_87_V    687 0.3435
#> 744   DPB1_96_R   1341 0.6705
#> 745   DPB1_96_K    647 0.3235
#> 746   DPB1_96_*     12 0.0060
#> 747  DPB1_170_T   1341 0.6705
#> 748  DPB1_170_I    647 0.3235
#> 749  DPB1_170_*     12 0.0060
#> 750  DPB1_178_L   1773 0.8865
#> 751  DPB1_178_M    215 0.1075
#> 752  DPB1_178_*     12 0.0060
#> 753  DPB1_194_R   1909 0.9545
#> 754  DPB1_194_Q     77 0.0385
#> 755  DPB1_194_*     14 0.0070
#> 756  DPB1_205_V   1806 0.9030
#> 757  DPB1_205_M    180 0.0900
#> 758  DPB1_205_*     14 0.0070
#> 759  DPB1_215_I   1972 0.9860
#> 760  DPB1_215_T     14 0.0070
#> 761  DPB1_215_*     14 0.0070
#> 762  DQA1_-16_M   1138 0.5690
#> 763  DQA1_-16_L    861 0.4305
#> 764  DQA1_-16_*      1 0.0005
#> 765  DQA1_-13_T    312 0.1560
#> 766  DQA1_-13_A   1687 0.8435
#> 767  DQA1_-13_*      1 0.0005
#> 768   DQA1_-7_V   1922 0.9610
#> 769   DQA1_-7_M     77 0.0385
#> 770   DQA1_-7_*      1 0.0005
#> 771   DQA1_-6_M   1968 0.9840
#> 772   DQA1_-6_T     31 0.0155
#> 773   DQA1_-6_*      1 0.0005
#> 774    DQA1_1_E   1993 0.9965
#> 775    DQA1_1_K      6 0.0030
#> 776    DQA1_1_*      1 0.0005
#> 777    DQA1_2_D   1922 0.9610
#> 778    DQA1_2_G     77 0.0385
#> 779    DQA1_2_*      1 0.0005
#> 780   DQA1_11_Y   1220 0.6100
#> 781   DQA1_11_C    780 0.3900
#> 782   DQA1_18_S   1220 0.6100
#> 783   DQA1_18_F    780 0.3900
#> 784   DQA1_25_Y   1570 0.7850
#> 785   DQA1_25_F    430 0.2150
#> 786   DQA1_26_T   1674 0.8370
#> 787   DQA1_26_S    326 0.1630
#> 788   DQA1_34_Q   1149 0.5745
#> 789   DQA1_34_E    851 0.4255
#> 790   DQA1_40_G    638 0.3190
#> 791   DQA1_40_E   1362 0.6810
#> 792   DQA1_41_R   1848 0.9240
#> 793   DQA1_41_K    152 0.0760
#> 794   DQA1_44_T   1999 0.9995
#> 795   DQA1_44_A      1 0.0005
#> 796   DQA1_45_V   1220 0.6100
#> 797   DQA1_45_A    780 0.3900
#> 798   DQA1_47_C    638 0.3190
#> 799   DQA1_47_K    256 0.1280
#> 800   DQA1_47_R    780 0.3900
#> 801   DQA1_47_Q    326 0.1630
#> 802   DQA1_48_L   1220 0.6100
#> 803   DQA1_48_W    780 0.3900
#> 804   DQA1_50_V    638 0.3190
#> 805   DQA1_50_L    582 0.2910
#> 806   DQA1_50_E    780 0.3900
#> 807   DQA1_51_L    638 0.3190
#> 808   DQA1_51_F   1362 0.6810
#> 809   DQA1_52_R    964 0.4820
#> 810   DQA1_52_H    256 0.1280
#> 811   DQA1_52_S    780 0.3900
#> 812   DQA1_53_Q    638 0.3190
#> 813   DQA1_53_R    582 0.2910
#> 814   DQA1_53_K    780 0.3900
#> 815   DQA1_54_F   1744 0.8720
#> 816   DQA1_54_L    256 0.1280
#> 817   DQA1_55_R   1220 0.6100
#> 818   DQA1_55_G    780 0.3900
#> 819   DQA1_56_.    894 0.4470
#> 820   DQA1_56_G    780 0.3900
#> 821   DQA1_56_R    326 0.1630
#> 822   DQA1_61_F   1220 0.6100
#> 823   DQA1_61_G    780 0.3900
#> 824   DQA1_64_T   1220 0.6100
#> 825   DQA1_64_R    780 0.3900
#> 826   DQA1_66_I   1220 0.6100
#> 827   DQA1_66_M    780 0.3900
#> 828   DQA1_69_L   1138 0.5690
#> 829   DQA1_69_A    780 0.3900
#> 830   DQA1_69_T     82 0.0410
#> 831   DQA1_75_S    556 0.2780
#> 832   DQA1_75_I   1444 0.7220
#> 833   DQA1_76_L    894 0.4470
#> 834   DQA1_76_M    780 0.3900
#> 835   DQA1_76_V    326 0.1630
#> 836   DQA1_80_S   1220 0.6100
#> 837   DQA1_80_Y    780 0.3900
#> 838   DQA1_84_A   1998 0.9990
#> 839   DQA1_84_T      2 0.0010
#> 840  DQA1_102_L   1998 0.9990
#> 841  DQA1_102_V      1 0.0005
#> 842  DQA1_102_*      1 0.0005
#> 843  DQA1_107_I    556 0.2780
#> 844  DQA1_107_T   1443 0.7215
#> 845  DQA1_107_*      1 0.0005
#> 846  DQA1_129_H   1372 0.6860
#> 847  DQA1_129_Q    627 0.3135
#> 848  DQA1_129_*      1 0.0005
#> 849  DQA1_130_S   1847 0.9235
#> 850  DQA1_130_A    152 0.0760
#> 851  DQA1_130_*      1 0.0005
#> 852  DQA1_138_T   1997 0.9985
#> 853  DQA1_138_I      2 0.0010
#> 854  DQA1_138_*      1 0.0005
#> 855  DQA1_139_S   1998 0.9990
#> 856  DQA1_139_I      1 0.0005
#> 857  DQA1_139_*      1 0.0005
#> 858  DQA1_156_L    556 0.2780
#> 859  DQA1_156_F   1443 0.7215
#> 860  DQA1_156_*      1 0.0005
#> 861  DQA1_160_A   1814 0.9070
#> 862  DQA1_160_D    163 0.0815
#> 863  DQA1_160_S     22 0.0110
#> 864  DQA1_160_*      1 0.0005
#> 865  DQA1_161_E    556 0.2780
#> 866  DQA1_161_D   1443 0.7215
#> 867  DQA1_161_*      1 0.0005
#> 868  DQA1_163_S    556 0.2780
#> 869  DQA1_163_I   1443 0.7215
#> 870  DQA1_163_*      1 0.0005
#> 871  DQA1_175_K    556 0.2780
#> 872  DQA1_175_E    664 0.3320
#> 873  DQA1_175_Q    779 0.3895
#> 874  DQA1_175_*      1 0.0005
#> 875  DQA1_187_A   1673 0.8365
#> 876  DQA1_187_T    326 0.1630
#> 877  DQA1_187_*      1 0.0005
#> 878  DQA1_199_A   1947 0.9735
#> 879  DQA1_199_T     52 0.0260
#> 880  DQA1_199_*      1 0.0005
#> 881  DQA1_207_V   1641 0.8205
#> 882  DQA1_207_M    358 0.1790
#> 883  DQA1_207_*      1 0.0005
#> 884  DQA1_208_G   1998 0.9990
#> 885  DQA1_208_C      1 0.0005
#> 886  DQA1_208_*      1 0.0005
#> 887  DQA1_215_F   1417 0.7085
#> 888  DQA1_215_L    582 0.2910
#> 889  DQA1_215_*      1 0.0005
#> 890  DQA1_218_R   1220 0.6100
#> 891  DQA1_218_Q    779 0.3895
#> 892  DQA1_218_*      1 0.0005
#> 893  DQB1_-27_A   1647 0.8235
#> 894  DQB1_-27_S    326 0.1630
#> 895  DQB1_-27_*     27 0.0135
#> 896  DQB1_-21_D    743 0.3715
#> 897  DQB1_-21_G   1230 0.6150
#> 898  DQB1_-21_*     27 0.0135
#> 899  DQB1_-18_V   1084 0.5420
#> 900  DQB1_-18_A    889 0.4445
#> 901  DQB1_-18_*     27 0.0135
#> 902  DQB1_-17_A   1924 0.9620
#> 903  DQB1_-17_P     49 0.0245
#> 904  DQB1_-17_*     27 0.0135
#> 905  DQB1_-10_A   1554 0.7770
#> 906  DQB1_-10_S    419 0.2095
#> 907  DQB1_-10_*     27 0.0135
#> 908   DQB1_-9_M   1647 0.8235
#> 909   DQB1_-9_I    326 0.1630
#> 910   DQB1_-9_*     27 0.0135
#> 911   DQB1_-6_S    743 0.3715
#> 912   DQB1_-6_T   1230 0.6150
#> 913   DQB1_-6_*     27 0.0135
#> 914   DQB1_-5_L    417 0.2085
#> 915   DQB1_-5_P   1230 0.6150
#> 916   DQB1_-5_S    326 0.1630
#> 917   DQB1_-5_*     27 0.0135
#> 918   DQB1_-4_L    743 0.3715
#> 919   DQB1_-4_V   1230 0.6150
#> 920   DQB1_-4_*     27 0.0135
#> 921    DQB1_3_S   1924 0.9620
#> 922    DQB1_3_P     49 0.0245
#> 923    DQB1_3_*     27 0.0135
#> 924    DQB1_9_F    302 0.1510
#> 925    DQB1_9_Y   1649 0.8245
#> 926    DQB1_9_L     49 0.0245
#> 927   DQB1_13_G   1508 0.7540
#> 928   DQB1_13_A    492 0.2460
#> 929   DQB1_14_M   1673 0.8365
#> 930   DQB1_14_L    327 0.1635
#> 931   DQB1_23_R   1974 0.9870
#> 932   DQB1_23_L     26 0.0130
#> 933   DQB1_26_L   1091 0.5455
#> 934   DQB1_26_Y    492 0.2460
#> 935   DQB1_26_G    417 0.2085
#> 936   DQB1_28_T   1580 0.7900
#> 937   DQB1_28_S    420 0.2100
#> 938   DQB1_30_Y   1073 0.5365
#> 939   DQB1_30_S    420 0.2100
#> 940   DQB1_30_H    507 0.2535
#> 941   DQB1_37_Y   1531 0.7655
#> 942   DQB1_37_I    420 0.2100
#> 943   DQB1_37_D     49 0.0245
#> 944   DQB1_38_A   1204 0.6020
#> 945   DQB1_38_V    796 0.3980
#> 946   DQB1_45_G   1557 0.7785
#> 947   DQB1_45_E    443 0.2215
#> 948   DQB1_46_V   1580 0.7900
#> 949   DQB1_46_E    420 0.2100
#> 950   DQB1_47_Y   1580 0.7900
#> 951   DQB1_47_F    420 0.2100
#> 952   DQB1_52_P   1580 0.7900
#> 953   DQB1_52_L    420 0.2100
#> 954   DQB1_53_Q    793 0.3965
#> 955   DQB1_53_L   1207 0.6035
#> 956   DQB1_55_R    880 0.4400
#> 957   DQB1_55_L    420 0.2100
#> 958   DQB1_55_P    700 0.3500
#> 959   DQB1_56_P   1913 0.9565
#> 960   DQB1_56_L     87 0.0435
#> 961   DQB1_57_D   1032 0.5160
#> 962   DQB1_57_A    601 0.3005
#> 963   DQB1_57_V    295 0.1475
#> 964   DQB1_57_S     72 0.0360
#> 965   DQB1_66_E   1440 0.7200
#> 966   DQB1_66_D    560 0.2800
#> 967   DQB1_67_V   1440 0.7200
#> 968   DQB1_67_I    560 0.2800
#> 969   DQB1_70_G    651 0.3255
#> 970   DQB1_70_R   1258 0.6290
#> 971   DQB1_70_E     91 0.0455
#> 972   DQB1_71_T   1166 0.5830
#> 973   DQB1_71_K    420 0.2100
#> 974   DQB1_71_D     91 0.0455
#> 975   DQB1_71_A    323 0.1615
#> 976   DQB1_74_E   1166 0.5830
#> 977   DQB1_74_A    420 0.2100
#> 978   DQB1_74_S    414 0.2070
#> 979   DQB1_75_L   1166 0.5830
#> 980   DQB1_75_V    834 0.4170
#> 981   DQB1_77_T   1253 0.6265
#> 982   DQB1_77_R    747 0.3735
#> 983   DQB1_84_E    793 0.3965
#> 984   DQB1_84_Q   1207 0.6035
#> 985   DQB1_85_V    793 0.3965
#> 986   DQB1_85_L   1207 0.6035
#> 987   DQB1_86_A    704 0.3520
#> 988   DQB1_86_E   1207 0.6035
#> 989   DQB1_86_G     89 0.0445
#> 990   DQB1_87_F    377 0.1885
#> 991   DQB1_87_L   1207 0.6035
#> 992   DQB1_87_Y    416 0.2080
#> 993   DQB1_89_G    793 0.3965
#> 994   DQB1_89_T   1207 0.6035
#> 995   DQB1_90_I    793 0.3965
#> 996   DQB1_90_T   1207 0.6035
#> 997  DQB1_116_V   1673 0.8365
#> 998  DQB1_116_I    327 0.1635
#> 999  DQB1_125_G    466 0.2330
#> 1000 DQB1_125_A   1207 0.6035
#> 1001 DQB1_125_S    327 0.1635
#> 1002 DQB1_126_Q   1928 0.9640
#> 1003 DQB1_126_H     72 0.0360
#> 1004 DQB1_130_R   1879 0.9395
#> 1005 DQB1_130_Q    121 0.0605
#> 1006 DQB1_135_D   1798 0.8990
#> 1007 DQB1_135_G    202 0.1010
#> 1008 DQB1_140_A   1213 0.6065
#> 1009 DQB1_140_T    787 0.3935
#> 1010 DQB1_142_V   1999 0.9995
#> 1011 DQB1_142_I      1 0.0005
#> 1012 DQB1_167_R   1508 0.7540
#> 1013 DQB1_167_H    492 0.2460
#> 1014 DQB1_177_H   1999 0.9995
#> 1015 DQB1_177_Q      1 0.0005
#> 1016 DQB1_182_S   1213 0.6065
#> 1017 DQB1_182_N    787 0.3935
#> 1018 DQB1_185_T   1645 0.8225
#> 1019 DQB1_185_I    355 0.1775
#> 1020 DQB1_197_S   1924 0.9620
#> 1021 DQB1_197_N     49 0.0245
#> 1022 DQB1_197_*     27 0.0135
#> 1023 DQB1_203_V    743 0.3715
#> 1024 DQB1_203_I   1230 0.6150
#> 1025 DQB1_203_*     27 0.0135
#> 1026 DQB1_220_R    792 0.3960
#> 1027 DQB1_220_H   1181 0.5905
#> 1028 DQB1_220_*     27 0.0135
#> 1029 DQB1_221_Q    792 0.3960
#> 1030 DQB1_221_H   1181 0.5905
#> 1031 DQB1_221_*     27 0.0135
#> 1032 DQB1_224_Q   1647 0.8235
#> 1033 DQB1_224_R    326 0.1630
#> 1034 DQB1_224_*     27 0.0135
#> 1035 DQB1_227_.   1876 0.9380
#> 1036 DQB1_227_P     97 0.0485
#> 1037 DQB1_227_*     27 0.0135
#> 1038 DQB1_228_.   1876 0.9380
#> 1039 DQB1_228_Q     97 0.0485
#> 1040 DQB1_228_*     27 0.0135
#> 1041 DQB1_229_.   1876 0.9380
#> 1042 DQB1_229_G     97 0.0485
#> 1043 DQB1_229_*     27 0.0135
#> 1044 DQB1_230_.   1876 0.9380
#> 1045 DQB1_230_P     97 0.0485
#> 1046 DQB1_230_*     27 0.0135
#> 1047 DQB1_231_.   1876 0.9380
#> 1048 DQB1_231_P     97 0.0485
#> 1049 DQB1_231_*     27 0.0135
#> 1050 DQB1_232_.   1876 0.9380
#> 1051 DQB1_232_P     97 0.0485
#> 1052 DQB1_232_*     27 0.0135
#> 1053 DQB1_233_.   1876 0.9380
#> 1054 DQB1_233_A     97 0.0485
#> 1055 DQB1_233_*     27 0.0135
#> 1056 DQB1_234_.   1876 0.9380
#> 1057 DQB1_234_G     97 0.0485
#> 1058 DQB1_234_*     27 0.0135
#> 1059  DRA_217_L    711 0.3555
#> 1060  DRA_217_V   1289 0.6445
#> 1061 DRB1_-25_K   1058 0.5290
#> 1062 DRB1_-25_R    926 0.4630
#> 1063 DRB1_-25_*     16 0.0080
#> 1064 DRB1_-24_L   1697 0.8485
#> 1065 DRB1_-24_F    287 0.1435
#> 1066 DRB1_-24_*     16 0.0080
#> 1067 DRB1_-17_T    490 0.2450
#> 1068 DRB1_-17_A   1494 0.7470
#> 1069 DRB1_-17_*     16 0.0080
#> 1070 DRB1_-16_A   1058 0.5290
#> 1071 DRB1_-16_V    924 0.4620
#> 1072 DRB1_-16_I      2 0.0010
#> 1073 DRB1_-16_*     16 0.0080
#> 1074  DRB1_-1_S    269 0.1345
#> 1075  DRB1_-1_A   1715 0.8575
#> 1076  DRB1_-1_*     16 0.0080
#> 1077   DRB1_4_R   1703 0.8515
#> 1078   DRB1_4_Q    281 0.1405
#> 1079   DRB1_4_*     16 0.0080
#> 1080   DRB1_9_W    751 0.3755
#> 1081   DRB1_9_E   1217 0.6085
#> 1082   DRB1_9_K     32 0.0160
#> 1083  DRB1_10_Q   1072 0.5360
#> 1084  DRB1_10_Y    906 0.4530
#> 1085  DRB1_10_E     22 0.0110
#> 1086  DRB1_11_P    325 0.1625
#> 1087  DRB1_11_G    249 0.1245
#> 1088  DRB1_11_S    906 0.4530
#> 1089  DRB1_11_D     32 0.0160
#> 1090  DRB1_11_V    311 0.1555
#> 1091  DRB1_11_L    177 0.0885
#> 1092  DRB1_12_K   1094 0.5470
#> 1093  DRB1_12_T    906 0.4530
#> 1094  DRB1_13_R    325 0.1625
#> 1095  DRB1_13_Y    249 0.1245
#> 1096  DRB1_13_S    760 0.3800
#> 1097  DRB1_13_F    231 0.1155
#> 1098  DRB1_13_H    289 0.1445
#> 1099  DRB1_13_G    146 0.0730
#> 1100  DRB1_14_E   1751 0.8755
#> 1101  DRB1_14_K    249 0.1245
#> 1102  DRB1_16_H   1852 0.9260
#> 1103  DRB1_16_Y    146 0.0730
#> 1104  DRB1_16_Q      2 0.0010
#> 1105  DRB1_25_R   1751 0.8755
#> 1106  DRB1_25_Q    249 0.1245
#> 1107  DRB1_26_F   1495 0.7475
#> 1108  DRB1_26_Y    248 0.1240
#> 1109  DRB1_26_L    257 0.1285
#> 1110  DRB1_28_D   1446 0.7230
#> 1111  DRB1_28_E    522 0.2610
#> 1112  DRB1_28_H     32 0.0160
#> 1113  DRB1_30_Y   1458 0.7290
#> 1114  DRB1_30_L    249 0.1245
#> 1115  DRB1_30_G     32 0.0160
#> 1116  DRB1_30_R     22 0.0110
#> 1117  DRB1_30_C    177 0.0885
#> 1118  DRB1_30_H     62 0.0310
#> 1119  DRB1_31_F   1769 0.8845
#> 1120  DRB1_31_I    209 0.1045
#> 1121  DRB1_31_V     22 0.0110
#> 1122  DRB1_32_Y   1436 0.7180
#> 1123  DRB1_32_H    564 0.2820
#> 1124  DRB1_33_N   1711 0.8555
#> 1125  DRB1_33_H    289 0.1445
#> 1126  DRB1_37_S    510 0.2550
#> 1127  DRB1_37_F    302 0.1510
#> 1128  DRB1_37_N    463 0.2315
#> 1129  DRB1_37_Y    667 0.3335
#> 1130  DRB1_37_L     58 0.0290
#> 1131  DRB1_38_V   1920 0.9600
#> 1132  DRB1_38_A     22 0.0110
#> 1133  DRB1_38_L     58 0.0290
#> 1134  DRB1_40_F   1978 0.9890
#> 1135  DRB1_40_Y     22 0.0110
#> 1136  DRB1_47_F    993 0.4965
#> 1137  DRB1_47_Y   1007 0.5035
#> 1138  DRB1_57_D   1469 0.7345
#> 1139  DRB1_57_V    339 0.1695
#> 1140  DRB1_57_A     52 0.0260
#> 1141  DRB1_57_S    140 0.0700
#> 1142  DRB1_58_A   1749 0.8745
#> 1143  DRB1_58_E    251 0.1255
#> 1144  DRB1_60_Y   1610 0.8050
#> 1145  DRB1_60_S    339 0.1695
#> 1146  DRB1_60_H     51 0.0255
#> 1147  DRB1_67_I    839 0.4195
#> 1148  DRB1_67_L    749 0.3745
#> 1149  DRB1_67_F    412 0.2060
#> 1150  DRB1_70_Q    935 0.4675
#> 1151  DRB1_70_D    957 0.4785
#> 1152  DRB1_70_R    108 0.0540
#> 1153  DRB1_71_A    269 0.1345
#> 1154  DRB1_71_R   1137 0.5685
#> 1155  DRB1_71_E    245 0.1225
#> 1156  DRB1_71_K    349 0.1745
#> 1157  DRB1_73_A   1531 0.7655
#> 1158  DRB1_73_G    469 0.2345
#> 1159  DRB1_74_A   1299 0.6495
#> 1160  DRB1_74_Q    249 0.1245
#> 1161  DRB1_74_E    147 0.0735
#> 1162  DRB1_74_R    220 0.1100
#> 1163  DRB1_74_L     85 0.0425
#> 1164  DRB1_77_T   1780 0.8900
#> 1165  DRB1_77_N    220 0.1100
#> 1166  DRB1_78_Y   1718 0.8590
#> 1167  DRB1_78_V    282 0.1410
#> 1168  DRB1_82_N   1999 0.9995
#> 1169  DRB1_82_S      1 0.0005
#> 1170  DRB1_85_V   1904 0.9520
#> 1171  DRB1_85_A     96 0.0480
#> 1172  DRB1_86_V    958 0.4790
#> 1173  DRB1_86_G   1042 0.5210
#> 1174  DRB1_96_Q    335 0.1675
#> 1175  DRB1_96_H   1186 0.5930
#> 1176  DRB1_96_Y    288 0.1440
#> 1177  DRB1_96_E    177 0.0885
#> 1178  DRB1_96_*     14 0.0070
#> 1179  DRB1_98_K   1417 0.7085
#> 1180  DRB1_98_E    569 0.2845
#> 1181  DRB1_98_*     14 0.0070
#> 1182 DRB1_104_S   1417 0.7085
#> 1183 DRB1_104_A    569 0.2845
#> 1184 DRB1_104_*     14 0.0070
#> 1185 DRB1_112_H   1979 0.9895
#> 1186 DRB1_112_Y      7 0.0035
#> 1187 DRB1_112_*     14 0.0070
#> 1188 DRB1_120_S   1676 0.8380
#> 1189 DRB1_120_N    310 0.1550
#> 1190 DRB1_120_*     14 0.0070
#> 1191 DRB1_133_L    313 0.1565
#> 1192 DRB1_133_R   1673 0.8365
#> 1193 DRB1_133_*     14 0.0070
#> 1194 DRB1_140_A    771 0.3855
#> 1195 DRB1_140_T   1215 0.6075
#> 1196 DRB1_140_*     14 0.0070
#> 1197 DRB1_142_M    313 0.1565
#> 1198 DRB1_142_V   1673 0.8365
#> 1199 DRB1_142_*     14 0.0070
#> 1200 DRB1_149_Q   1081 0.5405
#> 1201 DRB1_149_H    905 0.4525
#> 1202 DRB1_149_*     14 0.0070
#> 1203 DRB1_166_R   1964 0.9820
#> 1204 DRB1_166_Q     22 0.0110
#> 1205 DRB1_166_*     14 0.0070
#> 1206 DRB1_180_V   1698 0.8490
#> 1207 DRB1_180_L    288 0.1440
#> 1208 DRB1_180_*     14 0.0070
#> 1209 DRB1_181_T   1683 0.8415
#> 1210 DRB1_181_M    303 0.1515
#> 1211 DRB1_181_*     14 0.0070
#> 1212 DRB1_189_R   1908 0.9540
#> 1213 DRB1_189_S     77 0.0385
#> 1214 DRB1_189_*     15 0.0075
#> 1215 DRB1_207_V   1984 0.9920
#> 1216 DRB1_207_*     15 0.0075
#> 1217 DRB1_207_M      1 0.0005
#> 1218 DRB1_231_Q   1962 0.9810
#> 1219 DRB1_231_P     22 0.0110
#> 1220 DRB1_231_*     16 0.0080
#> 1221 DRB1_233_T   1157 0.5785
#> 1222 DRB1_233_R    827 0.4135
#> 1223 DRB1_233_*     16 0.0080