Calculate HLA allele frequencies

getHlaFrequencies calculates allele frequencies in HLA calls data frame.

getHlaFrequencies(
  hla_calls,
  carrier_frequency = FALSE,
  compare = FALSE,
  ref_pop = c("USA NMDP African American pop 2", "USA NMDP Chinese",
    "USA NMDP European Caucasian", "USA NMDP Hispanic South or Central American",
    "USA NMDP Japanese", "USA NMDP North American Amerindian",
    "USA NMDP South Asian Indian"),
  ref = allele_frequencies
)

Arguments

hla_calls
HLA calls data frame, as returned by readHlaCalls function.

carrier_frequency
Logical flag indicating if carrier frequency should be returned.

compare
Logical flag indicating if hla_calls frequencies should be compared to reference frequencies given in ref.

ref_pop
Character vector giving names of reference populations in ref to compare with. Optionally vector can be named, then those names will be used as population names.

ref
Data frame giving reference allele frequencies. See allele_frequencies for an example.

Value

Data frame with each row holding HLA allele, it's count and frequency.

Details

Both gene copies are taken into consideration for frequencies calculation, frequency = n / (2 * j) where n is the number of allele occurrences and j is the number of samples in hla_calls.

Examples

getHlaFrequencies(MiDAS_tut_HLA)

#>          allele Counts   Freq
#> 1       A*01:01    236 0.1180
#> 2       A*01:02      2 0.0010
#> 3      A*01:234      1 0.0005
#> 4      A*01:237      1 0.0005
#> 5      A*01:274      4 0.0020
#> 6       A*01:43      1 0.0005
#> 7       A*01:98      1 0.0005
#> 8       A*02:01    486 0.2430
#> 9       A*02:02      2 0.0010
#> 10      A*02:03      7 0.0035
#> 11      A*02:05     17 0.0085
#> 12      A*02:06     22 0.0110
#> 13      A*02:07      7 0.0035
#> 14      A*02:08      2 0.0010
#> 15      A*02:10      3 0.0015
#> 16     A*02:211      1 0.0005
#> 17      A*02:22      1 0.0005
#> 18     A*02:324      1 0.0005
#> 19      A*02:60      1 0.0005
#> 20     A*02:614      1 0.0005
#> 21     A*02:642      1 0.0005
#> 22     A*02:686      1 0.0005
#> 23     A*02:753      1 0.0005
#> 24     A*02:763      1 0.0005
#> 25      A*03:01    199 0.0995
#> 26      A*03:02     13 0.0065
#> 27     A*03:205      1 0.0005
#> 28      A*03:24      1 0.0005
#> 29     A*03:312      3 0.0015
#> 30      A*03:72      2 0.0010
#> 31      A*11:01    114 0.0570
#> 32      A*11:02      1 0.0005
#> 33      A*11:03      1 0.0005
#> 34      A*11:12      1 0.0005
#> 35     A*11:306      3 0.0015
#> 36      A*11:43      1 0.0005
#> 37     A*11:50Q      1 0.0005
#> 38      A*11:88      3 0.0015
#> 39      A*23:01     47 0.0235
#> 40      A*23:17      1 0.0005
#> 41      A*23:18      1 0.0005
#> 42      A*24:02    228 0.1140
#> 43      A*24:03      2 0.0010
#> 44      A*24:04      1 0.0005
#> 45      A*24:07      2 0.0010
#> 46     A*24:104      2 0.0010
#> 47      A*24:20      4 0.0020
#> 48      A*24:50      4 0.0020
#> 49      A*24:95      1 0.0005
#> 50      A*25:01     52 0.0260
#> 51      A*25:09      2 0.0010
#> 52      A*25:25      1 0.0005
#> 53      A*26:01     90 0.0450
#> 54      A*26:02      4 0.0020
#> 55      A*26:03      2 0.0010
#> 56      A*26:08      3 0.0015
#> 57      A*26:16      1 0.0005
#> 58      A*29:01      6 0.0030
#> 59      A*29:02     45 0.0225
#> 60     A*29:119      3 0.0015
#> 61      A*30:01     37 0.0185
#> 62      A*30:02     44 0.0220
#> 63      A*30:04      3 0.0015
#> 64      A*31:01     58 0.0290
#> 65      A*31:08      1 0.0005
#> 66      A*31:21      1 0.0005
#> 67      A*32:01     58 0.0290
#> 68      A*32:02      1 0.0005
#> 69     A*32:116      4 0.0020
#> 70      A*33:01     20 0.0100
#> 71      A*33:03     16 0.0080
#> 72      A*33:05      2 0.0010
#> 73     A*33:125      1 0.0005
#> 74     A*33:163      1 0.0005
#> 75      A*33:51      1 0.0005
#> 76      A*34:01      1 0.0005
#> 77      A*34:02      6 0.0030
#> 78      A*66:01     10 0.0050
#> 79      A*68:01     53 0.0265
#> 80      A*68:02     21 0.0105
#> 81      A*68:04      1 0.0005
#> 82      A*68:16      1 0.0005
#> 83     A*68:164      1 0.0005
#> 84      A*68:17      1 0.0005
#> 85      A*68:23      1 0.0005
#> 86      A*68:66      1 0.0005
#> 87      A*69:01      6 0.0030
#> 88      A*74:01      2 0.0010
#> 89      B*07:02    152 0.0760
#> 90      B*07:04      1 0.0005
#> 91      B*07:05      5 0.0025
#> 92      B*07:06      5 0.0025
#> 93      B*07:10      1 0.0005
#> 94      B*07:14      1 0.0005
#> 95      B*07:22      1 0.0005
#> 96      B*07:47      1 0.0005
#> 97      B*08:01    128 0.0640
#> 98     B*08:156      2 0.0010
#> 99      B*08:94      1 0.0005
#> 100     B*13:01      5 0.0025
#> 101     B*13:02     74 0.0370
#> 102    B*13:117      1 0.0005
#> 103     B*14:01     22 0.0110
#> 104     B*14:02     40 0.0200
#> 105     B*14:28      1 0.0005
#> 106     B*15:01     86 0.0430
#> 107     B*15:02      6 0.0030
#> 108     B*15:03      8 0.0040
#> 109     B*15:07      2 0.0010
#> 110     B*15:09      2 0.0010
#> 111     B*15:10      4 0.0020
#> 112     B*15:11      3 0.0015
#> 113     B*15:12      1 0.0005
#> 114     B*15:15      1 0.0005
#> 115     B*15:17     10 0.0050
#> 116     B*15:18      3 0.0015
#> 117     B*15:20      1 0.0005
#> 118    B*15:212      1 0.0005
#> 119    B*15:220      1 0.0005
#> 120     B*15:25      3 0.0015
#> 121    B*15:256      1 0.0005
#> 122     B*15:27      1 0.0005
#> 123     B*15:39      1 0.0005
#> 124     B*15:63      1 0.0005
#> 125     B*18:01    129 0.0645
#> 126     B*18:02      1 0.0005
#> 127     B*18:03      1 0.0005
#> 128     B*18:07      1 0.0005
#> 129     B*27:02      9 0.0045
#> 130     B*27:03      3 0.0015
#> 131     B*27:04      2 0.0010
#> 132     B*27:05     59 0.0295
#> 133     B*27:06      1 0.0005
#> 134    B*27:127      2 0.0010
#> 135     B*27:14      1 0.0005
#> 136     B*35:01     84 0.0420
#> 137     B*35:02     14 0.0070
#> 138     B*35:03     39 0.0195
#> 139     B*35:05      2 0.0010
#> 140     B*35:08     15 0.0075
#> 141     B*35:09      1 0.0005
#> 142     B*35:17      1 0.0005
#> 143    B*35:194      1 0.0005
#> 144     B*35:41      1 0.0005
#> 145     B*35:42      1 0.0005
#> 146     B*35:43      1 0.0005
#> 147     B*35:60      1 0.0005
#> 148     B*35:67      1 0.0005
#> 149     B*35:71      1 0.0005
#> 150     B*37:01     25 0.0125
#> 151     B*37:68      1 0.0005
#> 152     B*38:01     53 0.0265
#> 153     B*38:02      5 0.0025
#> 154     B*39:01     29 0.0145
#> 155     B*39:05      1 0.0005
#> 156     B*39:06     10 0.0050
#> 157     B*39:09      1 0.0005
#> 158     B*40:01     58 0.0290
#> 159     B*40:02     28 0.0140
#> 160     B*40:06     17 0.0085
#> 161    B*40:117      1 0.0005
#> 162     B*40:21      1 0.0005
#> 163     B*40:32      1 0.0005
#> 164    B*40:353      1 0.0005
#> 165     B*40:49      1 0.0005
#> 166     B*41:01      9 0.0045
#> 167     B*41:02      9 0.0045
#> 168     B*42:01      5 0.0025
#> 169     B*42:02      2 0.0010
#> 170     B*44:02    169 0.0845
#> 171     B*44:03     87 0.0435
#> 172     B*44:04      1 0.0005
#> 173     B*44:05      5 0.0025
#> 174     B*44:06      1 0.0005
#> 175     B*44:27     10 0.0050
#> 176    B*44:270      2 0.0010
#> 177     B*44:37      1 0.0005
#> 178     B*44:55      1 0.0005
#> 179     B*45:01     15 0.0075
#> 180     B*46:01     11 0.0055
#> 181     B*47:01      6 0.0030
#> 182     B*47:04      2 0.0010
#> 183     B*48:01     11 0.0055
#> 184     B*49:01     43 0.0215
#> 185     B*50:01     31 0.0155
#> 186     B*50:02      1 0.0005
#> 187     B*51:01    132 0.0660
#> 188     B*51:02      1 0.0005
#> 189     B*51:05      1 0.0005
#> 190     B*51:07      4 0.0020
#> 191     B*51:08      6 0.0030
#> 192     B*51:09      1 0.0005
#> 193   B*51:173Q      1 0.0005
#> 194    B*51:203      1 0.0005
#> 195    B*51:230      1 0.0005
#> 196     B*52:01     43 0.0215
#> 197     B*53:01      8 0.0040
#> 198     B*54:01     12 0.0060
#> 199     B*55:01     32 0.0160
#> 200     B*55:02      8 0.0040
#> 201     B*56:01     17 0.0085
#> 202     B*56:03      1 0.0005
#> 203     B*56:09      1 0.0005
#> 204     B*57:01     81 0.0405
#> 205     B*57:02      7 0.0035
#> 206     B*57:03      6 0.0030
#> 207     B*58:01     22 0.0110
#> 208     B*58:02      5 0.0025
#> 209     B*59:01      2 0.0010
#> 210     B*67:01      1 0.0005
#> 211     B*73:01      2 0.0010
#> 212     C*01:02    100 0.0500
#> 213     C*01:03      2 0.0010
#> 214    C*01:103      1 0.0005
#> 215     C*01:22      1 0.0005
#> 216     C*01:51      1 0.0005
#> 217     C*02:02     88 0.0440
#> 218     C*02:06      1 0.0005
#> 219     C*02:10      8 0.0040
#> 220     C*02:20      1 0.0005
#> 221     C*03:02     13 0.0065
#> 222     C*03:03     82 0.0410
#> 223     C*03:04     86 0.0430
#> 224     C*03:05      1 0.0005
#> 225     C*03:06      1 0.0005
#> 226    C*03:344      4 0.0020
#> 227     C*03:39      3 0.0015
#> 228     C*03:85      1 0.0005
#> 229     C*04:01    247 0.1235
#> 230     C*04:03      3 0.0015
#> 231    C*04:277      1 0.0005
#> 232     C*05:01    159 0.0795
#> 233     C*05:05      1 0.0005
#> 234     C*05:37      1 0.0005
#> 235     C*06:02    195 0.0975
#> 236     C*06:09      1 0.0005
#> 237     C*07:01    243 0.1215
#> 238     C*07:02    219 0.1095
#> 239     C*07:04     34 0.0170
#> 240     C*07:18     10 0.0050
#> 241    C*07:204      1 0.0005
#> 242     C*07:51      2 0.0010
#> 243    C*07:578      7 0.0035
#> 244    C*07:615      1 0.0005
#> 245    C*07:696      2 0.0010
#> 246     C*08:01     22 0.0110
#> 247     C*08:02     72 0.0360
#> 248     C*08:03      9 0.0045
#> 249     C*12:02     43 0.0215
#> 250     C*12:03    142 0.0710
#> 251    C*12:132      1 0.0005
#> 252    C*12:167      1 0.0005
#> 253     C*12:28      1 0.0005
#> 254     C*14:02     29 0.0145
#> 255     C*14:03      5 0.0025
#> 256     C*15:02     47 0.0235
#> 257     C*15:05     12 0.0060
#> 258     C*15:07      1 0.0005
#> 259     C*15:09      1 0.0005
#> 260     C*16:01     49 0.0245
#> 261     C*16:02     16 0.0080
#> 262     C*16:04      4 0.0020
#> 263     C*17:01     11 0.0055
#> 264     C*17:03     10 0.0050
#> 265     C*18:02      3 0.0015
#> 266  DPA1*01:03   1522 0.7610
#> 267  DPA1*01:04     12 0.0060
#> 268  DPA1*01:05      1 0.0005
#> 269  DPA1*01:11      1 0.0005
#> 270  DPA1*02:01    308 0.1540
#> 271  DPA1*02:02    125 0.0625
#> 272  DPA1*02:06      9 0.0045
#> 273  DPA1*02:07     13 0.0065
#> 274  DPA1*03:01      8 0.0040
#> 275  DPA1*04:01      1 0.0005
#> 276  DPB1*01:01     74 0.0370
#> 277  DPB1*02:01    304 0.1520
#> 278  DPB1*02:02     30 0.0150
#> 279  DPB1*03:01    137 0.0685
#> 280  DPB1*04:01    685 0.3425
#> 281  DPB1*04:02    202 0.1010
#> 282  DPB1*05:01    107 0.0535
#> 283  DPB1*06:01     19 0.0095
#> 284  DPB1*09:01     20 0.0100
#> 285  DPB1*10:01     25 0.0125
#> 286 DPB1*104:01     48 0.0240
#> 287 DPB1*105:01     14 0.0070
#> 288 DPB1*107:01      1 0.0005
#> 289  DPB1*11:01     40 0.0200
#> 290 DPB1*113:01      2 0.0010
#> 291 DPB1*124:01     12 0.0060
#> 292 DPB1*126:01      7 0.0035
#> 293  DPB1*13:01     51 0.0255
#> 294 DPB1*131:01      2 0.0010
#> 295 DPB1*133:01      1 0.0005
#> 296 DPB1*135:01      6 0.0030
#> 297 DPB1*138:01      1 0.0005
#> 298  DPB1*14:01     39 0.0195
#> 299  DPB1*15:01     13 0.0065
#> 300 DPB1*150:01      1 0.0005
#> 301 DPB1*158:01      1 0.0005
#> 302  DPB1*16:01     10 0.0050
#> 303 DPB1*162:01      5 0.0025
#> 304  DPB1*17:01     49 0.0245
#> 305  DPB1*18:01      1 0.0005
#> 306  DPB1*19:01      7 0.0035
#> 307  DPB1*20:01      7 0.0035
#> 308  DPB1*21:01      4 0.0020
#> 309  DPB1*22:01      1 0.0005
#> 310  DPB1*23:01     15 0.0075
#> 311 DPB1*234:01      1 0.0005
#> 312 DPB1*259:01      1 0.0005
#> 313 DPB1*271:01      4 0.0020
#> 314 DPB1*294:01      1 0.0005
#> 315  DPB1*33:01      1 0.0005
#> 316 DPB1*350:01      7 0.0035
#> 317 DPB1*351:01      3 0.0015
#> 318  DPB1*36:01      1 0.0005
#> 319  DPB1*38:01      1 0.0005
#> 320 DPB1*398:01      1 0.0005
#> 321  DPB1*40:01      1 0.0005
#> 322  DPB1*41:01      1 0.0005
#> 323 DPB1*414:01      7 0.0035
#> 324  DPB1*45:01      1 0.0005
#> 325 DPB1*461:01      1 0.0005
#> 326 DPB1*462:01      1 0.0005
#> 327 DPB1*463:01      4 0.0020
#> 328  DPB1*47:01      1 0.0005
#> 329  DPB1*51:01      1 0.0005
#> 330 DPB1*519:01      1 0.0005
#> 331 DPB1*547:01      1 0.0005
#> 332  DPB1*57:01      3 0.0015
#> 333 DPB1*651:01      4 0.0020
#> 334 DPB1*702:01      7 0.0035
#> 335 DPB1*744:01      1 0.0005
#> 336  DPB1*80:01      1 0.0005
#> 337  DPB1*88:01      1 0.0005
#> 338  DPB1*92:01      1 0.0005
#> 339  DQA1*01:01    192 0.0960
#> 340  DQA1*01:02    358 0.1790
#> 341  DQA1*01:03    150 0.0750
#> 342  DQA1*01:04     52 0.0260
#> 343  DQA1*01:05     25 0.0125
#> 344  DQA1*01:06      1 0.0005
#> 345  DQA1*01:10      2 0.0010
#> 346  DQA1*02:01    256 0.1280
#> 347  DQA1*03:01    163 0.0815
#> 348  DQA1*03:02     31 0.0155
#> 349  DQA1*03:03    132 0.0660
#> 350  DQA1*04:01     58 0.0290
#> 351  DQA1*04:02      2 0.0010
#> 352  DQA1*05:01    222 0.1110
#> 353  DQA1*05:03     20 0.0100
#> 354  DQA1*05:05    305 0.1525
#> 355  DQA1*05:06      1 0.0005
#> 356  DQA1*05:07      1 0.0005
#> 357  DQA1*05:08      1 0.0005
#> 358  DQA1*05:09      6 0.0030
#> 359  DQA1*06:01     22 0.0110
#> 360  DQB1*02:01    218 0.1090
#> 361  DQB1*02:02    201 0.1005
#> 362  DQB1*02:10      1 0.0005
#> 363  DQB1*03:01    403 0.2015
#> 364  DQB1*03:02    169 0.0845
#> 365  DQB1*03:03     78 0.0390
#> 366  DQB1*03:04      7 0.0035
#> 367  DQB1*03:05      3 0.0015
#> 368  DQB1*03:19     11 0.0055
#> 369  DQB1*03:22     22 0.0110
#> 370 DQB1*03:251      2 0.0010
#> 371  DQB1*03:96      5 0.0025
#> 372  DQB1*04:01     26 0.0130
#> 373  DQB1*04:02     58 0.0290
#> 374  DQB1*04:23      3 0.0015
#> 375  DQB1*05:01    206 0.1030
#> 376  DQB1*05:02     68 0.0340
#> 377  DQB1*05:03     48 0.0240
#> 378  DQB1*05:04      4 0.0020
#> 379 DQB1*05:107      1 0.0005
#> 380  DQB1*06:01     49 0.0245
#> 381  DQB1*06:02    215 0.1075
#> 382  DQB1*06:03    113 0.0565
#> 383  DQB1*06:04     70 0.0350
#> 384  DQB1*06:07      1 0.0005
#> 385  DQB1*06:09     18 0.0090
#> 386   DRA*01:01   1289 0.6445
#> 387   DRA*01:02    711 0.3555
#> 388  DRB1*01:01    130 0.0650
#> 389  DRB1*01:02     38 0.0190
#> 390  DRB1*01:03      9 0.0045
#> 391  DRB1*03:01    215 0.1075
#> 392  DRB1*03:02      4 0.0020
#> 393  DRB1*03:07      1 0.0005
#> 394  DRB1*04:01    101 0.0505
#> 395  DRB1*04:02     20 0.0100
#> 396  DRB1*04:03     36 0.0180
#> 397  DRB1*04:04     48 0.0240
#> 398  DRB1*04:05     49 0.0245
#> 399  DRB1*04:06      8 0.0040
#> 400  DRB1*04:07     16 0.0080
#> 401  DRB1*04:08      7 0.0035
#> 402  DRB1*04:10      2 0.0010
#> 403  DRB1*04:13      1 0.0005
#> 404  DRB1*04:92      1 0.0005
#> 405  DRB1*07:01    249 0.1245
#> 406  DRB1*08:01     32 0.0160
#> 407  DRB1*08:02     14 0.0070
#> 408  DRB1*08:03     18 0.0090
#> 409  DRB1*08:04      9 0.0045
#> 410  DRB1*08:06      2 0.0010
#> 411  DRB1*08:11      1 0.0005
#> 412  DRB1*08:18      1 0.0005
#> 413  DRB1*08:40      1 0.0005
#> 414  DRB1*08:77      7 0.0035
#> 415  DRB1*09:01     32 0.0160
#> 416  DRB1*10:01     22 0.0110
#> 417  DRB1*11:01    137 0.0685
#> 418  DRB1*11:02      8 0.0040
#> 419  DRB1*11:03     12 0.0060
#> 420  DRB1*11:04     93 0.0465
#> 421  DRB1*11:14      1 0.0005
#> 422  DRB1*12:01     38 0.0190
#> 423  DRB1*12:02     18 0.0090
#> 424  DRB1*12:10      2 0.0010
#> 425  DRB1*13:01    105 0.0525
#> 426  DRB1*13:02     89 0.0445
#> 427  DRB1*13:03     26 0.0130
#> 428  DRB1*13:05      4 0.0020
#> 429  DRB1*13:12      2 0.0010
#> 430  DRB1*13:27      1 0.0005
#> 431  DRB1*14:01      7 0.0035
#> 432  DRB1*14:02      3 0.0015
#> 433  DRB1*14:03      2 0.0010
#> 434  DRB1*14:04      3 0.0015
#> 435  DRB1*14:05      2 0.0010
#> 436  DRB1*14:06      6 0.0030
#> 437  DRB1*14:07      1 0.0005
#> 438  DRB1*14:18      1 0.0005
#> 439  DRB1*14:54     40 0.0200
#> 440  DRB1*15:01    228 0.1140
#> 441  DRB1*15:02     37 0.0185
#> 442  DRB1*15:03      4 0.0020
#> 443  DRB1*16:01     39 0.0195
#> 444  DRB1*16:02      5 0.0025
#> 445  DRB1*16:15     10 0.0050
#> 446  DRB1*16:32      1 0.0005
#> 447  DRB1*16:42      1 0.0005