Get omnibus groups from MiDAS object.

Get omnibus groups from MiDAS object.

getOmnibusGroups(object, experiment)

Arguments

object
MiDAS object.

experiment
String specifying experiment.

Value

List of omnibus groups for a given experiment.

Details

For some experiments features can be naturally divided into groups (here called omnibus groups). For example, in "hla_aa" experiment features can be grouped by amino acid position ("B_46_E", "B_46_A") can be grouped into B_46 group). Such groups can be then used to perform omnibus test, see runMiDAS for more details.

Examples

getOmnibusGroups(object = MiDAS_tut_object,
                 experiment = "hla_aa")

#> $`A_-22`
#> [1] "A_-22_V" "A_-22_*" "A_-22_I"
#> 
#> $`A_-21`
#> [1] "A_-21_M" "A_-21_V" "A_-21_*"
#> 
#> $`A_-18`
#> [1] "A_-18_R" "A_-18_*" "A_-18_P"
#> 
#> $`A_-16`
#> [1] "A_-16_L" "A_-16_*" "A_-16_I"
#> 
#> $`A_-15`
#> [1] "A_-15_V" "A_-15_L" "A_-15_*"
#> 
#> $`A_-11`
#> [1] "A_-11_S" "A_-11_L" "A_-11_*"
#> 
#> $`A_-5`
#> [1] "A_-5_T" "A_-5_*" "A_-5_P"
#> 
#> $`A_-2`
#> [1] "A_-2_W" "A_-2_*" "A_-2_R"
#> 
#> $A_3
#> [1] "A_3_H" "A_3_Q"
#> 
#> $A_6
#> [1] "A_6_R" "A_6_G"
#> 
#> $A_9
#> [1] "A_9_Y" "A_9_F" "A_9_S" "A_9_T"
#> 
#> $A_12
#> [1] "A_12_V" "A_12_M"
#> 
#> $A_17
#> [1] "A_17_R" "A_17_S"
#> 
#> $A_19
#> [1] "A_19_E" "A_19_K"
#> 
#> $A_29
#> [1] "A_29_D" "A_29_A"
#> 
#> $A_43
#> [1] "A_43_Q" "A_43_R"
#> 
#> $A_44
#> [1] "A_44_R" "A_44_K"
#> 
#> $A_54
#> [1] "A_54_Q" "A_54_R"
#> 
#> $A_56
#> [1] "A_56_G" "A_56_R"
#> 
#> $A_62
#> [1] "A_62_R" "A_62_G" "A_62_E" "A_62_Q" "A_62_L"
#> 
#> $A_63
#> [1] "A_63_N" "A_63_E" "A_63_Q"
#> 
#> $A_65
#> [1] "A_65_R" "A_65_G"
#> 
#> $A_66
#> [1] "A_66_N" "A_66_K"
#> 
#> $A_67
#> [1] "A_67_V" "A_67_M"
#> 
#> $A_70
#> [1] "A_70_H" "A_70_Q"
#> 
#> $A_73
#> [1] "A_73_T" "A_73_I"
#> 
#> $A_74
#> [1] "A_74_D" "A_74_H"
#> 
#> $A_76
#> [1] "A_76_E" "A_76_V" "A_76_A"
#> 
#> $A_77
#> [1] "A_77_S" "A_77_D" "A_77_N"
#> 
#> $A_79
#> [1] "A_79_R" "A_79_G"
#> 
#> $A_80
#> [1] "A_80_I" "A_80_T"
#> 
#> $A_81
#> [1] "A_81_A" "A_81_L"
#> 
#> $A_82
#> [1] "A_82_L" "A_82_R"
#> 
#> $A_83
#> [1] "A_83_R" "A_83_G"
#> 
#> $A_90
#> [1] "A_90_D" "A_90_A"
#> 
#> $A_95
#> [1] "A_95_I" "A_95_V" "A_95_L"
#> 
#> $A_97
#> [1] "A_97_R" "A_97_M" "A_97_I"
#> 
#> $A_99
#> [1] "A_99_Y" "A_99_F" "A_99_C"
#> 
#> $A_102
#> [1] "A_102_D" "A_102_H"
#> 
#> $A_103
#> [1] "A_103_V" "A_103_L"
#> 
#> $A_105
#> [1] "A_105_P" "A_105_S"
#> 
#> $A_107
#> [1] "A_107_G" "A_107_W"
#> 
#> $A_109
#> [1] "A_109_F" "A_109_L"
#> 
#> $A_114
#> [1] "A_114_Q" "A_114_H" "A_114_R" "A_114_E"
#> 
#> $A_116
#> [1] "A_116_D" "A_116_Y" "A_116_H" "A_116_N" "A_116_F" "A_116_V"
#> 
#> $A_127
#> [1] "A_127_N" "A_127_K"
#> 
#> $A_142
#> [1] "A_142_I" "A_142_T"
#> 
#> $A_144
#> [1] "A_144_Q" "A_144_K"
#> 
#> $A_145
#> [1] "A_145_R" "A_145_H"
#> 
#> $A_149
#> [1] "A_149_T" "A_149_A"
#> 
#> $A_150
#> [1] "A_150_A" "A_150_V"
#> 
#> $A_151
#> [1] "A_151_H" "A_151_R" "A_151_L"
#> 
#> $A_152
#> [1] "A_152_E" "A_152_V" "A_152_A" "A_152_R" "A_152_W"
#> 
#> $A_156
#> [1] "A_156_W" "A_156_L" "A_156_Q" "A_156_R"
#> 
#> $A_158
#> [1] "A_158_A" "A_158_V"
#> 
#> $A_161
#> [1] "A_161_E" "A_161_D"
#> 
#> $A_163
#> [1] "A_163_R" "A_163_T" "A_163_E"
#> 
#> $A_164
#> [1] "A_164_C" "A_164_F"
#> 
#> $A_166
#> [1] "A_166_E" "A_166_D"
#> 
#> $A_167
#> [1] "A_167_W" "A_167_G"
#> 
#> $A_171
#> [1] "A_171_Y" "A_171_H"
#> 
#> $A_182
#> [1] "A_182_T" "A_182_A"
#> 
#> $A_184
#> [1] "A_184_A" "A_184_P" "A_184_*"
#> 
#> $A_186
#> [1] "A_186_K" "A_186_R" "A_186_*"
#> 
#> $A_193
#> [1] "A_193_A" "A_193_P" "A_193_*"
#> 
#> $A_194
#> [1] "A_194_V" "A_194_I" "A_194_*"
#> 
#> $A_207
#> [1] "A_207_S" "A_207_G" "A_207_*"
#> 
#> $A_219
#> [1] "A_219_R" "A_219_*" "A_219_Q"
#> 
#> $A_231
#> [1] "A_231_V" "A_231_*" "A_231_M"
#> 
#> $A_245
#> [1] "A_245_A" "A_245_V" "A_245_*"
#> 
#> $A_246
#> [1] "A_246_S" "A_246_A" "A_246_*"
#> 
#> $A_253
#> [1] "A_253_Q" "A_253_E" "A_253_*"
#> 
#> $A_276
#> [1] "A_276_P" "A_276_L" "A_276_*"
#> 
#> $A_282
#> [1] "A_282_I" "A_282_V" "A_282_*"
#> 
#> $A_283
#> [1] "A_283_P" "A_283_H" "A_283_*"
#> 
#> $A_284
#> [1] "A_284_I" "A_284_*" "A_284_T"
#> 
#> $A_288
#> [1] "A_288_I" "A_288_*" "A_288_L"
#> 
#> $A_294
#> [1] "A_294_F" "A_294_L" "A_294_*"
#> 
#> $A_297
#> [1] "A_297_V" "A_297_M" "A_297_*"
#> 
#> $A_298
#> [1] "A_298_I" "A_298_F" "A_298_*"
#> 
#> $A_299
#> [1] "A_299_A" "A_299_T" "A_299_*"
#> 
#> $A_307
#> [1] "A_307_M" "A_307_R" "A_307_*"
#> 
#> $A_311
#> [1] "A_311_K" "A_311_N" "A_311_*"
#> 
#> $A_321
#> [1] "A_321_S" "A_321_T" "A_321_*"
#> 
#> $A_334
#> [1] "A_334_M" "A_334_V" "A_334_*"
#> 
#> $`B_-23`
#> [1] "B_-23_L" "B_-23_R" "B_-23_*"
#> 
#> $`B_-21`
#> [1] "B_-21_M" "B_-21_T" "B_-21_*"
#> 
#> $`B_-16`
#> [1] "B_-16_V" "B_-16_L" "B_-16_*"
#> 
#> $`B_-11`
#> [1] "B_-11_S" "B_-11_W" "B_-11_*"
#> 
#> $`B_-10`
#> [1] "B_-10_A" "B_-10_G" "B_-10_*"
#> 
#> $`B_-9`
#> [1] "B_-9_A" "B_-9_*" "B_-9_V"
#> 
#> $`B_-8`
#> [1] "B_-8_L" "B_-8_V" "B_-8_*"
#> 
#> $B_9
#> [1] "B_9_Y" "B_9_H" "B_9_D"
#> 
#> $B_11
#> [1] "B_11_S" "B_11_A"
#> 
#> $B_12
#> [1] "B_12_V" "B_12_M"
#> 
#> $B_24
#> [1] "B_24_S" "B_24_T" "B_24_A"
#> 
#> $B_30
#> [1] "B_30_D" "B_30_G"
#> 
#> $B_32
#> [1] "B_32_Q" "B_32_L"
#> 
#> $B_33
#> [1] "B_33_F" "B_33_S"
#> 
#> $B_41
#> [1] "B_41_A" "B_41_T"
#> 
#> $B_45
#> [1] "B_45_E" "B_45_K" "B_45_M" "B_45_T" "B_45_G"
#> 
#> $B_46
#> [1] "B_46_E" "B_46_A"
#> 
#> $B_52
#> [1] "B_52_I" "B_52_V"
#> 
#> $B_59
#> [1] "B_59_Y" "B_59_H"
#> 
#> $B_62
#> [1] "B_62_R" "B_62_G"
#> 
#> $B_63
#> [1] "B_63_N" "B_63_E"
#> 
#> $B_65
#> [1] "B_65_Q" "B_65_R"
#> 
#> $B_66
#> [1] "B_66_I" "B_66_K" "B_66_N"
#> 
#> $B_67
#> [1] "B_67_Y" "B_67_S" "B_67_F" "B_67_C" "B_67_M"
#> 
#> $B_69
#> [1] "B_69_A" "B_69_T" "B_69_R"
#> 
#> $B_70
#> [1] "B_70_Q" "B_70_N" "B_70_K" "B_70_S"
#> 
#> $B_71
#> [1] "B_71_A" "B_71_T"
#> 
#> $B_72
#> [1] "B_72_Q" "B_72_R"
#> 
#> $B_74
#> [1] "B_74_D" "B_74_Y"
#> 
#> $B_76
#> [1] "B_76_E" "B_76_V"
#> 
#> $B_77
#> [1] "B_77_S" "B_77_N" "B_77_D" "B_77_G"
#> 
#> $B_80
#> [1] "B_80_N" "B_80_T" "B_80_I"
#> 
#> $B_81
#> [1] "B_81_L" "B_81_A"
#> 
#> $B_82
#> [1] "B_82_R" "B_82_L"
#> 
#> $B_83
#> [1] "B_83_G" "B_83_R"
#> 
#> $B_90
#> [1] "B_90_A" "B_90_D"
#> 
#> $B_94
#> [1] "B_94_T" "B_94_I"
#> 
#> $B_95
#> [1] "B_95_L" "B_95_W" "B_95_I"
#> 
#> $B_97
#> [1] "B_97_S" "B_97_R" "B_97_T" "B_97_N" "B_97_V" "B_97_W"
#> 
#> $B_99
#> [1] "B_99_Y" "B_99_S" "B_99_F"
#> 
#> $B_101
#> [1] "B_101_C" "B_101_Y"
#> 
#> $B_103
#> [1] "B_103_V" "B_103_L" "B_103_M"
#> 
#> $B_109
#> [1] "B_109_L" "B_109_F"
#> 
#> $B_113
#> [1] "B_113_H" "B_113_Y"
#> 
#> $B_114
#> [1] "B_114_D" "B_114_N" "B_114_H"
#> 
#> $B_116
#> [1] "B_116_Y" "B_116_L" "B_116_S" "B_116_F" "B_116_D"
#> 
#> $B_118
#> [1] "B_118_Y" "B_118_C"
#> 
#> $B_131
#> [1] "B_131_R" "B_131_S"
#> 
#> $B_137
#> [1] "B_137_D" "B_137_N"
#> 
#> $B_143
#> [1] "B_143_T" "B_143_S"
#> 
#> $B_145
#> [1] "B_145_R" "B_145_L"
#> 
#> $B_147
#> [1] "B_147_W" "B_147_L"
#> 
#> $B_152
#> [1] "B_152_E" "B_152_V"
#> 
#> $B_156
#> [1] "B_156_R" "B_156_L" "B_156_W" "B_156_D"
#> 
#> $B_158
#> [1] "B_158_A" "B_158_T"
#> 
#> $B_163
#> [1] "B_163_E" "B_163_L" "B_163_T"
#> 
#> $B_166
#> [1] "B_166_E" "B_166_D"
#> 
#> $B_167
#> [1] "B_167_W" "B_167_S" "B_167_G"
#> 
#> $B_171
#> [1] "B_171_Y" "B_171_H"
#> 
#> $B_177
#> [1] "B_177_D" "B_177_E"
#> 
#> $B_178
#> [1] "B_178_K" "B_178_T"
#> 
#> $B_180
#> [1] "B_180_E" "B_180_Q"
#> 
#> $B_194
#> [1] "B_194_I" "B_194_V" "B_194_*"
#> 
#> $B_199
#> [1] "B_199_A" "B_199_V" "B_199_*"
#> 
#> $B_211
#> [1] "B_211_A" "B_211_*" "B_211_G"
#> 
#> $B_239
#> [1] "B_239_R" "B_239_*" "B_239_G"
#> 
#> $B_245
#> [1] "B_245_A" "B_245_*" "B_245_T"
#> 
#> $B_253
#> [1] "B_253_E" "B_253_*" "B_253_Q"
#> 
#> $B_267
#> [1] "B_267_P" "B_267_*" "B_267_Q"
#> 
#> $B_268
#> [1] "B_268_K" "B_268_*" "B_268_E"
#> 
#> $B_270
#> [1] "B_270_L" "B_270_*" "B_270_C"
#> 
#> $B_275
#> [1] "B_275_E" "B_275_*" "B_275_K"
#> 
#> $B_282
#> [1] "B_282_V" "B_282_I" "B_282_*"
#> 
#> $B_295
#> [1] "B_295_." "B_295_*" "B_295_V"
#> 
#> $B_296
#> [1] "B_296_A" "B_296_*" "B_296_V"
#> 
#> $B_297
#> [1] "B_297_V" "B_297_*" "B_297_T"
#> 
#> $B_299
#> [1] "B_299_V" "B_299_*" "B_299_A"
#> 
#> $B_300
#> [1] "B_300_I" "B_300_*" "B_300_V"
#> 
#> $B_301
#> [1] "B_301_G" "B_301_*" "B_301_V"
#> 
#> $B_306
#> [1] "B_306_A" "B_306_T" "B_306_*"
#> 
#> $B_308
#> [1] "B_308_M" "B_308_*" "B_308_V"
#> 
#> $B_326
#> [1] "B_326_C" "B_326_S" "B_326_*"
#> 
#> $B_339
#> [1] "B_339_A" "B_339_*" "B_339_D"
#> 
#> $`C_-18`
#> [1] "C_-18_R" "C_-18_*" "C_-18_Q"
#> 
#> $`C_-17`
#> [1] "C_-17_T" "C_-17_A" "C_-17_*"
#> 
#> $`C_-15`
#> [1] "C_-15_I" "C_-15_L" "C_-15_*"
#> 
#> $`C_-9`
#> [1] "C_-9_A" "C_-9_G" "C_-9_*"
#> 
#> $`C_-5`
#> [1] "C_-5_T" "C_-5_*" "C_-5_I"
#> 
#> $`C_-1`
#> [1] "C_-1_A" "C_-1_*" "C_-1_T"
#> 
#> $C_1
#> [1] "C_1_C" "C_1_G" "C_1_*"
#> 
#> $C_6
#> [1] "C_6_R" "C_6_K"
#> 
#> $C_9
#> [1] "C_9_Y" "C_9_D" "C_9_F" "C_9_S"
#> 
#> $C_11
#> [1] "C_11_A" "C_11_S"
#> 
#> $C_14
#> [1] "C_14_R" "C_14_W"
#> 
#> $C_16
#> [1] "C_16_G" "C_16_S"
#> 
#> $C_21
#> [1] "C_21_R" "C_21_H"
#> 
#> $C_24
#> [1] "C_24_A" "C_24_S"
#> 
#> $C_35
#> [1] "C_35_R" "C_35_Q"
#> 
#> $C_49
#> [1] "C_49_A" "C_49_E"
#> 
#> $C_66
#> [1] "C_66_K" "C_66_N"
#> 
#> $C_68
#> [1] "C_68_K" "C_68_N"
#> 
#> $C_73
#> [1] "C_73_A" "C_73_T"
#> 
#> $C_77
#> [1] "C_77_S" "C_77_N"
#> 
#> $C_80
#> [1] "C_80_N" "C_80_K"
#> 
#> $C_90
#> [1] "C_90_A" "C_90_D"
#> 
#> $C_91
#> [1] "C_91_G" "C_91_R"
#> 
#> $C_94
#> [1] "C_94_T" "C_94_I"
#> 
#> $C_95
#> [1] "C_95_L" "C_95_I" "C_95_F"
#> 
#> $C_97
#> [1] "C_97_W" "C_97_R" "C_97_S"
#> 
#> $C_99
#> [1] "C_99_Y" "C_99_C" "C_99_F" "C_99_S"
#> 
#> $C_103
#> [1] "C_103_L" "C_103_V"
#> 
#> $C_113
#> [1] "C_113_Y" "C_113_H"
#> 
#> $C_114
#> [1] "C_114_D" "C_114_N" "C_114_V"
#> 
#> $C_116
#> [1] "C_116_S" "C_116_F" "C_116_Y" "C_116_L" "C_116_H"
#> 
#> $C_127
#> [1] "C_127_N" "C_127_K"
#> 
#> $C_138
#> [1] "C_138_T" "C_138_K"
#> 
#> $C_143
#> [1] "C_143_T" "C_143_S"
#> 
#> $C_147
#> [1] "C_147_W" "C_147_L"
#> 
#> $C_152
#> [1] "C_152_E" "C_152_A" "C_152_T"
#> 
#> $C_156
#> [1] "C_156_W" "C_156_R" "C_156_L" "C_156_D" "C_156_Q"
#> 
#> $C_163
#> [1] "C_163_T" "C_163_L" "C_163_E"
#> 
#> $C_169
#> [1] "C_169_R" "C_169_H"
#> 
#> $C_170
#> [1] "C_170_R" "C_170_G"
#> 
#> $C_173
#> [1] "C_173_E" "C_173_K"
#> 
#> $C_175
#> [1] "C_175_G" "C_175_R"
#> 
#> $C_177
#> [1] "C_177_E" "C_177_K"
#> 
#> $C_184
#> [1] "C_184_H" "C_184_P" "C_184_*" "C_184_R"
#> 
#> $C_186
#> [1] "C_186_K" "C_186_*" "C_186_R"
#> 
#> $C_193
#> [1] "C_193_P" "C_193_L" "C_193_*"
#> 
#> $C_194
#> [1] "C_194_V" "C_194_L" "C_194_*"
#> 
#> $C_211
#> [1] "C_211_A" "C_211_T" "C_211_*"
#> 
#> $C_219
#> [1] "C_219_R" "C_219_W" "C_219_*"
#> 
#> $C_248
#> [1] "C_248_V" "C_248_M" "C_248_*"
#> 
#> $C_253
#> [1] "C_253_E" "C_253_Q" "C_253_*"
#> 
#> $C_261
#> [1] "C_261_V" "C_261_M" "C_261_*"
#> 
#> $C_267
#> [1] "C_267_P" "C_267_Q" "C_267_*"
#> 
#> $C_270
#> [1] "C_270_L" "C_270_*" "C_270_C"
#> 
#> $C_273
#> [1] "C_273_R" "C_273_S" "C_273_*"
#> 
#> $C_275
#> [1] "C_275_E" "C_275_K" "C_275_G" "C_275_*"
#> 
#> $C_284
#> [1] "C_284_I" "C_284_*" "C_284_N"
#> 
#> $C_285
#> [1] "C_285_V" "C_285_M" "C_285_*" "C_285_L"
#> 
#> $C_289
#> [1] "C_289_A" "C_289_*" "C_289_S"
#> 
#> $C_291
#> [1] "C_291_L" "C_291_*" "C_291_P"
#> 
#> $C_295
#> [1] "C_295_A" "C_295_V" "C_295_*"
#> 
#> $C_301
#> [1] "C_301_." "C_301_A"
#> 
#> $C_302
#> [1] "C_302_." "C_302_V"
#> 
#> $C_303
#> [1] "C_303_." "C_303_L"
#> 
#> $C_304
#> [1] "C_304_." "C_304_A"
#> 
#> $C_305
#> [1] "C_305_." "C_305_V"
#> 
#> $C_306
#> [1] "C_306_." "C_306_L"
#> 
#> $C_309
#> [1] "C_309_V" "C_309_M" "C_309_*"
#> 
#> $C_310
#> [1] "C_310_M" "C_310_V" "C_310_*"
#> 
#> $C_311
#> [1] "C_311_A" "C_311_T" "C_311_*"
#> 
#> $C_312
#> [1] "C_312_V" "C_312_A" "C_312_*"
#> 
#> $C_313
#> [1] "C_313_V" "C_313_M" "C_313_*"
#> 
#> $C_314
#> [1] "C_314_M" "C_314_*" "C_314_I"
#> 
#> $C_315
#> [1] "C_315_C" "C_315_*" "C_315_H"
#> 
#> $C_327
#> [1] "C_327_C" "C_327_*" "C_327_Y"
#> 
#> $C_330
#> [1] "C_330_A" "C_330_V" "C_330_*"
#> 
#> $C_332
#> [1] "C_332_S" "C_332_C" "C_332_*"
#> 
#> $C_345
#> [1] "C_345_A" "C_345_T" "C_345_*"
#> 
#> $DPA1_11
#> [1] "DPA1_11_A" "DPA1_11_M"
#> 
#> $DPA1_18
#> [1] "DPA1_18_P" "DPA1_18_T"
#> 
#> $DPA1_28
#> [1] "DPA1_28_E" "DPA1_28_D"
#> 
#> $DPA1_31
#> [1] "DPA1_31_M" "DPA1_31_Q"
#> 
#> $DPA1_50
#> [1] "DPA1_50_Q" "DPA1_50_R"
#> 
#> $DPA1_66
#> [1] "DPA1_66_L" "DPA1_66_S"
#> 
#> $DPA1_72
#> [1] "DPA1_72_T" "DPA1_72_I"
#> 
#> $DPA1_73
#> [1] "DPA1_73_L" "DPA1_73_A"
#> 
#> $DPA1_83
#> [1] "DPA1_83_T" "DPA1_83_A"
#> 
#> $DPA1_91
#> [1] "DPA1_91_V" "DPA1_91_M" "DPA1_91_*"
#> 
#> $DPA1_96
#> [1] "DPA1_96_P" "DPA1_96_*" "DPA1_96_A"
#> 
#> $DPA1_111
#> [1] "DPA1_111_K" "DPA1_111_R" "DPA1_111_*"
#> 
#> $DPA1_127
#> [1] "DPA1_127_L" "DPA1_127_P" "DPA1_127_*"
#> 
#> $DPA1_160
#> [1] "DPA1_160_F" "DPA1_160_V" "DPA1_160_*"
#> 
#> $DPA1_190
#> [1] "DPA1_190_T" "DPA1_190_M" "DPA1_190_*" "DPA1_190_A"
#> 
#> $DPA1_228
#> [1] "DPA1_228_T" "DPA1_228_P" "DPA1_228_*"
#> 
#> $`DPB1_-14`
#> [1] "DPB1_-14_T" "DPB1_-14_*" "DPB1_-14_M"
#> 
#> $DPB1_8
#> [1] "DPB1_8_L" "DPB1_8_V"
#> 
#> $DPB1_9
#> [1] "DPB1_9_F" "DPB1_9_Y" "DPB1_9_H"
#> 
#> $DPB1_11
#> [1] "DPB1_11_G" "DPB1_11_L"
#> 
#> $DPB1_17
#> [1] "DPB1_17_A" "DPB1_17_P"
#> 
#> $DPB1_33
#> [1] "DPB1_33_E" "DPB1_33_Q"
#> 
#> $DPB1_35
#> [1] "DPB1_35_F" "DPB1_35_L" "DPB1_35_Y"
#> 
#> $DPB1_36
#> [1] "DPB1_36_V" "DPB1_36_A"
#> 
#> $DPB1_55
#> [1] "DPB1_55_D" "DPB1_55_E" "DPB1_55_A"
#> 
#> $DPB1_56
#> [1] "DPB1_56_E" "DPB1_56_A"
#> 
#> $DPB1_57
#> [1] "DPB1_57_E" "DPB1_57_D"
#> 
#> $DPB1_65
#> [1] "DPB1_65_I" "DPB1_65_L" "DPB1_65_F"
#> 
#> $DPB1_69
#> [1] "DPB1_69_E" "DPB1_69_K" "DPB1_69_R"
#> 
#> $DPB1_76
#> [1] "DPB1_76_M" "DPB1_76_V" "DPB1_76_I"
#> 
#> $DPB1_84
#> [1] "DPB1_84_G" "DPB1_84_D" "DPB1_84_V"
#> 
#> $DPB1_85
#> [1] "DPB1_85_G" "DPB1_85_E"
#> 
#> $DPB1_86
#> [1] "DPB1_86_P" "DPB1_86_A"
#> 
#> $DPB1_87
#> [1] "DPB1_87_M" "DPB1_87_V"
#> 
#> $DPB1_96
#> [1] "DPB1_96_R" "DPB1_96_K" "DPB1_96_*"
#> 
#> $DPB1_170
#> [1] "DPB1_170_T" "DPB1_170_I" "DPB1_170_*"
#> 
#> $DPB1_178
#> [1] "DPB1_178_L" "DPB1_178_M" "DPB1_178_*"
#> 
#> $DPB1_194
#> [1] "DPB1_194_R" "DPB1_194_Q" "DPB1_194_*"
#> 
#> $DPB1_205
#> [1] "DPB1_205_V" "DPB1_205_M" "DPB1_205_*"
#> 
#> $DPB1_215
#> [1] "DPB1_215_I" "DPB1_215_T" "DPB1_215_*"
#> 
#> $`DQA1_-16`
#> [1] "DQA1_-16_M" "DQA1_-16_L" "DQA1_-16_*"
#> 
#> $`DQA1_-13`
#> [1] "DQA1_-13_T" "DQA1_-13_A" "DQA1_-13_*"
#> 
#> $`DQA1_-7`
#> [1] "DQA1_-7_V" "DQA1_-7_M" "DQA1_-7_*"
#> 
#> $`DQA1_-6`
#> [1] "DQA1_-6_M" "DQA1_-6_T" "DQA1_-6_*"
#> 
#> $DQA1_1
#> [1] "DQA1_1_E" "DQA1_1_K" "DQA1_1_*"
#> 
#> $DQA1_2
#> [1] "DQA1_2_D" "DQA1_2_G" "DQA1_2_*"
#> 
#> $DQA1_11
#> [1] "DQA1_11_Y" "DQA1_11_C"
#> 
#> $DQA1_18
#> [1] "DQA1_18_S" "DQA1_18_F"
#> 
#> $DQA1_25
#> [1] "DQA1_25_Y" "DQA1_25_F"
#> 
#> $DQA1_26
#> [1] "DQA1_26_T" "DQA1_26_S"
#> 
#> $DQA1_34
#> [1] "DQA1_34_Q" "DQA1_34_E"
#> 
#> $DQA1_40
#> [1] "DQA1_40_G" "DQA1_40_E"
#> 
#> $DQA1_41
#> [1] "DQA1_41_R" "DQA1_41_K"
#> 
#> $DQA1_44
#> [1] "DQA1_44_T" "DQA1_44_A"
#> 
#> $DQA1_45
#> [1] "DQA1_45_V" "DQA1_45_A"
#> 
#> $DQA1_47
#> [1] "DQA1_47_C" "DQA1_47_K" "DQA1_47_R" "DQA1_47_Q"
#> 
#> $DQA1_48
#> [1] "DQA1_48_L" "DQA1_48_W"
#> 
#> $DQA1_50
#> [1] "DQA1_50_V" "DQA1_50_L" "DQA1_50_E"
#> 
#> $DQA1_51
#> [1] "DQA1_51_L" "DQA1_51_F"
#> 
#> $DQA1_52
#> [1] "DQA1_52_R" "DQA1_52_H" "DQA1_52_S"
#> 
#> $DQA1_53
#> [1] "DQA1_53_Q" "DQA1_53_R" "DQA1_53_K"
#> 
#> $DQA1_54
#> [1] "DQA1_54_F" "DQA1_54_L"
#> 
#> $DQA1_55
#> [1] "DQA1_55_R" "DQA1_55_G"
#> 
#> $DQA1_56
#> [1] "DQA1_56_." "DQA1_56_G" "DQA1_56_R"
#> 
#> $DQA1_61
#> [1] "DQA1_61_F" "DQA1_61_G"
#> 
#> $DQA1_64
#> [1] "DQA1_64_T" "DQA1_64_R"
#> 
#> $DQA1_66
#> [1] "DQA1_66_I" "DQA1_66_M"
#> 
#> $DQA1_69
#> [1] "DQA1_69_L" "DQA1_69_A" "DQA1_69_T"
#> 
#> $DQA1_75
#> [1] "DQA1_75_S" "DQA1_75_I"
#> 
#> $DQA1_76
#> [1] "DQA1_76_L" "DQA1_76_M" "DQA1_76_V"
#> 
#> $DQA1_80
#> [1] "DQA1_80_S" "DQA1_80_Y"
#> 
#> $DQA1_84
#> [1] "DQA1_84_A" "DQA1_84_T"
#> 
#> $DQA1_102
#> [1] "DQA1_102_L" "DQA1_102_V" "DQA1_102_*"
#> 
#> $DQA1_107
#> [1] "DQA1_107_I" "DQA1_107_T" "DQA1_107_*"
#> 
#> $DQA1_129
#> [1] "DQA1_129_H" "DQA1_129_Q" "DQA1_129_*"
#> 
#> $DQA1_130
#> [1] "DQA1_130_S" "DQA1_130_A" "DQA1_130_*"
#> 
#> $DQA1_138
#> [1] "DQA1_138_T" "DQA1_138_I" "DQA1_138_*"
#> 
#> $DQA1_139
#> [1] "DQA1_139_S" "DQA1_139_I" "DQA1_139_*"
#> 
#> $DQA1_156
#> [1] "DQA1_156_L" "DQA1_156_F" "DQA1_156_*"
#> 
#> $DQA1_160
#> [1] "DQA1_160_A" "DQA1_160_D" "DQA1_160_S" "DQA1_160_*"
#> 
#> $DQA1_161
#> [1] "DQA1_161_E" "DQA1_161_D" "DQA1_161_*"
#> 
#> $DQA1_163
#> [1] "DQA1_163_S" "DQA1_163_I" "DQA1_163_*"
#> 
#> $DQA1_175
#> [1] "DQA1_175_K" "DQA1_175_E" "DQA1_175_Q" "DQA1_175_*"
#> 
#> $DQA1_187
#> [1] "DQA1_187_A" "DQA1_187_T" "DQA1_187_*"
#> 
#> $DQA1_199
#> [1] "DQA1_199_A" "DQA1_199_T" "DQA1_199_*"
#> 
#> $DQA1_207
#> [1] "DQA1_207_V" "DQA1_207_M" "DQA1_207_*"
#> 
#> $DQA1_208
#> [1] "DQA1_208_G" "DQA1_208_C" "DQA1_208_*"
#> 
#> $DQA1_215
#> [1] "DQA1_215_F" "DQA1_215_L" "DQA1_215_*"
#> 
#> $DQA1_218
#> [1] "DQA1_218_R" "DQA1_218_Q" "DQA1_218_*"
#> 
#> $`DQB1_-27`
#> [1] "DQB1_-27_A" "DQB1_-27_S" "DQB1_-27_*"
#> 
#> $`DQB1_-21`
#> [1] "DQB1_-21_D" "DQB1_-21_G" "DQB1_-21_*"
#> 
#> $`DQB1_-18`
#> [1] "DQB1_-18_V" "DQB1_-18_A" "DQB1_-18_*"
#> 
#> $`DQB1_-17`
#> [1] "DQB1_-17_A" "DQB1_-17_P" "DQB1_-17_*"
#> 
#> $`DQB1_-10`
#> [1] "DQB1_-10_A" "DQB1_-10_S" "DQB1_-10_*"
#> 
#> $`DQB1_-9`
#> [1] "DQB1_-9_M" "DQB1_-9_I" "DQB1_-9_*"
#> 
#> $`DQB1_-6`
#> [1] "DQB1_-6_S" "DQB1_-6_T" "DQB1_-6_*"
#> 
#> $`DQB1_-5`
#> [1] "DQB1_-5_L" "DQB1_-5_P" "DQB1_-5_S" "DQB1_-5_*"
#> 
#> $`DQB1_-4`
#> [1] "DQB1_-4_L" "DQB1_-4_V" "DQB1_-4_*"
#> 
#> $DQB1_3
#> [1] "DQB1_3_S" "DQB1_3_P" "DQB1_3_*"
#> 
#> $DQB1_9
#> [1] "DQB1_9_F" "DQB1_9_Y" "DQB1_9_L"
#> 
#> $DQB1_13
#> [1] "DQB1_13_G" "DQB1_13_A"
#> 
#> $DQB1_14
#> [1] "DQB1_14_M" "DQB1_14_L"
#> 
#> $DQB1_23
#> [1] "DQB1_23_R" "DQB1_23_L"
#> 
#> $DQB1_26
#> [1] "DQB1_26_L" "DQB1_26_Y" "DQB1_26_G"
#> 
#> $DQB1_28
#> [1] "DQB1_28_T" "DQB1_28_S"
#> 
#> $DQB1_30
#> [1] "DQB1_30_Y" "DQB1_30_S" "DQB1_30_H"
#> 
#> $DQB1_37
#> [1] "DQB1_37_Y" "DQB1_37_I" "DQB1_37_D"
#> 
#> $DQB1_38
#> [1] "DQB1_38_A" "DQB1_38_V"
#> 
#> $DQB1_45
#> [1] "DQB1_45_G" "DQB1_45_E"
#> 
#> $DQB1_46
#> [1] "DQB1_46_V" "DQB1_46_E"
#> 
#> $DQB1_47
#> [1] "DQB1_47_Y" "DQB1_47_F"
#> 
#> $DQB1_52
#> [1] "DQB1_52_P" "DQB1_52_L"
#> 
#> $DQB1_53
#> [1] "DQB1_53_Q" "DQB1_53_L"
#> 
#> $DQB1_55
#> [1] "DQB1_55_R" "DQB1_55_L" "DQB1_55_P"
#> 
#> $DQB1_56
#> [1] "DQB1_56_P" "DQB1_56_L"
#> 
#> $DQB1_57
#> [1] "DQB1_57_D" "DQB1_57_A" "DQB1_57_V" "DQB1_57_S"
#> 
#> $DQB1_66
#> [1] "DQB1_66_E" "DQB1_66_D"
#> 
#> $DQB1_67
#> [1] "DQB1_67_V" "DQB1_67_I"
#> 
#> $DQB1_70
#> [1] "DQB1_70_G" "DQB1_70_R" "DQB1_70_E"
#> 
#> $DQB1_71
#> [1] "DQB1_71_T" "DQB1_71_K" "DQB1_71_D" "DQB1_71_A"
#> 
#> $DQB1_74
#> [1] "DQB1_74_E" "DQB1_74_A" "DQB1_74_S"
#> 
#> $DQB1_75
#> [1] "DQB1_75_L" "DQB1_75_V"
#> 
#> $DQB1_77
#> [1] "DQB1_77_T" "DQB1_77_R"
#> 
#> $DQB1_84
#> [1] "DQB1_84_E" "DQB1_84_Q"
#> 
#> $DQB1_85
#> [1] "DQB1_85_V" "DQB1_85_L"
#> 
#> $DQB1_86
#> [1] "DQB1_86_A" "DQB1_86_E" "DQB1_86_G"
#> 
#> $DQB1_87
#> [1] "DQB1_87_F" "DQB1_87_L" "DQB1_87_Y"
#> 
#> $DQB1_89
#> [1] "DQB1_89_G" "DQB1_89_T"
#> 
#> $DQB1_90
#> [1] "DQB1_90_I" "DQB1_90_T"
#> 
#> $DQB1_116
#> [1] "DQB1_116_V" "DQB1_116_I"
#> 
#> $DQB1_125
#> [1] "DQB1_125_G" "DQB1_125_A" "DQB1_125_S"
#> 
#> $DQB1_126
#> [1] "DQB1_126_Q" "DQB1_126_H"
#> 
#> $DQB1_130
#> [1] "DQB1_130_R" "DQB1_130_Q"
#> 
#> $DQB1_135
#> [1] "DQB1_135_D" "DQB1_135_G"
#> 
#> $DQB1_140
#> [1] "DQB1_140_A" "DQB1_140_T"
#> 
#> $DQB1_142
#> [1] "DQB1_142_V" "DQB1_142_I"
#> 
#> $DQB1_167
#> [1] "DQB1_167_R" "DQB1_167_H"
#> 
#> $DQB1_177
#> [1] "DQB1_177_H" "DQB1_177_Q"
#> 
#> $DQB1_182
#> [1] "DQB1_182_S" "DQB1_182_N"
#> 
#> $DQB1_185
#> [1] "DQB1_185_T" "DQB1_185_I"
#> 
#> $DQB1_197
#> [1] "DQB1_197_S" "DQB1_197_N" "DQB1_197_*"
#> 
#> $DQB1_203
#> [1] "DQB1_203_V" "DQB1_203_I" "DQB1_203_*"
#> 
#> $DQB1_220
#> [1] "DQB1_220_R" "DQB1_220_H" "DQB1_220_*"
#> 
#> $DQB1_221
#> [1] "DQB1_221_Q" "DQB1_221_H" "DQB1_221_*"
#> 
#> $DQB1_224
#> [1] "DQB1_224_Q" "DQB1_224_R" "DQB1_224_*"
#> 
#> $DQB1_227
#> [1] "DQB1_227_." "DQB1_227_P" "DQB1_227_*"
#> 
#> $DQB1_228
#> [1] "DQB1_228_." "DQB1_228_Q" "DQB1_228_*"
#> 
#> $DQB1_229
#> [1] "DQB1_229_." "DQB1_229_G" "DQB1_229_*"
#> 
#> $DQB1_230
#> [1] "DQB1_230_." "DQB1_230_P" "DQB1_230_*"
#> 
#> $DQB1_231
#> [1] "DQB1_231_." "DQB1_231_P" "DQB1_231_*"
#> 
#> $DQB1_232
#> [1] "DQB1_232_." "DQB1_232_P" "DQB1_232_*"
#> 
#> $DQB1_233
#> [1] "DQB1_233_." "DQB1_233_A" "DQB1_233_*"
#> 
#> $DQB1_234
#> [1] "DQB1_234_." "DQB1_234_G" "DQB1_234_*"
#> 
#> $DRA_217
#> [1] "DRA_217_L" "DRA_217_V"
#> 
#> $`DRB1_-25`
#> [1] "DRB1_-25_K" "DRB1_-25_R" "DRB1_-25_*"
#> 
#> $`DRB1_-24`
#> [1] "DRB1_-24_L" "DRB1_-24_F" "DRB1_-24_*"
#> 
#> $`DRB1_-17`
#> [1] "DRB1_-17_T" "DRB1_-17_A" "DRB1_-17_*"
#> 
#> $`DRB1_-16`
#> [1] "DRB1_-16_A" "DRB1_-16_V" "DRB1_-16_I" "DRB1_-16_*"
#> 
#> $`DRB1_-1`
#> [1] "DRB1_-1_S" "DRB1_-1_A" "DRB1_-1_*"
#> 
#> $DRB1_4
#> [1] "DRB1_4_R" "DRB1_4_Q" "DRB1_4_*"
#> 
#> $DRB1_9
#> [1] "DRB1_9_W" "DRB1_9_E" "DRB1_9_K"
#> 
#> $DRB1_10
#> [1] "DRB1_10_Q" "DRB1_10_Y" "DRB1_10_E"
#> 
#> $DRB1_11
#> [1] "DRB1_11_P" "DRB1_11_G" "DRB1_11_S" "DRB1_11_D" "DRB1_11_V" "DRB1_11_L"
#> 
#> $DRB1_12
#> [1] "DRB1_12_K" "DRB1_12_T"
#> 
#> $DRB1_13
#> [1] "DRB1_13_R" "DRB1_13_Y" "DRB1_13_S" "DRB1_13_F" "DRB1_13_H" "DRB1_13_G"
#> 
#> $DRB1_14
#> [1] "DRB1_14_E" "DRB1_14_K"
#> 
#> $DRB1_16
#> [1] "DRB1_16_H" "DRB1_16_Y" "DRB1_16_Q"
#> 
#> $DRB1_25
#> [1] "DRB1_25_R" "DRB1_25_Q"
#> 
#> $DRB1_26
#> [1] "DRB1_26_F" "DRB1_26_Y" "DRB1_26_L"
#> 
#> $DRB1_28
#> [1] "DRB1_28_D" "DRB1_28_E" "DRB1_28_H"
#> 
#> $DRB1_30
#> [1] "DRB1_30_Y" "DRB1_30_L" "DRB1_30_G" "DRB1_30_R" "DRB1_30_C" "DRB1_30_H"
#> 
#> $DRB1_31
#> [1] "DRB1_31_F" "DRB1_31_I" "DRB1_31_V"
#> 
#> $DRB1_32
#> [1] "DRB1_32_Y" "DRB1_32_H"
#> 
#> $DRB1_33
#> [1] "DRB1_33_N" "DRB1_33_H"
#> 
#> $DRB1_37
#> [1] "DRB1_37_S" "DRB1_37_F" "DRB1_37_N" "DRB1_37_Y" "DRB1_37_L"
#> 
#> $DRB1_38
#> [1] "DRB1_38_V" "DRB1_38_A" "DRB1_38_L"
#> 
#> $DRB1_40
#> [1] "DRB1_40_F" "DRB1_40_Y"
#> 
#> $DRB1_47
#> [1] "DRB1_47_F" "DRB1_47_Y"
#> 
#> $DRB1_57
#> [1] "DRB1_57_D" "DRB1_57_V" "DRB1_57_A" "DRB1_57_S"
#> 
#> $DRB1_58
#> [1] "DRB1_58_A" "DRB1_58_E"
#> 
#> $DRB1_60
#> [1] "DRB1_60_Y" "DRB1_60_S" "DRB1_60_H"
#> 
#> $DRB1_67
#> [1] "DRB1_67_I" "DRB1_67_L" "DRB1_67_F"
#> 
#> $DRB1_70
#> [1] "DRB1_70_Q" "DRB1_70_D" "DRB1_70_R"
#> 
#> $DRB1_71
#> [1] "DRB1_71_A" "DRB1_71_R" "DRB1_71_E" "DRB1_71_K"
#> 
#> $DRB1_73
#> [1] "DRB1_73_A" "DRB1_73_G"
#> 
#> $DRB1_74
#> [1] "DRB1_74_A" "DRB1_74_Q" "DRB1_74_E" "DRB1_74_R" "DRB1_74_L"
#> 
#> $DRB1_77
#> [1] "DRB1_77_T" "DRB1_77_N"
#> 
#> $DRB1_78
#> [1] "DRB1_78_Y" "DRB1_78_V"
#> 
#> $DRB1_82
#> [1] "DRB1_82_N" "DRB1_82_S"
#> 
#> $DRB1_85
#> [1] "DRB1_85_V" "DRB1_85_A"
#> 
#> $DRB1_86
#> [1] "DRB1_86_V" "DRB1_86_G"
#> 
#> $DRB1_96
#> [1] "DRB1_96_Q" "DRB1_96_H" "DRB1_96_Y" "DRB1_96_E" "DRB1_96_*"
#> 
#> $DRB1_98
#> [1] "DRB1_98_K" "DRB1_98_E" "DRB1_98_*"
#> 
#> $DRB1_104
#> [1] "DRB1_104_S" "DRB1_104_A" "DRB1_104_*"
#> 
#> $DRB1_112
#> [1] "DRB1_112_H" "DRB1_112_Y" "DRB1_112_*"
#> 
#> $DRB1_120
#> [1] "DRB1_120_S" "DRB1_120_N" "DRB1_120_*"
#> 
#> $DRB1_133
#> [1] "DRB1_133_L" "DRB1_133_R" "DRB1_133_*"
#> 
#> $DRB1_140
#> [1] "DRB1_140_A" "DRB1_140_T" "DRB1_140_*"
#> 
#> $DRB1_142
#> [1] "DRB1_142_M" "DRB1_142_V" "DRB1_142_*"
#> 
#> $DRB1_149
#> [1] "DRB1_149_Q" "DRB1_149_H" "DRB1_149_*"
#> 
#> $DRB1_166
#> [1] "DRB1_166_R" "DRB1_166_Q" "DRB1_166_*"
#> 
#> $DRB1_180
#> [1] "DRB1_180_V" "DRB1_180_L" "DRB1_180_*"
#> 
#> $DRB1_181
#> [1] "DRB1_181_T" "DRB1_181_M" "DRB1_181_*"
#> 
#> $DRB1_189
#> [1] "DRB1_189_R" "DRB1_189_S" "DRB1_189_*"
#> 
#> $DRB1_207
#> [1] "DRB1_207_V" "DRB1_207_*" "DRB1_207_M"
#> 
#> $DRB1_231
#> [1] "DRB1_231_Q" "DRB1_231_P" "DRB1_231_*"
#> 
#> $DRB1_233
#> [1] "DRB1_233_T" "DRB1_233_R" "DRB1_233_*"
#>